### 2025-04-06 #8006 @editorialbot
[PRE REVIEW]: OpenSim Creator: A Graphical User Interface for Building OpenSim Musculoskeletal Models
**Submitting author:** @adamkewley (<a href="http://orcid.org/0000-0002-6505-5374">Adam Kewley</a>)
**Repository:** https://github.com/ComputationalBiomechanicsLab/opensim-creator
**Branch with paper.md** (empty if default branch): 2025_joss-submission
**Version:** 0.5.21
