### 2025-04-07 #8006 @adamkewley
I will go and fix the DOIs. I have contacted suitable potential reviewers that I sourced from the similar historical papers (above) and https://reviewers.joss.theoj.org/reviewers/ - I'll list their names as suggested reviewers if they are willing!
