### 2025-04-07 3cafa5f
Fix DOIs, as detected by @editorialbot in openjournals/joss-reviews#8006
