+++ paper/paper.bib
@@ -18,7 +18,7 @@
   number = {1},
   pages = {207-227},
   keywords = {3D modelling, adduction–abduction, hominoid shoulder, moment arms, musculoskeletal model, scapula functional morphology, western lowland gorilla},
-  doi = {https://doi.org/10.1111/joa.13412},
+  doi = {10.1111/joa.13412},
   url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/joa.13412},
   eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/joa.13412},
   year = {2021}
@@ -101,7 +101,7 @@
   pages = {112560},
   year = {2025},
   issn = {0021-9290},
-  doi = {https://doi.org/10.1016/j.jbiomech.2025.112560},
+  doi = {10.1016/j.jbiomech.2025.112560},
   url = {https://www.sciencedirect.com/science/article/pii/S0021929025000715},
   author = {Eline {van der Kruk}},
   keywords = {Biasmechanics, Bias, Gender, Sex}
