### 2025-04-07 cd3c3eb
Add std::generator shim (WIP)
