### 2025-04-07 2694b6a
Add std::generator shim (WIP)
