### 2025-04-08 ca25231
Squashed 'third_party/SDL/' changes from 535d80bade..5de97ca3b9
5de97ca3b9 Add OFN_OVERWRITEPROMPT to window save dialogs
877399b2b2 Updated to version 3.2.10 for release
3f059376ac wayland: Commit the confinement region upon creation
30a6d01cc5 wayland: Defer creating cursor confinement regions until the surface is mapped
b6fb47aa70 Sync SDL3 wiki -> header
7ecae856e5 Update include/SDL3/SDL_power.h
5db7694f22 Add performance disclaimer in `SDL_GetPowerInfo` documentation
366294e1c5 Fixed build
4d966ff546 Set the alpha opaque when blending to an RGB format
6b665066f0 Added SIMD blitters for 8
[…]
