### 2025-04-09 10f755b
Squashed 'third_party/SDL/' changes from 5de97ca3b97..f303ccca64d
f303ccca64d Updated the report size for Ultimate 2 Wireless firmware v1.02
0add037803e [Windows]: dialog: Add `OFN_OVERWRITEPROMPT` to save dialogs (#12782)
a390f5716e6 docs: improve man page generation
93ac1e6894f Fixed non-XBox controllers being detected as Series X controllers
d29bb902f3c Don't treat the Keychron K1 Pro System Control keyboard as a joystick
ab57ef9d7bb update sdlgenblit.pl after PR/12769
b0a0d236d85 fix sdlgenblit.pl so that it actually matches the generated source
0fdfa925f27 rename local pointer vars 'pixel
[…]
