+++ libopensimcreator/Documents/MeshWarper/UndoableTPSDocumentActions.cpp
@@ -273,170 +273,124 @@ void osc::ActionSaveLandmarksToCSV(
     TPSDocumentInputIdentifier which,
     lm::LandmarkCSVFlags flags)
 {
-    const std::optional<std::filesystem::path> maybeCSVPath =
-        prompt_user_for_file_save_location_add_extension_if_necessary("csv");
-    if (!maybeCSVPath)
+    App::upd().prompt_user_to_save_file_with_specific_extension([pairs = doc.landmarkPairs, which, flags](std::filesystem::path p)
     {
-        return;  // user didn't select a save location
-    }
-
-    std::ofstream fout{*maybeCSVPath};
-    if (!fout)
-    {
-        return;  // couldn't open file for writing
-    }
-
-    lm::WriteLandmarksToCSV(fout, [which, it = doc.landmarkPairs.begin(), end = doc.landmarkPairs.end()]() mutable
-    {
-        std::optional<lm::Landmark> rv;
-        for (; !rv && it != end; ++it)
-        {
-            if (auto location = GetLocation(*it, which))
-            {
-                rv = lm::Landmark{std::string{it->name}, *location};
-            }
+        std::ofstream fout{p};
+        if (not fout) {
+            return;  // couldn't open file for writing
         }
-        return rv;
-    }, flags);
+
+        lm::WriteLandmarksToCSV(fout, [which, pairs = std::move(pairs), flags]() mutable
+        {
+            std::optional<lm::Landmark> rv;
+            for (const TPSDocumentLandmarkPair& pair : pairs) {
+                if (const auto location = GetLocation(pair, which)) {
+                    rv = lm::Landmark{std::string{pair.name}, *location};
+                }
+            }
+            return rv;
+        }, flags);
+    }, "csv");
 }
 
 void osc::ActionSaveNonParticipatingLandmarksToCSV(
     const TPSDocument& doc,
     lm::LandmarkCSVFlags flags)
 {
-    const std::optional<std::filesystem::path> maybeCSVPath =
-        prompt_user_for_file_save_location_add_extension_if_necessary("csv");
-    if (!maybeCSVPath)
+    App::upd().prompt_user_to_save_file_with_specific_extension([nplms = doc.nonParticipatingLandmarks, flags](std::filesystem::path p)
     {
-        return;  // user didn't select a save location
-    }
-
-    std::ofstream fout{*maybeCSVPath};
-    if (!fout)
-    {
-        return;  // couldn't open file for writing
-    }
-
-    lm::WriteLandmarksToCSV(fout, [it = doc.nonParticipatingLandmarks.begin(), end = doc.nonParticipatingLandmarks.end()]() mutable
-    {
-        std::optional<lm::Landmark> rv;
-        if (it != end) {
-            rv = lm::Landmark{std::string{it->name}, it->location};
+        std::ofstream fout{p};
+        if (not fout) {
+            return;  // couldn't open file for writing
         }
-        ++it;
-        return rv;
-    }, flags);
+
+        lm::WriteLandmarksToCSV(fout, [it = nplms.begin(), end = nplms.end()]() mutable
+        {
+            std::optional<lm::Landmark> rv;
+            if (it != end) {
+                rv = lm::Landmark{std::string{it->name}, it->location};
+            }
+            ++it;
+            return rv;
+        }, flags);
+    }, "csv");
 }
 
 void osc::ActionSavePairedLandmarksToCSV(const TPSDocument& doc, lm::LandmarkCSVFlags flags)
 {
-    const std::optional<std::filesystem::path> maybeCSVPath =
-        prompt_user_for_file_save_location_add_extension_if_necessary("csv");
-    if (!maybeCSVPath)
[…]
