### 2025-04-10 3009a1d
Remove unused lambda capture
