+++ libopensimcreator/UI/ModelEditor/ModelEditorMainMenu.cpp
@@ -150,7 +150,7 @@ private:
                 ui::draw_tooltip_if_item_hovered("Simulate Against All Integrators", "Simulate the given model against all available SimTK integrators. This takes the current simulation parameters and permutes the integrator, reporting the overall simulation wall-time to the user. It's an advanced feature that's handy for developers to figure out which integrator best-suits a particular model");
 
                 if (ui::draw_menu_item("Export Model Graph as Dotviz")) {
-                    ActionExportModelGraphToDotviz(*m_Model);
+                    ActionExportModelGraphToDotviz(m_Model);
                 }
                 ui::draw_tooltip_if_item_hovered("Writes the model's data topology graph in dotviz format, so that it can be visualized in external tooling such as Graphviz Online");
 
