### 2025-04-10 2285376
Add support for watching OpenSim::Force record outputs
