### 2025-04-10 6708242
Refurbish build_windows.py, add support for --system-version, set SDK to 10.0.26100.0
