+++ scripts/build_windows.py
@@ -20,6 +20,15 @@ def _run(s: str, extra_env_vars={}):
     logging.info(f"running: {s}")
     subprocess.run(s, check=True, env={**extra_env_vars, **os.environ})
 
+def _run_cmake_configure(source_dir, binary_dir, generator, architecture, cache_variables, extra_env_vars={}):
+    all_cache_variables = ' '.join([f'-D{k}={v}' for k, v in cache_variables.items()])
+    maybe_arch_flag = f'-A {architecture}' if 'Visual Studio' in generator else ''
+    _run(f'cmake -S {source_dir} -B {binary_dir} -G {generator} {maybe_arch_flag} {all_cache_variables}', extra_env_vars)
+
+def _run_cmake_build(binary_dir, config, concurrency, target=None):
+    maybe_target_flag = f'--target {target}' if target else ''
+    _run(f'cmake --build {binary_dir} --config {config} -j{concurrency} {maybe_target_flag}')
+
 def _log_dir_contents(path: str):
     logging.info(f"listing {path}")
     for el in os.listdir(path):
@@ -35,6 +44,7 @@ class BuildConfiguration:
             build_concurrency=multiprocessing.cpu_count(),
             build_target="package",
             build_docs="OFF",
+            system_version=None,
             build_skip_osc=False):
 
         self.base_build_type = os.getenv("OSC_BASE_BUILD_TYPE", base_build_type)
@@ -43,7 +53,9 @@ class BuildConfiguration:
         self.concurrency = int(os.getenv("OSC_BUILD_CONCURRENCY", build_concurrency))
         self.build_target = os.getenv("OSC_BUILD_TARGET", build_target)
         self.build_docs = _is_truthy_envvar(os.getenv("OSC_BUILD_DOCS", build_docs))
-        self.generator_flags = f'-G"Visual Studio 17 2022" -A x64'
+        self.generator = 'Visual Studio 17 2022'
+        self.architecture = 'x64'
+        self.system_version = os.getenv('OSC_SYSTEM_VERSION', system_version)
         self.build_dir = build_dir
         self.skip_osc = build_skip_osc
 
@@ -79,12 +91,31 @@ def log_build_params(conf: BuildConfiguration):
 
 def build_osc_dependencies(conf: BuildConfiguration):
     with Section("build osc dependencies"):
-        _run(
-            f'cmake -S third_party/ -B "{conf.get_dependencies_build_dir()}" {conf.generator_flags} -DCMAKE_BUILD_TYPE={conf.get_osc_deps_build_type()} -DCMAKE_INSTALL_PREFIX="{conf.get_dependencies_install_dir()}" -DOSCDEPS_BUILD_OPENBLAS=ON',
+        # calculate cache variables
+        cache_variables = {
+            'CMAKE_BUILD_TYPE': conf.get_osc_deps_build_type(),
+            'CMAKE_INSTALL_PREFIX': conf.get_dependencies_install_dir(),
+            'OSCDEPS_BUILD_OPENBLAS': 'ON'
+        }
+        if conf.system_version:
+            cache_variables['CMAKE_SYSTEM_VERSION'] = conf.system_version
+
+        # configure
+        _run_cmake_configure(
+            source_dir='third_party',
+            binary_dir=conf.get_dependencies_build_dir(),
+            generator=conf.generator,
+            architecture=conf.architecture,
+            cache_variables=cache_variables,
             # note: this is necessary because OpenSim transitively uses `spdlog`, which has deprecated this
-            extra_env_vars={'CXXFLAGS': '/D_SILENCE_STDEXT_ARR_ITERS_DEPRECATION_WARNING'},
+            extra_env_vars={'CXXFLAGS': '/D_SILENCE_STDEXT_ARR_ITERS_DEPRECATION_WARNING'}
         )
[…]
