+++ .github/workflows/continuous-integration-workflow.yml
@@ -5,8 +5,8 @@ env:
 
 jobs:
   windows2022msvc:
-    name: Windows 2025 (amd64, MSVC)
-    runs-on: windows-2025
+    name: Windows 2022 (amd64, MSVC)
+    runs-on: windows-2022
     steps:
 
     - uses: actions/checkout@v4
@@ -19,7 +19,10 @@ jobs:
         cd $env:GITHUB_WORKSPACE\osc
 
         # run build script
-        python scripts\\build_windows.py -j2 || exit /b
+        #
+        # --system-version is necessary with `windows-2022` runner because the
+        # default SDK doesn't include `gameinput.h` correctly (see: https://github.com/actions/runner-images/issues/10980)
+        python scripts\\build_windows.py --system-version=10.0.26100.0 -j2 || exit /b
 
     - uses: actions/upload-artifact@v4
       with:
