### 2025-04-10 1dfc77e
Add escapes to generator string
