### 2025-04-16 #8006 @davidpagnon
Not sure it is appropriate to do so, but I just saw my name here, and if needed, I'm willing to review this paper.
