### 2025-04-20 605dc60
Clean up scripts
