+++ scripts/setup_ubuntu-20.04.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+# `setup_ubuntu-20.04.sh`: sets up an Ubuntu 20.04 machine
+# with the necessary dependencies to build OpenSim Creator.
+
 # setup system dependencies
 sudo apt install -y cmake pkg-config libgtk-3-dev libblas-dev liblapack-dev clang-11 clang-tidy-11 libstdc++-10-dev xdg-desktop-portal-gtk
 
@@ -11,4 +14,3 @@ cd cmake-3.31.2
 CXX=clang++-11 CC=clang-11 ./bootstrap
 CXX=clang++-11 CC=clang-11 make -j$(nproc) && sudo make install
 cd -
-
