+++ scripts/macos_check-sdk.py
@@ -15,4 +15,3 @@ args = parser.parse_args()
 
 sdk_version = subprocess.check_output(f'otool -l {args.binary} | grep sdk', shell=True).decode('utf-8').split(' ')[-1].strip()
 assert sdk_version == args.expected_sdk_version, f'{sdk_version} is not the expected SDK version ({args.expected_sdk_version})'
-
