+++ scripts/build_mac-arm64.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+# `build_mac-arm64.sh`: performs an end-to-end build of OpenSim Creator for
+# arm64 (i.e. Apple Silicon) MacOS systems.
+
 export OSC_CMAKE_CONFIG_EXTRA="-Werror=dev -DCMAKE_OSX_SYSROOT=/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX14.2.sdk -DCMAKE_OSX_DEPLOYMENT_TARGET=14 -DCMAKE_OSX_ARCHITECTURES=arm64"
 export OSC_BUILD_CONCURRENCY=$(sysctl -n hw.ncpu)
 ./scripts/build_mac.sh
