+++ scripts/env_ubuntu.sh
@@ -1,4 +1,3 @@
 # a session DBUS must be launched in order for the
 # file dialogs to work
 eval $(dbus-launch --sh-syntax)
-
