+++ libopensimcreator/UI/ModelEditor/ModelMusclePlotPanel.cpp
@@ -935,26 +935,23 @@ namespace
         }
     }
 
-    void TrySavePlotToCSV(const OpenSim::Coordinate& coord, const PlotParameters& params, const Plot& plot, const std::filesystem::path& outPath)
+    void WritePlotAsCSV(
+        const OpenSim::Coordinate& coord,
+        const PlotParameters& params,
+        const Plot& plot,
+        std::ostream& out)
     {
-        std::ofstream fileOutputStream{outPath};
-        if (!fileOutputStream)
-        {
-            return;  // error opening outfile
-        }
-
         // write header
         write_csv_row(
-            fileOutputStream,
+            out,
             std::to_array({ ComputePlotXAxisTitle(params, coord), ComputePlotYAxisTitle(params) })
         );
 
         // write data rows
         auto lock = plot.lockDataPoints();
-        for (const PlotDataPoint& p : *lock)
-        {
+        for (const PlotDataPoint& p : *lock) {
             write_csv_row(
-                fileOutputStream,
+                out,
                 std::to_array({ std::to_string(p.x), std::to_string(p.y) })
             );
         }
@@ -962,13 +959,21 @@ namespace
 
     void ActionPromptUserToSavePlotToCSV(const OpenSim::Coordinate& coord, const PlotParameters& params, const Plot& plot)
     {
-        const std::optional<std::filesystem::path> maybeCSVPath =
-            prompt_user_for_file_save_location_add_extension_if_necessary("csv");
+        // Pre-write the CSV in memory so that the asynchronous user request isn't dependent on
+        // any UI state.
+        std::stringstream ss;
+        WritePlotAsCSV(coord, params, plot, ss);
 
-        if (maybeCSVPath)
+        // Asynchronously request the save location from the user and write it
+        App::upd().prompt_user_to_save_file_with_specific_extension([csv = std::move(ss).str()](std::filesystem::path p)
         {
-            TrySavePlotToCSV(coord, params, plot, *maybeCSVPath);
-        }
+            std::ofstream ofs{p};
+            if (not ofs) {
+                log_error("%s: cannot open path for writing", p.string().c_str());
+                return;
+            }
+            ofs << csv;
+        }, "csv");
     }
 
     // holds a collection of plotlines that are to-be-drawn on the plot
@@ -1320,24 +1325,18 @@ namespace
         return it != cursors.end() ? it->peekX() : std::optional<float>{};
     }
 
-    // try to save the given collection of plotlines to an on-disk CSV file
+    // tries to write the given collection of plotlines in a CSV format to the output stream.
     //
     // the resulting CSV may be sparsely populated, because each line may have a different
-    // number of, and location of, values
-    void TrySavePlotLinesToCSV(
+    // number of, and location of, values.
+    void TryWritePlotLinesAsCSV(
         const OpenSim::Coordinate& coord,
         const PlotParameters& params,
         const PlotLines& lines,
-        const std::filesystem::path& outPath)
+        std::ostream& out)
     {
-        std::ofstream outputFileStream{outPath};
-        if (!outputFileStream)
-        {
-            return;  // error opening outfile
-        }
-
         // write header
-        write_csv_row(outputFileStream, GetAllCSVHeaders(coord, params, lines));
+        write_csv_row(out, GetAllCSVHeaders(coord, params, lines));
 
         // get incrementable cursors to all curves in the plot
         std::vector<LineCursor> cursors = GetCursorsToAllPlotLines(lines);
@@ -1348,8 +1347,7 @@ namespace
         // keep an eye out for the *next* lowest X value as we iterate
         std::optional<float> maybeNextX;
 
-        while (maybeX)
-        {
+        while (maybeX) {
             std::vector<std::string> cols;
[…]
