+++ libopensimcreator/UI/Simulation/ISimulatorUIAPI.cpp
@@ -4,6 +4,7 @@
 #include <libopensimcreator/Documents/Simulation/ISimulation.h>
 #include <libopensimcreator/Documents/Simulation/SimulationHelpers.h>
 
+#include <liboscar/Platform/App.h>
 #include <liboscar/Platform/Log.h>
 #include <liboscar/Platform/os.h>
 #include <liboscar/Utils/Assertions.h>
@@ -13,46 +14,56 @@
 #include <fstream>
 #include <optional>
 #include <span>
-#include <string>
+#include <sstream>
 #include <string_view>
+#include <string>
+#include <utility>
 
 using namespace osc;
 
 namespace
 {
-    std::optional<std::filesystem::path> TryExportOutputsToCSV(
+    void TryExportOutputsToCSV(
         const ISimulation& simulation,
-        std::span<const OutputExtractor> outputs)
+        std::span<const OutputExtractor> outputs,
+        bool openInDefaultApp)
     {
-        // prompt user for save location
-        std::optional<std::filesystem::path> path =
-            prompt_user_for_file_save_location_add_extension_if_necessary("csv");
-        if (not path) {
-            return std::nullopt;  // user probably cancelled out
+        // Pre-write the CSV in-memory so that the asynchronous user prompt isn't
+        // dependent on a bunch of state.
+        std::stringstream ss;
+        {
+            const auto guard = simulation.getModel();
+            WriteOutputsAsCSV(*guard, outputs, simulation.getAllSimulationReports(), ss);
         }
 
-        // open output file
-        std::ofstream fout{*path};
-        if (not fout) {
-            log_error("%s: error opening file for writing", path->string().c_str());
-            return std::nullopt;  // error opening output file for writing
-        }
-        fout.exceptions(std::ios_base::badbit | std::ios_base::failbit);
+        // Asynchronously prompt the user to select a save location and write the CSV
+        // to it. If requested, open it in the user's default application.
+        App::upd().prompt_user_to_save_file_with_specific_extension([openInDefaultApp, csv = std::move(ss).str()](std::filesystem::path p)
+        {
+            // Open+write output file
+            {
+                std::ofstream ofs{p};
+                if (not ofs) {
+                    log_error("%s: error opening file for writing", p.string().c_str());
+                    return;  // error opening output file for writing
+                }
 
-        // write output
-        const auto guard = simulation.getModel();
-        WriteOutputsAsCSV(*guard, outputs, simulation.getAllSimulationReports(), fout);
+                ofs << csv;
+            }
 
-        return path;
+            if (openInDefaultApp) {
+                open_file_in_os_default_application(p);
+            }
+        }, "csv");
     }
 }
 
-std::optional<std::filesystem::path> osc::ISimulatorUIAPI::tryPromptToSaveOutputsAsCSV(std::span<const OutputExtractor> outputs) const
+void osc::ISimulatorUIAPI::tryPromptToSaveOutputsAsCSV(std::span<const OutputExtractor> outputs, bool openInDefaultApp) const
 {
-    return TryExportOutputsToCSV(getSimulation(), outputs);
+    TryExportOutputsToCSV(getSimulation(), outputs, openInDefaultApp);
 }
 
-std::optional<std::filesystem::path> osc::ISimulatorUIAPI::tryPromptToSaveAllOutputsAsCSV(std::span<const OutputExtractor> outputs) const
+void osc::ISimulatorUIAPI::tryPromptToSaveAllOutputsAsCSV(std::span<const OutputExtractor> outputs, bool openInDefaultApp) const
 {
-    return TryExportOutputsToCSV(getSimulation(), outputs);
+    TryExportOutputsToCSV(getSimulation(), outputs, openInDefaultApp);
 }
