### 2025-04-21 2d3efef
Drop Ubuntu 20.04 from github CI
