+++ .github/workflows/continuous-integration-workflow.yml
@@ -30,41 +30,6 @@ jobs:
         path: osc/osc-build/*windows-amd64.exe
         if-no-files-found: error
 
-  ubuntu2004clang:
-    name: Ubuntu 20.04 (amd64, clang-11)
-    runs-on: ubuntu-20.04
-    steps:
-
-    - uses: actions/checkout@v4
-      with:
-        path: osc
-        submodules: 'recursive'
-
-    - name: setup system
-      run: |
-        # install system dependencies
-        sudo apt update
-        sudo apt-get install -y build-essential cmake libgtk-3-dev liblapack-dev
-        sudo apt-get install -y xvfb  # virtual desktop, for UI tests
-
-    - name: run build script
-      run: |
-        # set compiler, concurrency, etc.
-        export OSC_BUILD_CONCURRENCY=$(nproc)
-        export CC=clang
-        export CXX=clang++
-
-        cd $GITHUB_WORKSPACE/osc
-
-        # run buildscript under virtual desktop (for UI tests)
-        xvfb-run ./scripts/build_ubuntu-20.04.sh
-
-    - uses: actions/upload-artifact@v4
-      with:
-        name: opensimcreator-debian-amd64.zip
-        path: osc/osc-build/*.deb
-        if-no-files-found: error
-
   macos11clang:
     name: MacOS 13 (amd64, XCode 15.2, targeting 14.2)
     runs-on: macos-13
