+++ libopensimcreator/UI/ModelEditor/ModelEditorTab.h
@@ -1,7 +1,10 @@
 #pragma once
 
 #include <liboscar/UI/Tabs/Tab.h>
+#include <liboscar/UI/Tabs/TabSaveResult.h>
 #include <liboscar/Utils/CStringView.h>
+
+#include <future>
 
 namespace OpenSim { class Model; }
 namespace osc { class Widget; }
@@ -32,7 +35,7 @@ namespace osc
 
     private:
         bool impl_is_unsaved() const final;
-        bool impl_try_save() final;
+        std::future<TabSaveResult> impl_try_save() final;
         void impl_on_mount() final;
         void impl_on_unmount() final;
         bool impl_on_event(Event&) final;

[…]
