### 2025-04-22 526e30e
Make dialog cancellation explicit in prompt_user_to_save_file_async
