+++ .github/workflows/continuous-integration-workflow.yml
@@ -70,7 +70,7 @@ jobs:
       run: |
         # install system dependencies
         sudo apt update
-        sudo apt-get install -y build-essential cmake libgtk-3-dev liblapack-dev
+        sudo apt-get install -y build-essential cmake liblapack-dev
         sudo apt-get install -y xvfb  # virtual desktop, for tests
 
     - name: run build script
