### 2025-04-22 9ef08a4
Fix Ubuntu dependencies
