+++ scripts/setup_ubuntu-24.04.sh
@@ -5,4 +5,4 @@
 
 # WSL2 on Ubuntu24: ln -sf  /mnt/wslg/runtime-dir/wayland-* $XDG_RUNTIME_DIR/
 
-sudo apt-get install -y clang clang-tidy cmake pkg-config libblas-dev liblapack-dev xdg-desktop-portal-gtk
+sudo apt-get install -y clang clang-tidy cmake pkg-config libblas-dev liblapack-dev xdg-desktop-portal-gtk libopengl-dev libgl1-mesa-dev libx11-dev libxext-dev
