+++ .github/workflows/continuous-integration-workflow.yml
@@ -70,7 +70,7 @@ jobs:
       run: |
         # install system dependencies
         sudo apt update
-        sudo apt-get install -y build-essential cmake liblapack-dev
+        sudo apt-get install -y build-essential cmake liblapack-dev libopengl-dev libgl1-mesa-dev libx11-dev libxext-dev
         sudo apt-get install -y xvfb  # virtual desktop, for tests
 
     - name: run build script
