### 2025-04-22 a845445
Move get/set_initial_directory_to_show to osc::App, rather than a global
