+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -47,24 +47,16 @@
 // NOLINTBEGIN
 
 #ifndef IMGUI_DEFINE_MATH_OPERATORS
-#define IMGUI_DEFINE_MATH_OPERATORS
+    #define IMGUI_DEFINE_MATH_OPERATORS
 #endif
 #include "imgui.h"
 #include "imgui_internal.h"
 #include "ImGuizmo.h"
 
-#if defined(_MSC_VER) || defined(__MINGW32__)
-#include <malloc.h>
-#endif
-#if !defined(_MSC_VER) && !defined(__MINGW64_VERSION_MAJOR)
-#define _malloca(x) alloca(x)
-#define _freea(x)
-#endif
-
 #include <limits>
+#include <vector>
 
-// includes patches for multiview from
-// https://github.com/CedricGuillemet/ImGuizmo/issues/15
+using namespace ImGuizmo;
 
 namespace
 {
@@ -72,37 +64,34 @@ namespace
     {
         return std::numeric_limits<ImGuiID>::max();
     }
-}
 
-namespace ImGuizmo
-{
-   static const float ZPI = 3.14159265358979323846f;
-   static const float RAD2DEG = (180.f / ZPI);
-   static const float DEG2RAD = (ZPI / 180.f);
-   const float screenRotateSize = 0.06f;
-   // scale a bit so translate axis do not touch when in universal
-   const float rotationDisplayFactor = 1.2f;
+    const float ZPI = 3.14159265358979323846f;
+    const float RAD2DEG = (180.f / ZPI);
+    const float DEG2RAD = (ZPI / 180.f);
+    const float screenRotateSize = 0.06f;
+    // scale a bit so translate axis do not touch when in universal
+    const float rotationDisplayFactor = 1.2f;
 
-   static OPERATION operator&(OPERATION lhs, OPERATION rhs)
-   {
-     return static_cast<OPERATION>(static_cast<int>(lhs) & static_cast<int>(rhs));
-   }
+    OPERATION operator&(OPERATION lhs, OPERATION rhs)
+    {
+        return static_cast<OPERATION>(static_cast<int>(lhs) & static_cast<int>(rhs));
+    }
 
-   static bool operator!=(OPERATION lhs, int rhs)
-   {
-     return static_cast<int>(lhs) != rhs;
-   }
+    bool operator!=(OPERATION lhs, int rhs)
+    {
+        return static_cast<int>(lhs) != rhs;
+    }
 
-   static bool Intersects(OPERATION lhs, OPERATION rhs)
-   {
-     return (lhs & rhs) != 0;
-   }
+    bool Intersects(OPERATION lhs, OPERATION rhs)
+    {
+        return (lhs & rhs) != 0;
+    }
 
-   // True if lhs contains rhs
-   static bool Contains(OPERATION lhs, OPERATION rhs)
-   {
-     return (lhs & rhs) == rhs;
-   }
+    // True if lhs contains rhs
+    bool Contains(OPERATION lhs, OPERATION rhs)
+    {
+        return (lhs & rhs) == rhs;
+    }
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
    // utility and math
@@ -221,8 +210,7 @@ namespace ImGuizmo
    template <typename T> bool IsWithin(T x, T y, T z) { return (x >= y) && (x <= z); }
 
    struct matrix_t;
-   struct vec_t
-   {
+   struct vec_t {
    public:
       float x = 0.0f, y = 0.0f, z = 0.0f, w = 0.0f;
 
@@ -335,13 +323,11 @@ namespace ImGuizmo
[…]
