+++ liboscar-demos/ImGuizmoDemoTab.cpp
@@ -13,7 +13,9 @@ public:
 
     explicit Impl(ImGuizmoDemoTab& owner, Widget* parent) :
         TabPrivate{owner, parent, static_label()}
-    {}
+    {
+        basic_material_.set_transparent(true);
+    }
 
     void on_draw()
     {
@@ -21,8 +23,40 @@ public:
         const Rect viewport_ui_rect = ui::get_main_viewport_workspace_uiscreenspace_rect();
         const Mat4 projection_matrix = scene_camera_.projection_matrix(aspect_ratio_of(viewport_ui_rect));
 
-        ui::gizmo_demo_draw_grid(identity<Mat4>(), view_matrix, projection_matrix, 100.0f, viewport_ui_rect);
-        ui::gizmo_demo_draw_cube(model_matrix_, view_matrix, projection_matrix, viewport_ui_rect);
+        // Render 3D scene: a grid floor and a cube that has a different color per face
+        {
+            Camera render_camera;
+            render_camera.set_view_matrix_override(view_matrix);
+            render_camera.set_projection_matrix_override(projection_matrix);
+            render_camera.set_pixel_rect(viewport_ui_rect);
+
+            for (size_t i = 0; i < 6; ++i) {
+                // axis-aligned vector
+                Vec3 v;
+                v[i % 3] = i/3 ? -1.0f : 1.0f;
+
+                const Mat4 xform = model_matrix_ * translate(identity<Mat4>(), 0.5f*v) * mat4_cast(rotation(plane_.normal(), v));
+                const Color color = Color{0.4f}.with_element(i % 3, 0.8f);
+                graphics::draw(
+                    plane_,
+                    xform,
+                    basic_material_,
+                    render_camera,
+                    MeshBasicMaterial::PropertyBlock{color}
+                );
+            }
+            basic_material_.set_color(Color::white());
+            graphics::draw(
+                grid_,
+                {.rotation = rotation(grid_.normal(), {0.0f, 1.0f, 0.0f})},
+                basic_material_,
+                render_camera,
+                MeshBasicMaterial::PropertyBlock{Color::white().with_alpha(0.1f)}
+            );
+            render_camera.render_to_screen();
+        }
+
+        // Draw UI overlays (incl. gizmo)
         gizmo_.draw_to_foreground(
             model_matrix_,
             view_matrix,
@@ -46,6 +80,9 @@ private:
 
     ui::Gizmo gizmo_;
     Mat4 model_matrix_ = identity<Mat4>();
+    GridGeometry grid_{{.size = 20.0f, .num_divisions = 100}};
+    PlaneGeometry plane_;
+    MeshBasicMaterial basic_material_;
 };
 
 
