+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -53,6 +53,7 @@
 #include "imgui_internal.h"
 #include "ImGuizmo.h"
 
+#include <functional>
 #include <limits>
 #include <vector>
 
@@ -60,17 +61,55 @@ using namespace ImGuizmo;
 
 namespace
 {
+    struct Context;
+    constinit Context* gCurrentContext = nullptr;
+
+    constexpr float ZPI = 3.14159265358979323846f;
+    constexpr float RAD2DEG = (180.f / ZPI);
+    constexpr float DEG2RAD = (ZPI / 180.f);
+    constexpr float screenRotateSize = 0.06f;
+    // scale a bit so translate axis do not touch when in universal
+    constexpr float rotationDisplayFactor = 1.2f;
+    const char* translationInfoMask[] = {
+        "X : %5.3f",
+        "Y : %5.3f",
+        "Z : %5.3f",
+        "Y : %5.3f Z : %5.3f",
+        "X : %5.3f Z : %5.3f",
+        "X : %5.3f Y : %5.3f",
+        "X : %5.3f Y : %5.3f Z : %5.3f"
+    };
+    const char* const scaleInfoMask[] = {
+        "X : %5.2f",
+        "Y : %5.2f",
+        "Z : %5.2f",
+        "XYZ : %5.2f"
+    };
+    const char* const rotationInfoMask[] = {
+        "X : %5.2f deg %5.2f rad",
+        "Y : %5.2f deg %5.2f rad",
+        "Z : %5.2f deg %5.2f rad",
+        "Screen : %5.2f deg %5.2f rad"
+    };
+    constexpr int translationInfoIndex[] = {
+        0,0,0,
+        1,0,0,
+        2,0,0,
+        1,2,0,
+        0,2,0,
+        0,1,0,
+        0,1,2
+    };
+    constexpr float quadMin = 0.5f;
+    constexpr float quadMax = 0.8f;
+    constexpr float quadUV[8] = { quadMin, quadMin, quadMin, quadMax, quadMax, quadMax, quadMax, quadMin };
+    constexpr int halfCircleSegmentCount = 64;
+    constexpr float snapTension = 0.5f;
+
     constexpr ImGuiID blank_id()
     {
         return std::numeric_limits<ImGuiID>::max();
     }
-
-    const float ZPI = 3.14159265358979323846f;
-    const float RAD2DEG = (180.f / ZPI);
-    const float DEG2RAD = (ZPI / 180.f);
-    const float screenRotateSize = 0.06f;
-    // scale a bit so translate axis do not touch when in universal
-    const float rotationDisplayFactor = 1.2f;
 
     OPERATION operator&(OPERATION lhs, OPERATION rhs)
     {
@@ -295,6 +334,8 @@ namespace
    vec_t vec_t::operator + (const vec_t& v) const { return makeVect(x + v.x, y + v.y, z + v.z, w + v.w); }
    vec_t vec_t::operator * (const vec_t& v) const { return makeVect(x * v.x, y * v.y, z * v.z, w * v.w); }
    vec_t vec_t::Abs() const { return makeVect(fabsf(x), fabsf(y), fabsf(z)); }
+
+   const vec_t directionUnary[3] = { makeVect(1.f, 0.f, 0.f), makeVect(0.f, 1.f, 0.f), makeVect(0.f, 0.f, 1.f) };
 
    vec_t Normalized(const vec_t& v) { vec_t res; res = v; res.Normalize(); return res; }
    vec_t Cross(const vec_t& v1, const vec_t& v2)
@@ -796,55 +837,15 @@ namespace
       // check to not have multiple gizmo highlighted at the same time
[…]
