+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -2453,25 +2453,6 @@ namespace
        ImGuiID id = ImHashData(&n, sizeof(n), seed);
        return id;
    }
-   ImGuiID GetID(const char* str, const char* str_end)
-   {
-       ImGuiID seed = gCurrentContext->mIDStack.back();
-       ImGuiID id = ImHashStr(str, str_end ? (str_end - str) : 0, seed);
-       return id;
-   }
-
-   ImGuiID GetID(const char* str)
-   {
-       return GetID(str, nullptr);
-   }
-
-   ImGuiID GetID(const void* ptr)
-   {
-       ImGuiID seed = gCurrentContext->mIDStack.back();
-       ImGuiID id = ImHashData(&ptr, sizeof(void*), seed);
-       return id;
-   }
-
 }
 
 void ImGuizmo::CreateContext()
@@ -2491,7 +2472,14 @@ void ImGuizmo::SetDrawlist(ImDrawList* drawlist)
 
 void ImGuizmo::BeginFrame()
 {
-    const ImU32 flags = ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoInputs | ImGuiWindowFlags_NoSavedSettings | ImGuiWindowFlags_NoFocusOnAppearing | ImGuiWindowFlags_NoBringToFrontOnFocus;
+    const ImU32 flags =
+        ImGuiWindowFlags_NoTitleBar |
+        ImGuiWindowFlags_NoResize |
+        ImGuiWindowFlags_NoScrollbar |
+        ImGuiWindowFlags_NoInputs |
+        ImGuiWindowFlags_NoSavedSettings |
+        ImGuiWindowFlags_NoFocusOnAppearing |
+        ImGuiWindowFlags_NoBringToFrontOnFocus;
 
 #ifdef IMGUI_HAS_VIEWPORT
     ImGui::SetNextWindowSize(ImGui::GetMainViewport()->Size);
@@ -2517,52 +2505,43 @@ void ImGuizmo::BeginFrame()
 bool ImGuizmo::IsOver()
 {
     return
-        (Intersects(gCurrentContext->mOperation, TRANSLATE) && GetMoveType(gCurrentContext->mOperation, NULL) != MT_NONE) ||
-        (Intersects(gCurrentContext->mOperation, ROTATE)    && GetRotateType(gCurrentContext->mOperation) != MT_NONE) ||
-        (Intersects(gCurrentContext->mOperation, SCALE)     && GetScaleType(gCurrentContext->mOperation) != MT_NONE) ||
+        (Intersects(gCurrentContext->mOperation, TRANSLATE) and GetMoveType(gCurrentContext->mOperation, NULL) != MT_NONE) or
+        (Intersects(gCurrentContext->mOperation, ROTATE)    and GetRotateType(gCurrentContext->mOperation) != MT_NONE)     or
+        (Intersects(gCurrentContext->mOperation, SCALE)     and GetScaleType(gCurrentContext->mOperation) != MT_NONE)      or
         IsUsing();
 }
 
 bool ImGuizmo::IsOver(OPERATION op)
 {
-    if(IsUsing()) {
+    if (IsUsing()) {
         return true;
     }
-    if(Intersects(op, SCALE) && GetScaleType(op) != MT_NONE) {
+    if (Intersects(op, SCALE) and GetScaleType(op) != MT_NONE) {
         return true;
     }
-    if(Intersects(op, ROTATE) && GetRotateType(op) != MT_NONE) {
+    if (Intersects(op, ROTATE) and GetRotateType(op) != MT_NONE) {
[…]
