### 2025-04-23 e0bfb8e
Apply fixes for stuff found by clang-tidy
