+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -65,7 +65,6 @@ namespace
     constinit Context* gCurrentContext = nullptr;
 
     constexpr float ZPI = 3.14159265358979323846f;
-    constexpr float RAD2DEG = (180.f / ZPI);
     constexpr float DEG2RAD = (ZPI / 180.f);
     constexpr float screenRotateSize = 0.06f;
     // scale a bit so translate axis do not touch when in universal
@@ -156,91 +155,6 @@ namespace
       r[13] = a[12] * b[1] + a[13] * b[5] + a[14] * b[9] + a[15] * b[13];
       r[14] = a[12] * b[2] + a[13] * b[6] + a[14] * b[10] + a[15] * b[14];
       r[15] = a[12] * b[3] + a[13] * b[7] + a[14] * b[11] + a[15] * b[15];
-   }
-
-   void Frustum(float left, float right, float bottom, float top, float znear, float zfar, float* m16)
-   {
-      float temp, temp2, temp3, temp4;
-      temp = 2.0f * znear;
-      temp2 = right - left;
-      temp3 = top - bottom;
-      temp4 = zfar - znear;
-      m16[0] = temp / temp2;
-      m16[1] = 0.0;
-      m16[2] = 0.0;
-      m16[3] = 0.0;
-      m16[4] = 0.0;
-      m16[5] = temp / temp3;
-      m16[6] = 0.0;
-      m16[7] = 0.0;
-      m16[8] = (right + left) / temp2;
-      m16[9] = (top + bottom) / temp3;
-      m16[10] = (-zfar - znear) / temp4;
-      m16[11] = -1.0f;
-      m16[12] = 0.0;
-      m16[13] = 0.0;
-      m16[14] = (-temp * zfar) / temp4;
-      m16[15] = 0.0;
-   }
-
-   void Perspective(float fovyInDegrees, float aspectRatio, float znear, float zfar, float* m16)
-   {
-      float ymax, xmax;
-      ymax = znear * tanf(fovyInDegrees * DEG2RAD);
-      xmax = ymax * aspectRatio;
-      Frustum(-xmax, xmax, -ymax, ymax, znear, zfar, m16);
-   }
-
-   void Cross(const float* a, const float* b, float* r)
-   {
-      r[0] = a[1] * b[2] - a[2] * b[1];
-      r[1] = a[2] * b[0] - a[0] * b[2];
-      r[2] = a[0] * b[1] - a[1] * b[0];
-   }
-
-   float Dot(const float* a, const float* b)
-   {
-      return a[0] * b[0] + a[1] * b[1] + a[2] * b[2];
-   }
-
-   void Normalize(const float* a, float* r)
-   {
-      float il = 1.f / (sqrtf(Dot(a, a)) + FLT_EPSILON);
-      r[0] = a[0] * il;
-      r[1] = a[1] * il;
-      r[2] = a[2] * il;
-   }
-
-   void LookAt(const float* eye, const float* at, const float* up, float* m16)
-   {
-      float X[3], Y[3], Z[3], tmp[3];
-
-      tmp[0] = eye[0] - at[0];
-      tmp[1] = eye[1] - at[1];
-      tmp[2] = eye[2] - at[2];
-      Normalize(tmp, Z);
-      Normalize(up, Y);
-      Cross(Y, Z, tmp);
-      Normalize(tmp, X);
-      Cross(Z, X, tmp);
-      Normalize(tmp, Y);
-
-      m16[0] = X[0];
-      m16[1] = Y[0];
-      m16[2] = Z[0];
-      m16[3] = 0.0f;
-      m16[4] = X[1];
-      m16[5] = Y[1];
[…]
