### 2025-04-24 5b2a5b7
Fix ConcatenatingOutputExtractor not concatenating X against Y (#1025)
