+++ libopensimcreator/Documents/CustomComponents/BlankComponent.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <OpenSim/Common/Component.h>
+
+namespace osc
+{
+    class BlankComponent : public OpenSim::Component {
+        OpenSim_DECLARE_CONCRETE_OBJECT(BlankComponent, OpenSim::Component);
+    };
+}
