### 2025-04-24 b23133f
Squashed 'third_party/imgui/' changes from 11b3a7c8ca2..da3baddf9d2
da3baddf9d2 Patch dockspace viewport label for backwards compat with opensim-creator imgui.ini configs
198718bc3e7 Add CMakeLists.txt that's suitable for opensim-creator
4806a1924ff Merge branch 'master' into docking
f5befd2d29e Version 1.91.9b
cfed7a3a544 Tables: fixed assert/issues loading settings. (#8496, #7934)
93b446a9628 Backtrack of version tagging 1.92.0 > 1.91.9b for publishing a hotfix (#8496)
a7dc1847728 Moved ImDrawIdx definition lower in imgui.h. Moved ImTextureID slightly below in its own section.
bfaef8ca178 Ve
[…]
