### 2025-04-24 9fa5eba
Fix 'is visible' not toggling visibility due to ImGui regression (#1028)
