### 2025-04-24 421b3be
Fix visualize plane contact forces only works if the plane is the first element in HCF (#1026)
