### 2025-04-24 caa7d6c
Remove trailing semicolon
