+++ libopensimcreator/Documents/CustomComponents/BlankComponent.h
@@ -5,6 +5,6 @@
 namespace osc
 {
     class BlankComponent : public OpenSim::Component {
-        OpenSim_DECLARE_CONCRETE_OBJECT(BlankComponent, OpenSim::Component);
+        OpenSim_DECLARE_CONCRETE_OBJECT(BlankComponent, OpenSim::Component)
     };
 }
