### 2025-04-24 c6e685c
Make camera auto-focus only focus visible geometry (#1029)
