### 2025-04-24 87e2971
Bump version to 0.5.22 (#1031)
