### 2025-04-24 eeb8672
Add artifact upload for ubuntu-22.04 build
