### 2025-04-24 de5b331
Drop cpp20::bit_cast shim
