+++ liboscar/Shims.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <liboscar/Shims/Cpp20/bit.h>
 #include <liboscar/Shims/Cpp20/stop_token.h>
 #include <liboscar/Shims/Cpp20/thread.h>
 #include <liboscar/Shims/Cpp23/cstddef.h>
