### 2025-04-24 f279298
Add rudimentary working generator
