### 2025-04-24 5ce158d
Add alignas on byte array that will be casted to a float
