### 2025-04-25 08ed022
Update CHANGELOG.md for 0.5.22 (#1031)
