### 2025-04-25 069ad7f
Bump version to 0.5.23-dev (#1031)
