### 2025-04-25 87bf96d
Revert version to 0.5.22 (dev is invalid, #1031)
