### 2025-04-25 0b67292
Update CITATION.cff, README.md, and codemeta.json (#1031)
