+++ CITATION.cff
@@ -24,8 +24,8 @@ authors:
     orcid: 'https://orcid.org/0000-0003-4217-1580'
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.14951131
-    description: Zenodo entry containing 0.5.21 binaries
+    value: 10.5281/zenodo.15282275
+    description: Zenodo entry containing 0.5.22 binaries
   - type: url
     value: 'https://www.opensimcreator.com'
     description: Project URL
@@ -43,6 +43,6 @@ keywords:
   - Mechanical Modelling
   - Simulation Software
 license: Apache-2.0
-commit: a24769c54d60ccaaffbc2ddb4f2e515e84929478
-version: 0.5.21
-date-released: '2025-03-01'
+commit: 08ed02215d8d1a3d57ae8a358a43dee974223a26
+version: 0.5.22
+date-released: '2025-04-25'
