### 2025-04-28 a47aaec
Drop almost all mentions/uses of Ubuntu 20.04
