+++ scripts/env_ubuntu.sh
@@ -1,3 +1,5 @@
-# a session DBUS must be launched in order for the
-# file dialogs to work
+# Sets up an Ubuntu terminal with the necessary environment to run
+# OpenSim Creator correctly.
+
+# A session DBUS must be launched in order for the file dialogs to work.
 eval $(dbus-launch --sh-syntax)
