+++ scripts/generate_icon_files.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 
-# `generate_icon_files.py`: generates icon files from images that are compatible with
-# Mac's (icns) and Windows' (ico) icon systems
+# Generates icon files from images that are compatible with Mac's (icns) and
+# Windows' (ico) icon systems.
 #
-# usage:
+# WALKTHROUGH
 #
 # - open raw icon SVG in editor (e.g. inkscape)
 # - save the raw icon SVG as a high (e.g. 1024x1024) resolution PNG file (e.g. logoBIG.png)
@@ -20,7 +20,7 @@
 
 import argparse
 import os
-from PIL import Image  # `pip install Pillow`
+from PIL import Image  # must be installed via pip: `pip install Pillow`
 
 def convert(input_filename, output_filename):
     assert os.path.exists(input_filename)
