### 2025-04-29 f593b35
Add [[maybe_unused]] for Ubuntu24.04 compilation
