### 2025-05-01 2d1c3fe
Rename osc-build/ to build/ for easier clangd compatibility
