+++ docs/source/setup-development-environment.rst
@@ -20,7 +20,7 @@ Visual Studio 2022
   that automatically detects that the folder is a CMake project
 - Right-click the ``CMakeLists.txt`` file to edit settings or build the project
     - You may need to set your configure command arguments to point to the dependencies
-      install (e.g. ``-DCMAKE_PREFIX_PATH=$(projectDir)/osc-dependencies-install``)
+      install (e.g. ``-DCMAKE_PREFIX_PATH=$(projectDir)/osc-deps-install``)
 - Use the ``Switch between solutions and available views`` button in the
   ``Solution Explorer`` hierarchy tab to switch to the ``CMake Targets View``
 - Right-click the ``osc`` CMake target and ``Set As Startup Project``, so that
@@ -72,8 +72,8 @@ QtCreator
 
 - Run the appropriate (OS-dependent) buildscript (described above)
 - Open QtCreator and then open the ``opensim-creator`` source directory as a folder
-- For selecting a "kit", QtCreator *usually* detects that ``osc-build`` already
+- For selecting a "kit", QtCreator *usually* detects that ``build/`` already
   exists (side-effect of running the buildscript). You *may* need to "import existing
-  kit/build" and then select ``osc-build``, though
+  kit/build" and then select ``build/``, though
 - Once QtCreator knows your source dir (``opensim-creator/``) and build/kit
-  (``opensim-creator/osc-build``), it should be good to go
+  (``opensim-creator/build/``), it should be good to go
