### 2025-05-01 712066a
Add sustainability, archiving, and mirrors section to README.md
