### 2025-05-01 d1612c9
Update RELEASE_CHECKLIST.md to include Gitlab mirror step
