### 2025-05-06 d0ff906
Add throwaway wayland build scripts/notes
