### 2025-05-07 b9e9e73
Clean up scripts directort slightly
