### 2025-05-08 b0a61b7
Hotfix issue when loading multiple model files silmultaneously (#1036)
