### 2025-05-09 c940802
Rename osc-deps-{build,install} to third_party-{build,install}
