+++ scripts/build_linux_debugging.sh
@@ -30,10 +30,10 @@ EOF
 # configure+build dependencies
 CCFLAGS=-fsanitize=address CXXFLAGS=-fsanitize=address cmake \
     -S third_party/ \
-    -B osc-deps-build \
+    -B third_party-build \
     -DCMAKE_BUILD_TYPE=RelWithDebInfo \
-    -DCMAKE_INSTALL_PREFIX=${PWD}/osc-deps-install
-cmake --build osc-deps-build/ -v -j${OSC_BUILD_CONCURRENCY}
+    -DCMAKE_INSTALL_PREFIX="${PWD}/third_party-install"
+cmake --build third_party-build/ -v -j${OSC_BUILD_CONCURRENCY}
 
 # configure+build OpenSimCreator
 # also: `-DCMAKE_CXX_INCLUDE_WHAT_YOU_USE "include-what-you-use;-Xiwyu;any;-Xiwyu;iwyu;-Xiwyu;"`
@@ -44,8 +44,7 @@ CCFLAGS="-fsanitize=address -fno-sanitize-recover=all" CXXFLAGS="-fsanitize=addr
     -B build/ \
     -DCMAKE_BUILD_TYPE=RelWithDebInfo \
     -DOSC_FORCE_ASSERTS_ENABLED=ON \
-    -DCMAKE_PREFIX_PATH=${PWD}/osc-deps-install \
-    -DCMAKE_INSTALL_PREFIX=${PWD}/osc-install \
+    -DCMAKE_PREFIX_PATH="${PWD}/third_party-install" \
     -DCMAKE_CXX_CLANG_TIDY=${CLANG_TIDY}
 cmake --build build/ -j${OSC_BUILD_CONCURRENCY}
 
