### 2025-05-09 e61091a
Squashed 'third_party/libosim/simbody/' content from commit 07a4b9d6b6
git-subtree-dir: third_party/libosim/simbody
git-subtree-split: 07a4b9d6b67ea0cfe842de52aaed68cd788670b7
