+++ Simbody/tests/TestCableSpan_LengthDot.cpp
@@ -1,0 +1,293 @@
+/*-----------------------------------------------------------------------------
+                Simbody(tm) Test: Cable Forces
+-------------------------------------------------------------------------------
+ Copyright (c) 2024 Authors.
+ Authors: Pepijn van den Bos
+ Contributors:
+
+ Licensed under the Apache License, Version 2.0 (the "License"); you may
+ not use this file except in compliance with the License. You may obtain a
+ copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ ----------------------------------------------------------------------------*/
+
+#include "Simbody.h"
+
+using namespace SimTK;
+
+/**
+This file contains simple tests for checking computed forces and moments.
+**/
+
+// Helper function for the force assertion.
+void assertLengthDot(
+    const std::string& testCase,
+    MultibodySystem& system,
+    CableSpan& cable,
+    const std::function<void(State& s)>& setU,
+    Real expectedLengthDot)
+{
+    cable.setCurveSegmentAccuracy(1e-12);
+    cable.setSmoothnessTolerance(1e-8);
+
+    // Initialize the system and state.
+    system.realizeTopology();
+    State s = system.getDefaultState();
+    setU(s);
+    system.realize(s, Stage::Report);
+
+    const Real gotLengthDot = cable.calcLengthDot(s);
+
+    for (CableSpanObstacleIndex ix(0); ix < cable.getNumObstacles(); ++ix) {
+        SimTK_ASSERT2_ALWAYS(
+            cable.isInContactWithObstacle(s, ix),
+            "%s failed: Cable not in contact with obstacle %i",
+            testCase.c_str(),
+            ix);
+    }
+
+    SimTK_ASSERT4_ALWAYS(
+        std::abs(gotLengthDot - expectedLengthDot) < 1e-13,
+        "%s failed: expected lengthDot (=%f) does not match computed "
+        "lengthDot (=%f), with error %e",
+        testCase.c_str(),
+        expectedLengthDot,
+        gotLengthDot,
+        expectedLengthDot - gotLengthDot);
+}
+
+// A cable with both end points on the same body, and no obstacles, should
+// result in zero lengthening speed.
+void testCableLengthDotOnSameBody()
+{
+    // Create the system.
+    MultibodySystem system;
+    SimbodyMatterSubsystem matter(system);
+    CableSubsystem cables(system);
+
+    Body::Rigid aBody(MassProperties(1., Vec3(0), Inertia(1)));
+    MobilizedBody::Free aMovingBody(
+        matter.Ground(),
+        Vec3(0., 0., 0.),
+        aBody,
+        Transform());
+
+    // Construct a new cable.
+    CableSpan cable(
+        cables,
+        aMovingBody,
+        Vec3(-1., 0., 0.),
+        aMovingBody,
+        Vec3(0., 1., 0.));
+
+    // Moving body velocity:
+    const Vec6 u{
+        // angular velocity
+        1., 2., 3.,
+        // linear velocity
+        4., 5., 6.,
+    };
+    const Real expectedLengthDot = 0.;
+
+    assertLengthDot(
+        "testCableLengthDotOnSameBody",
+        system,
+        cable,
+        [&](State& s) { aMovingBody.setU(s, u); },
[…]
