+++ SimTKcommon/SmallMatrix/include/SimTKcommon/internal/Mat.h
@@ -1,0 +1,1560 @@
+#ifndef SimTK_SIMMATRIX_SMALLMATRIX_MAT_H_
+#define SimTK_SIMMATRIX_SMALLMATRIX_MAT_H_
+
+/* -------------------------------------------------------------------------- *
+ *                       Simbody(tm): SimTKcommon                             *
+ * -------------------------------------------------------------------------- *
+ * This is part of the SimTK biosimulation toolkit originating from           *
+ * Simbios, the NIH National Center for Physics-Based Simulation of           *
+ * Biological Structures at Stanford, funded under the NIH Roadmap for        *
+ * Medical Research, grant U54 GM072970. See https://simtk.org/home/simbody.  *
+ *                                                                            *
+ * Portions copyright (c) 2005-12 Stanford University and the Authors.        *
+ * Authors: Michael Sherman                                                   *
+ * Contributors:                                                              *
+ *                                                                            *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may    *
+ * not use this file except in compliance with the License. You may obtain a  *
+ * copy of the License at http://www.apache.org/licenses/LICENSE-2.0.         *
+ *                                                                            *
+ * Unless required by applicable law or agreed to in writing, software        *
+ * distributed under the License is distributed on an "AS IS" BASIS,          *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+ * See the License for the specific language governing permissions and        *
+ * limitations under the License.                                             *
+ * -------------------------------------------------------------------------- */
+
+/**@file
+ * This file declares class Mat<NROWS, NCOLS, ELEMENT_TYPE, COL_SPACING, ROW_SPACING>.
+ */
+
+#include "SimTKcommon/internal/common.h"
+
+namespace SimTK {
+
+/** @brief This class represents a small matrix whose size is known at compile 
+time, containing elements of any Composite Numerical Type (CNT) and engineered 
+to have no runtime overhead whatsoever. 
+
+@ingroup MatVecUtilities
+
+Memory layout defaults to packed, column-ordered storage but can be specified to
+have any regular row and column spacing. A %Mat object is itself a Composite 
+Numerical Type and can thus be the element type for other matrix and vector 
+types. Some common use cases are provided below.
+
+@tparam M   The number of rows in this matrix (no default).
+@tparam N   The number of columns in this matrix (no default).
+@tparam ELT The element type; default is Real.
+@tparam CS  Column spacing in memory as a multiple of element size (default M).
+@tparam RS  %Row spacing in memory as a multiple of element size (default 1).
+
+<b>Construction</b>
+
+A 3x3 identity matrix can be constructed in the following ways:
+\code
+Mat<3,3,Real>(1,0,0, 0,1,0, 0,0,1);          //row-major
+Mat33(1,0,0,0,1,0,0,0,1);
+Mat33(Vec3(1,0,0),Vec3(0,1,0),Vec3(0,0,1));  //column-by-column
+Mat33(1);
+\endcode
+Note that the default element type is Real, and that Mat33 is a typedef for
[…]
