+++ Simbody/Visualizer/simbody-visualizer/simbody-visualizer.cpp
@@ -1,0 +1,2954 @@
+/* -------------------------------------------------------------------------- *
+ *                               Simbody(tm)                                  *
+ * -------------------------------------------------------------------------- *
+ * This is part of the SimTK biosimulation toolkit originating from           *
+ * Simbios, the NIH National Center for Physics-Based Simulation of           *
+ * Biological Structures at Stanford, funded under the NIH Roadmap for        *
+ * Medical Research, grant U54 GM072970. See https://simtk.org/home/simbody.  *
+ *                                                                            *
+ * Portions copyright (c) 2010-14 Stanford University and the Authors.        *
+ * Authors: Peter Eastman                                                     *
+ * Contributors: Michael Sherman                                              *
+ *                                                                            *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may    *
+ * not use this file except in compliance with the License. You may obtain a  *
+ * copy of the License at http://www.apache.org/licenses/LICENSE-2.0.         *
+ *                                                                            *
+ * Unless required by applicable law or agreed to in writing, software        *
+ * distributed under the License is distributed on an "AS IS" BASIS,          *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+ * See the License for the specific language governing permissions and        *
+ * limitations under the License.                                             *
+ * -------------------------------------------------------------------------- */
+
+#include "SimTKcommon.h"
+#include "simbody/internal/Visualizer.h"
+#include "simbody/internal/Visualizer_InputListener.h"
+#include "../src/VisualizerProtocol.h"
+#include "lodepng.h"
+
+#include <cstdlib>
+#include <cmath>
+#include <string>
+#include <algorithm>
+#include <set>
+#include <vector>
+#include <utility>
+#include <limits>
+#include <cstdio>
+#include <cerrno>
+#include <cstring>
+#include <thread>
+#include <condition_variable>
+#include <sys/stat.h>
+#ifdef _WIN32
+    #include <direct.h>
+#endif
+
+// Get gl and glut using the appropriate platform-dependent incantations.
+#if defined(__APPLE__)
+    // OSX comes with a glut implementation. In OSX 10.9 and 10.10, this
+    // glut is deprecated and emits deprecation warnings.
+    #pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+    #include <GLUT/glut.h>
+#elif defined(_WIN32)
+    #include "glut32/glut.h"    // we have our own private headers
+    #include "glut32/glext.h"
+
+    // A Windows-only extension for disabling vsync, allowing unreasonably
+    // high frame rates.
+    PFNWGLSWAPINTERVALFARPROC wglSwapIntervalEXT;
+
+    // These will hold the dynamically-determined function addresses.
+
+    // These functions are needed for basic Visualizer functionality.
+    PFNGLGENBUFFERSPROC glGenBuffers;
+    PFNGLBINDBUFFERPROC glBindBuffer;
+    PFNGLBUFFERDATAPROC glBufferData;
+    PFNGLACTIVETEXTUREPROC glActiveTexture;
+
+    // These are needed only for saving images and movies.
+    // Use old EXT names for these so we only require OpenGL 2.0.
+    PFNGLGENFRAMEBUFFERSEXTPROC glGenFramebuffersEXT;
[…]

[…]
