### 2025-05-09 2969dba
Add OSC_INTERNAL_HIDE_WINDOW shim for testing on rate-limited X servers
