### 2025-05-12 4982aec
Fix MSVC debug-mode issuse with NodePath (#982)
