### 2025-05-12 3fcbdbe
Fix minor typo in docs
