### 2025-05-12 f163bc2
Add TypeInfoReference::operator< and TypeInfoReference tests
