### 2025-05-12 7411621
Format generator sources
