### 2025-05-12 360f33d
Re-enable VariableLengthArray (might work on current MacOS SDK)
