+++ liboscar/CMakeLists.txt
@@ -36,7 +36,6 @@ file(GLOB_RECURSE OSC_SOURCE_AND_HEADER_FILES
 )
 list(FILTER OSC_SOURCE_AND_HEADER_FILES EXCLUDE REGEX "testing")
 list(FILTER OSC_SOURCE_AND_HEADER_FILES EXCLUDE REGEX "\.tests\.cpp")
-list(FILTER OSC_SOURCE_AND_HEADER_FILES EXCLUDE REGEX "VariableLengthArray")  # hacked out: can't compile on MacOS yet
 
 add_library(oscar STATIC
     ${OSC_SOURCE_AND_HEADER_FILES}
