+++ .github/workflows/continuous-integration-workflow.yml
@@ -27,8 +27,8 @@ jobs:
         if-no-files-found: error
 
   macos11clang:
-    name: MacOS 13 (amd64, XCode 15.2, targeting 14.2)
-    runs-on: macos-13
+    name: MacOS 14 (amd64, XCode 15.2, targeting 14.2)
+    runs-on: macos-14
     steps:
 
     - uses: actions/checkout@v4
