### 2025-05-13 0f77327
Clean up compiler flags, disable frame pointer pinning (noone uses it)
