### 2025-05-13 a23217e
Update XCode version and ensure compiler targets 14.2
