### 2025-05-13 721d777
Downgrade runner image to macos-14 to encourage SDK 14.5 usage
