+++ docs/source/installation.rst
@@ -22,8 +22,8 @@ Installing on Windows (10 or newer)
   its installation location (defaults to ``C:\Program Files\OpenSimCreator\``).
 
 
-Installing on MacOS (Sonoma or newer)
---------------------------------------
+Installing on MacOS (Sonoma 14.5 or newer)
+------------------------------------------
 
 - Download a ``dmg``-based installer from the `GitHub Releases Page`_
 - Double click the ``dmg`` file to mount it
