+++ liboscar/Shims.h
@@ -2,7 +2,5 @@
 
 #include <liboscar/Shims/Cpp20/stop_token.h>
 #include <liboscar/Shims/Cpp20/thread.h>
-#include <liboscar/Shims/Cpp23/cstddef.h>
 #include <liboscar/Shims/Cpp23/numeric.h>
 #include <liboscar/Shims/Cpp23/ranges.h>
-#include <liboscar/Shims/Cpp23/utility.h>
