### 2025-05-13 7695fb4
Remove unnecessary warning flags
