+++ liboscar/Utils/StringHelpers.cpp
@@ -23,42 +23,12 @@ namespace
     {
         '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'
     });
-
-    std::string to_lowercase(std::string_view str)
-    {
-        std::string cpy{str};
-        rgs::transform(cpy, cpy.begin(), [](std::string::value_type c)
-        {
-            return static_cast<std::string::value_type>(std::tolower(c));
-        });
-        return cpy;
-    }
-}
-
-bool osc::contains(std::string_view sv, std::string_view substr)
-{
-    return std::search(sv.begin(), sv.end(), substr.begin(), substr.end()) != sv.end();
-}
-
-bool osc::contains(std::string_view sv, std::string_view::value_type c)
-{
-    return cpp23::contains(sv, c);
 }
 
 bool osc::contains_case_insensitive(std::string_view sv, std::string_view substr)
 {
-    if (substr.empty()) {
-        return true;
-    }
-
-    if (substr.size() > sv.size()) {
-        return false;
-    }
-
-    const std::string s = to_lowercase(sv);
-    const std::string ss = to_lowercase(substr);
-
-    return contains(s, ss);
+    const auto tolower = [](const auto c) { return std::tolower(c); };
+    return not std::ranges::search(sv, substr, std::ranges::equal_to{}, tolower, tolower).empty();
 }
 
 bool osc::is_string_case_insensitive_greater_than(std::string_view a, std::string_view b)
