### 2025-05-13 162426c
Update building-from-source.rst to reflect latest compiler toolchain versions
