+++ .github/workflows/continuous-integration-workflow.yml
@@ -5,7 +5,7 @@ env:
 
 jobs:
 
-  windows2022msvc:
+  windows10_msvc_amd64:
     name: Windows 2022 (amd64, MSVC)
     runs-on: windows-2022
     steps:
@@ -28,7 +28,7 @@ jobs:
         path: build/*windows-amd64.exe
         if-no-files-found: error
 
-  macos11clang:
+  macos14_xcode15-4_targeting14-5_amd64:
     name: MacOS 14 (amd64, XCode 15.4, targeting 14.5)
     runs-on: macos-14
     steps:
@@ -39,8 +39,11 @@ jobs:
       run: |
         cd $GITHUB_WORKSPACE
 
+        # configure buildscript
+        export OSC_CMAKE_CONFIG_EXTRA="-DCMAKE_OSX_ARCHITECTURES=x86_64 -DCMAKE_OSX_DEPLOYMENT_TARGET=14.5"
+
         # run buildscript
-        ./scripts/build_mac-amd64.sh
+        ./scripts/build_mac.sh
 
         # run after-build OS-specific test scripts
         ./scripts/macos_check-dependencies.py build/osc/osc
@@ -52,7 +55,34 @@ jobs:
         path: build/*.dmg
         if-no-files-found: error
 
-  ubuntu2204gcc:
+  macos14_xcode15-4_targeting14-5_arm64:
+    name: MacOS 14 (arm64, XCode 15.4, targeting 14.5)
+    runs-on: macos-14
+    steps:
+
+    - uses: actions/checkout@v4
+
+    - name: run build script
+      run: |
+        cd $GITHUB_WORKSPACE
+
+        # configure buildscript
+        export OSC_CMAKE_CONFIG_EXTRA="-DCMAKE_OSX_ARCHITECTURES=arm64 -DCMAKE_OSX_DEPLOYMENT_TARGET=14.5"
+
+        # run buildscript
+        ./scripts/build_mac.sh
+
+        # run after-build OS-specific test scripts
+        ./scripts/macos_check-dependencies.py build/osc/osc
+        ./scripts/macos_check-sdk.py 14.5 build/osc/osc
+
+    - uses: actions/upload-artifact@v4
+      with:
+        name: opensimcreator-macos-arm64.zip
+        path: build/*.dmg
+        if-no-files-found: error
+
+  ubuntu22-04_gcc-12_amd64:
     name: Ubuntu 22.04 (amd64, gcc-12)
     runs-on: ubuntu-22.04
     steps:
