### 2025-05-15 15c7376
Update get_osc_github_download_counts to include OS usage
