+++ scripts/get_osc_github_download_counts.py
@@ -43,37 +43,76 @@ urls = [
 
 version_pattern = re.compile(r"\d+\.\d+.\d+")
 
-rows = legacy_download_counters.copy()
-for url in urls:
-    resp = requests.get(url=url)
+def fetch_raw_rows(add_legacy_rows=True):
+    if add_legacy_rows:
+        rv = legacy_download_counters.copy()
+    else:
+        rv = []
+    for url in urls:
+        resp = requests.get(url=url)
 
-    for el in resp.json():
-        for asset in el["assets"]:
-            name = asset['name']
-            match = version_pattern.search(name)
-            if not match:
-                continue
-            version = match.group(0)
-            created_at = datetime.datetime.strptime(asset['created_at'], '%Y-%m-%dT%H:%M:%SZ')
-            downloads = int(asset['download_count'])
-            rows.append(Row(name, version, created_at, downloads))
+        for el in resp.json():
+            for asset in el["assets"]:
+                name = asset['name']
+                match = version_pattern.search(name)
+                if not match:
+                    continue
+                version = match.group(0)
+                created_at = datetime.datetime.strptime(asset['created_at'], '%Y-%m-%dT%H:%M:%SZ')
+                downloads = int(asset['download_count'])
+                rv.append(Row(name, version, created_at, downloads))
+    return rv
 
-# aggregate downloads by version number
-aggregate = []
-for key, group in itertools.groupby(rows, lambda row: row.version):
-    first = next(group)
-    name = first.name.split('-')[0]
-    version = first.version
-    created_at = first.created_at
-    downloads = functools.reduce(lambda acc,el: acc + el.downloads, group, first.downloads)
-    aggregate.append(Row(name, version, created_at, downloads))
-rows = aggregate
+def extract_extension(row):
+    name = row.name
+    if match := version_pattern.search(name):
+        name = row.name[match.end():]
+    return '.'.join(name.split('.')[1:]) if '.' in name else ''
 
-# sort by creation time
-rows = sorted(rows, key=lambda e: e.created_at)
+def print_download_analysis(rows):
+    # aggregate downloads by version number
+    aggregate = []
+    for key, group in itertools.groupby(rows, lambda row: row.version):
+        first = next(group)
+        name = first.name.split('-')[0]
+        version = first.version
+        created_at = first.created_at
+        downloads = functools.reduce(lambda acc,el: acc + el.downloads, group, first.downloads)
+        aggregate.append(Row(name, version, created_at, downloads))
+    rows = aggregate
 
-acc = 0
-print("download name,version,created at,num downloads,cumulative sum")
-for row in rows:
-    acc += row.downloads
-    print(f"{row.name},{row.version},{row.created_at.date()},{row.downloads},{acc}")
+    # sort by creation time
+    rows = sorted(rows, key=lambda e: e.created_at)
+
+    acc = 0
+    print("download name,version,created at,num downloads,cumulative sum")
+    for row in rows:
+        acc += row.downloads
+        print(f"{row.name},{row.version},{row.created_at.date()},{row.downloads},{acc}")
+
+def print_download_type_percentages(rows):
+    # filter out empty extensions
+    rows = list(filter(lambda row: len(extract_extension(row))>0, rows))
+
+    # sort by extension
+    rows = sorted(rows, key=extract_extension)
+
+    # aggregate by extension
+    aggregate = []
+    total_downloads = 0
+    for key, group in itertools.groupby(rows, extract_extension):
+        first = next(group)
+        name = extract_extension(first)
+        downloads = functools.reduce(lambda acc,el: acc + el.downloads, group, first.downloads)
+        total_downloads += downloads
[…]
