### 2025-05-15 394fb89
Format scripts a little
