+++ scripts/bundle_sources.sh
@@ -6,4 +6,3 @@
 set -xeuo pipefail
 
 git ls-files --recurse-submodules | tar caf ../opensimcreator-${1}-src.tar.xz --xform s:^:opensimcreator-${1}/: --verbatim-files-from -T-
-
