### 2025-05-15 03f1eed
Add --defaults to format_sources
