### 2025-05-15 6ca2781
Patch MSVC debug issue with VariableLengthArray
