+++ liboscar/Utils/VariableLengthArray.h
@@ -15,13 +15,13 @@ namespace osc
     // Provides a container that behaves as an on-stack/on-heap hybrid array that:
     //
     // - Allocates up to `N` elements on the stack without using a memory allocator.
-    // - Once the number of elements exeeds `N`, allocates all elements (incl. existing
+    // - Once the number of elements exceeds `N`, allocates all elements (incl. existing
     //   elements, which are moved) via an upstream memory resource, which defaults to
     //   a `std::pmr::new_delete_resource`
     //
     // This is handy when the caller believes that there's likely to be a (low) upper
     // bound on the number of elements in the container, but they cannot be 100 %
-    // certain that the number of elements will never exeed that bound. E.g. it might
+    // certain that the number of elements will never exceed that bound. E.g. it might
     // be fair to assume that a tree loaded from a data file never exceeds 16 levels
     // of depth for reasonable inputs, but there might exist unreasonable ones, so
     // having a safe fallback to an upstream allocator is useful.
@@ -117,14 +117,35 @@ namespace osc
         void push_back(T&& value) { vector_.push_back(std::move(value)); }
 
         friend bool operator==(const VariableLengthArray& lhs, const VariableLengthArray& rhs) { return lhs.vector_ == rhs.vector_; }
+
     private:
+// Fix debug ABI issue from MSVC: when compiling in debug mode, MSVC will allocate a
+// little container proxy at the start of the allocation unit (presumably, to track
+// allocations) and that causes problems when downstream code expects spillover to
+// occur at particular points (e.g. the test suite ensures this).
+//
+// see: https://developercommunity.visualstudio.com/t/debug-version-of-stl-is-not-excepion-safe-and-caus/77779
+#if defined(_MSC_VER) && _ITERATOR_DEBUG_LEVEL > 0
+        static inline constexpr size_t extra_bytes_ = []()
+        {
+            // from MSVC
+            constexpr size_t proxy_size = sizeof(std::_Container_proxy);
+            // ensure the extra bytes also account for the remainder of the buffer being used
+            // for over-aligned datastructures
+            constexpr size_t num_align_windows = (proxy_size + (alignof(T) - 1))/alignof(T);
+            return alignof(T) * num_align_windows;
+        }();
+#else
+        static inline constexpr size_t extra_bytes_ = 0;
+#endif
+
         // The object representation of elements, stored on-stack when `size() <= N`
-        alignas(T) std::array<std::byte, N * sizeof(T)> stack_data_;
+        alignas(T) std::array<std::byte, N * sizeof(T) + extra_bytes_> stack_data_;
 
         // A memory resource that uses `stack_data_` until `size() > N`, it then uses an upstream resource
         std::pmr::monotonic_buffer_resource pool{stack_data_.data(), stack_data_.size(), std::pmr::new_delete_resource()};
 
         // A vector that's backed by the above memory resource
-        std::pmr::vector<T> vector_ = std::pmr::vector<T>(&pool);
+        underlying_vector vector_ = underlying_vector(&pool);
     };
 }
