### 2025-05-15 5f36564
Upate station-defined-frames.rst with more explanation text
