+++ docs/source/station-defined-frames.rst
@@ -3,7 +3,7 @@ Station Defined Frames
 
 This tutorial focuses on the ``StationDefinedFrame`` component, which was
 merged into OpenSim in `opensim-core/pull/3694`_  and should be available in
-OpenSim >= v4.3.1. Specifically, this tutorial focuses on adding
+OpenSim >= v4.5.1. Specifically, this tutorial focuses on adding
 ``StationDefinedFrames`` via OpenSim Creator. If you're more interested
 in the underlying implementation, or its associated OpenSim API,
 then `StationDefinedFrame.h`_ provides a technical (C++) description.
@@ -55,22 +55,22 @@ are shown in :numref:`sdf-maths-figure`.
 .. figure:: _static/station-defined-frames/sdf-maths.svg
     :width: 60%
 
-    The relationship between stations and a ``StationDefinedFrame``. :math:`\vec{a}`,
-    :math:`\vec{b}`, :math:`\vec{c}`, and :math:`\vec{d}` are four stations in the
-    model that must be attached---either directly, or indirectly (e.g. via a
-    ``PhysicalOffsetFrame``)---to the same body. The ``StationDefinedFrame``
-    implementation uses the stations to derive :math:`f(\vec{v})`, its transform
-    function. The origin station, :math:`\vec{o}`, may be coincident with one of
+    The relationship between stations and a ``StationDefinedFrame``. :math:`\mathbf{a}`,
+    :math:`\mathbf{b}`, :math:`\mathbf{c}`, and :math:`\mathbf{o}` (origin) are four
+    stations in the model that must be attached---either directly, or indirectly (e.g.
+    via a ``PhysicalOffsetFrame``)---to the same body. The ``StationDefinedFrame``
+    implementation uses the stations to derive :math:`f(\mathbf{v})`, its transform
+    function. The origin station, :math:`\mathbf{o}`, may be coincident with one of
     the other stations.
 
 Practically speaking, this means is that ``StationDefinedFrame`` s let modellers
-define frames by choosing 3 or 4 stations (landmarks) on each body. Once that
-relationship is established, the resulting frame is automatically recalculated
+define frames by choosing/calculating 3 or 4 stations (landmarks) on each body. Once
+that relationship is established, the resulting frame is automatically recalculated
 whenever the the stations moved (e.g. due to scaling, warping, shear, etc.).
 
 
-A Simple Example
-----------------
+Example Walkthrough
+-------------------
 
 OpenSim Creator includes example models that use ``StationDefinedFrame``:
 
@@ -81,10 +81,9 @@ OpenSim Creator includes example models that use ``StationDefinedFrame``:
   ``StationDefinedFrame`` s that are chained and use stations attached via
   ``PhysicalOffsetFrame`` s.
 
-This example walks through how something like ``StationDefinedFrame.osim`` can be built from
-scratch, so that you can get an idea of how the mathematics (:numref:`sdf-maths-figure`) is exposed via
-OpenSim's component system. The next section, `A Practical Example`_, then shows how ``StationDefinedFrame`` s
-can be added to an existing model.
+This walkthrough outlines creating something similar to ``StationDefinedFrame.osim``, so
+that you can get an idea of how the mathematics (:numref:`sdf-maths-figure`) is exposed via
+OpenSim's component system.
 
 
 Make a One-Body Model
@@ -99,50 +98,116 @@ Make a One-Body Model
 .. figure:: _static/station-defined-frames/model-with-one-body.jpg
     :width: 60%
 
-    ``TODO`` model containing one body with a brick attached.
[…]
