### 2025-05-15 7bf0081
Ensure OSC_INTERNAL_HIDE_WINDOW is set in CI
