### 2025-05-15 66c46fc
Remove Windows test exclusions
