### 2025-05-15 7faf051
Fix cannot take screenshots while modal is shown (#1038)
