+++ docs/source/index.rst
@@ -55,6 +55,7 @@ Table of Contents
    the-mesh-importer-advanced
    the-mesh-warper
    preview-experimental-data
+   station-defined-frames
 
 .. toctree::
    :maxdepth: 2
@@ -78,7 +79,6 @@ Table of Contents
     :numbered:
    
     use-the-model-warper
-    station-defined-frames
 
 
 .. _Indices and Tables:
