### 2025-05-16 4d4ed74
Refactor some static_cast<{size_t,ptrdiff_t}> to C++23 literals
