### 2025-05-16 c3494ac
Remove now-invalid model warping documentation steps
