+++ docs/source/the-mesh-warper.rst
@@ -1,3 +1,5 @@
+.. _the-mesh-warper:
+
 The Mesh Warper
 ===============
 
