+++ docs/source/use-the-model-warper.rst
@@ -1,40 +1,32 @@
-Use the Model Warper
-====================
+The Model Warper
+================
 
 .. warning::
 
-    **Model warping is very 🪄 experimental 🪄.**
+    **The model warper is very very 🪄 experimental 🪄.**
 
-    We invite you to try model warping out and get a feel for how it might be
+    The model warper has been in development since 2023 and has undergone several
+    redesigns from its initial version (an extended form of :ref:`the-mesh-warper`)
+    to what it is now (a way of combining various scaling steps in a linear pipeline).
+    This is because because non-linear "model warping" combines a variety of scaling
+    algorithms in a model-specific way - our goal is to provide clear UI tooling
+    that makes combining those algorithms easier.
+
+    We invite you to try the model warper out and get a feel for how it might be
     useful. Maybe it's already useful enough for you to use it in something
     serious (some researchers already have 🎉).
 
-    However, be aware that some implementation details of model warping, such
-    as the structure of the ``.warpconfig.toml`` file, aren't stable yet. This
-    means that they may change as model warping is developed.
-
-    The documentation here is extremely work-in-progress, so expect many ``TODO`` s
-    and ``FIXME`` s. This will be improved as the model warping feature is developed.
-    We figured it's better to show you what's available as we develop it, rather
-    than only releasing once it's perfect.
-    
-
-In this tutorial, we will be using the model warping UI to warp an entire
-OpenSim model in a landmark-driven way. The benefit of this technique,
-compared to standard scaling, is that it makes non-uniform, subject-specific
-scaling possible.
+In this tutorial, we will be using the model warper to create a warping pipeline that
+can be used to warp an entire OpenSim model using experimental measurements, such as
+CT scans and weight measurements. The benefit of the model warper is that it lets you
+combine various, potentially non-uniform, scaling steps into a single warping pipeline
+that's standard, introspectible, and reusable.
 
 .. _model-warper-ui:
 .. figure:: _static/use-the-model-warper/model-warper.jpg
     :width: 60%
 
-    A screenshot of the model warping UI, showing how it can be used to warp
-    a source/reference/template OpenSim model (left) into a new model (right).
-    In this case, the warping relationship is possible because there are
-    corresponding landmarks available for all meshes in both the source and
-    target (which were placed as described in :doc:`the-mesh-warper`). Those landmarks
-    enable the non-uniform scaling shown above (e.g. the scapula has a
-    different size *and* shape).
+    TODO
 
 
 Prerequisites
@@ -43,262 +35,43 @@ Prerequisites
 * **You can diagnose and work with OpenSim models**. This tutorial assumes that
   you're able to diagnose the models that go into, and come out of, the model
   warping UI. If you don't feel comfortable with working on OpenSim models, then
-  we recommend going through earlier tutorials (e.g. :doc:`make-a-pendulum`, :doc:`make-a-bouncing-block`).
+  we recommend going through some model-building tutorials (e.g. :doc:`make-a-pendulum`,
+  :doc:`make-a-bouncing-block`).
 
-* **A basic understanding of the Thin-Plate Spline (TPS) technique**. The model
-  warper applies the TPS technique to multiple components in the source model.
-  Therefore, it's recommended that you have already gone through :doc:`the-mesh-warper`, which
-  outlines pairing landmarks between two corresponding meshes as inputs for the
-  TPS technique.
+* **A basic understanding of the Thin-Plate Spline (TPS) technique**. The walkthrough
+  in this tutorial uses the TPS technique to warp parts of the model. Therefore, it's
+  recommended that you have already gone through :doc:`the-mesh-warper`, which outlines
[…]
