### 2025-05-16 f6dcc35
Drop deprecated/prototype model warper UI and instead expose the new one
