+++ libopensimcreator/Documents/ModelWarperV2/ModelWarperConfiguration.tests.cpp
@@ -1,787 +1,0 @@
-#include "ModelWarperConfiguration.h"
-
-#include <libopensimcreator/Platform/OpenSimCreatorApp.h>
-#include <libopensimcreator/testing/TestOpenSimCreatorConfig.h>
-#include <libopensimcreator/Utils/LandmarkPair3D.h>
-#include <libopensimcreator/Utils/OpenSimHelpers.h>
-
-#include <gtest/gtest.h>
-#include <liboscar/Utils/StringHelpers.h>
-#include <liboscar/Utils/TemporaryFile.h>
-#include <OpenSim/Simulation/Model/Marker.h>
-#include <OpenSim/Simulation/Model/Model.h>
-
-#include <algorithm>
-#include <array>
-#include <cstring>
-#include <exception>
-#include <filesystem>
-#include <ranges>
-
-using namespace osc;
-using namespace osc::mow;
-namespace rgs = std::ranges;
-
-namespace
-{
-    std::filesystem::path GetFixturePath(const std::filesystem::path& subpath)
-    {
-        return std::filesystem::weakly_canonical(std::filesystem::path{OSC_TESTING_RESOURCES_DIR} / subpath);
-    }
-
-    // A testing class that can act as an inner node in a model's component tree.
-    class ContainerNode : public OpenSim::Component {
-        OpenSim_DECLARE_CONCRETE_OBJECT(ContainerNode, OpenSim::Component)
-    };
-}
-
-TEST(StrategyMatchQuality, NoneComparesLessThanWildcard)
-{
-    // A wildcard match is "better than" no match.
-    static_assert(StrategyMatchQuality::none() < StrategyMatchQuality::wildcard());
-}
-
-TEST(StrategyMatchQuality, WildcardComparesLessThanExact)
-{
-    // An exact match is "better than" a wildcard match.
-    static_assert(StrategyMatchQuality::wildcard() < StrategyMatchQuality::exact());
-}
-
-TEST(StrategyMatchQuality, NoneImplicitlyConvertsToFalse)
-{
-    // The truthiness of a `StrategyMatchQuality` corresponds to "is there any match?".
-    static_assert(not static_cast<bool>(StrategyMatchQuality::none()));
-}
-
-TEST(StrategyMatchQuality, WildcardImplicitlyConvertsToTrue)
-{
-    // The truthiness of a `StrategyMatchQuality` corresponds to "is there any match?".
-    static_assert(static_cast<bool>(StrategyMatchQuality::wildcard()));
-}
-
-TEST(StrategyMatchQuality, ExactImplicitlyConvertsToTrue)
-{
-    // The truthiness of a `StrategyMatchQuality` corresponds to "is there any match?".
-    static_assert(static_cast<bool>(StrategyMatchQuality::exact()));
-}
-
-TEST(RuntimeWarpParameters, ConstructedWithBlendFactorMakesGetBlendFactorReturnTheBlendFactor)
-{
-    const RuntimeWarpParameters params{0.3f};
-    ASSERT_EQ(params.getBlendFactor(), 0.3f);
-}
-
-TEST(WarpCache, CanDefaultConstruct)
-{
-    [[maybe_unused]] const WarpCache instance;
-}
-
-TEST(IdentityComponentWarper, CanDefaultConstruct)
-{
-    [[maybe_unused]] const IdentityComponentWarper instance;
-}
-
-TEST(IdentityComponentWarper, DoesNotChangeAnyComponentProperty)
-{
-    // An `IdentityComponentWarper` shouldn't do anything to a component when it "warps" it.
-
-    OpenSim::Model sourceModel;
-    OpenSim::Marker& sourceMarker = AddMarker(sourceModel, "marker", sourceModel.getGround(), SimTK::Vec3{0.0});
-    FinalizeConnections(sourceModel);
[…]
