+++ libopensimcreator/Documents/ModelWarper/TPSLandmarkPairWarperFactory.cpp
@@ -1,304 +1,0 @@
-#include "TPSLandmarkPairWarperFactory.h"
-
-#include <libopensimcreator/Documents/Landmarks/LandmarkHelpers.h>
-#include <libopensimcreator/Documents/ModelWarper/ValidationCheckState.h>
-#include <libopensimcreator/Documents/ModelWarper/WarpableModel.h>
-#include <libopensimcreator/Utils/TPS3D.h>
-
-#include <liboscar/Maths/Vec2.h>
-#include <liboscar/Platform/Log.h>
-#include <liboscar/Shims/Cpp23/ranges.h>
-#include <liboscar/Utils/Algorithms.h>
-
-#include <algorithm>
-#include <filesystem>
-#include <fstream>
-#include <functional>
-#include <optional>
-#include <span>
-#include <sstream>
-#include <string>
-#include <utility>
-#include <vector>
-
-using namespace osc;
-using namespace osc::lm;
-using namespace osc::mow;
-namespace rgs = std::ranges;
-
-namespace
-{
-    std::filesystem::path CalcExpectedAssociatedLandmarksFile(
-        const std::filesystem::path& meshAbsolutePath)
-    {
-        std::filesystem::path expected{meshAbsolutePath};
-        expected.replace_extension(".landmarks.csv");
-        expected = std::filesystem::weakly_canonical(expected);
-        return expected;
-    }
-
-    std::filesystem::path CalcExpectedDestinationMeshFilepath(
-        const std::filesystem::path& osimFilepath,
-        const std::filesystem::path& sourceMeshFilepath)
-    {
-        std::filesystem::path expected = osimFilepath.parent_path() / "DestinationGeometry" / sourceMeshFilepath.filename();
-        expected = std::filesystem::weakly_canonical(expected);
-        return expected;
-    }
-
-    std::vector<MaybeNamedLandmarkPair> PairLandmarks(std::vector<Landmark> a, std::vector<Landmark> b)
-    {
-        std::vector<MaybeNamedLandmarkPair> rv;
-        TryPairingLandmarks(std::move(a), std::move(b), [&rv](const MaybeNamedLandmarkPair& p) { rv.push_back(p); });
-        return rv;
-    }
-
-    std::vector<MaybeNamedLandmarkPair> TryLoadPairedLandmarks(
-        [[maybe_unused]] const std::optional<std::filesystem::path>& maybeSourceLandmarksCSV,
-        [[maybe_unused]] const std::optional<std::filesystem::path>& maybeDestinationLandmarksCSV)
-    {
-        std::vector<Landmark> src;
-        if (maybeSourceLandmarksCSV) {
-            try {
-                src = ReadLandmarksFromCSVIntoVectorOrThrow(*maybeSourceLandmarksCSV);
-            }
-            catch (const std::exception& ex) {
-                log_error("%s", ex.what());
-            }
-        }
-        std::vector<Landmark> dest;
-        if (maybeDestinationLandmarksCSV) {
-            try {
-                dest = ReadLandmarksFromCSVIntoVectorOrThrow(*maybeDestinationLandmarksCSV);
-            }
-            catch (const std::exception& ex) {
-                log_error("%s", ex.what());
-            }
-        }
-        return PairLandmarks(std::move(src), std::move(dest));
-    }
-
-    TPSCoefficients3D<float> TryCalcTPSCoefficients(std::span<const MaybeNamedLandmarkPair> maybePairs)
-    {
-        TPSCoefficientSolverInputs3D<float> rv;
[…]
