### 2025-05-16 7c54bd5
Hide the frame definition tab from the splash screen (#951)
