+++ liboscar/Graphics/GraphicsImplementation.cpp
@@ -456,7 +456,7 @@ namespace
         case SourceBlendingFactor::Zero:                return GL_ZERO;
         case SourceBlendingFactor::SourceAlpha:         return GL_SRC_ALPHA;
         case SourceBlendingFactor::OneMinusSourceAlpha: return GL_ONE_MINUS_SRC_ALPHA;
-        default:                                        return GL_ONE;
+        default:                                        std::unreachable();
         }
     }
 
@@ -468,7 +468,7 @@ namespace
         case DestinationBlendingFactor::Zero:                return GL_ZERO;
         case DestinationBlendingFactor::SourceAlpha:         return GL_SRC_ALPHA;
         case DestinationBlendingFactor::OneMinusSourceAlpha: return GL_ONE_MINUS_SRC_ALPHA;
-        default:                                             return GL_ONE;
+        default:                                             std::unreachable();
         }
     }
 
@@ -479,7 +479,7 @@ namespace
         case BlendingEquation::Add: return GL_FUNC_ADD;
         case BlendingEquation::Min: return GL_MIN;
         case BlendingEquation::Max: return GL_MAX;
-        default:                    return GL_FUNC_ADD;
+        default:                    std::unreachable();
         }
     }
 }
@@ -1000,7 +1000,7 @@ namespace
         case TextureWrapMode::Repeat: return GL_REPEAT;
         case TextureWrapMode::Clamp:  return GL_CLAMP_TO_EDGE;
         case TextureWrapMode::Mirror: return GL_MIRRORED_REPEAT;
-        default:                      return GL_REPEAT;
+        default:                      std::unreachable();
         }
     }
 
@@ -1028,7 +1028,7 @@ namespace
         case TextureFilterMode::Nearest: return GL_NEAREST;
         case TextureFilterMode::Linear:  return GL_LINEAR;
         case TextureFilterMode::Mipmap:  return GL_LINEAR_MIPMAP_LINEAR;
-        default:                         return GL_LINEAR;
+        default:                         std::unreachable();
         }
     }
 
@@ -1040,7 +1040,7 @@ namespace
         case TextureFilterMode::Nearest: return GL_NEAREST;
         case TextureFilterMode::Linear:  return GL_LINEAR;
         case TextureFilterMode::Mipmap:  return GL_LINEAR;
-        default:                         return GL_LINEAR;
+        default:                         std::unreachable();
         }
     }
 }
@@ -1768,10 +1768,10 @@ std::optional<TextureFormat> osc::to_texture_format(size_t num_components, Textu
 
     static_assert(num_options<TextureFormat>() == 7);
     switch (num_components) {
-    case 1: return format_is_byte_oriented ? TextureFormat::R8     : std::optional<TextureFormat>{};
-    case 2: return format_is_byte_oriented ? TextureFormat::RG16   : TextureFormat::RGFloat;
-    case 3: return format_is_byte_oriented ? TextureFormat::RGB24  : TextureFormat::RGBFloat;
-    case 4: return format_is_byte_oriented ? TextureFormat::RGBA32 : TextureFormat::RGBAFloat;
+    case 1:  return format_is_byte_oriented ? TextureFormat::R8     : std::optional<TextureFormat>{};
+    case 2:  return format_is_byte_oriented ? TextureFormat::RG16   : TextureFormat::RGFloat;
+    case 3:  return format_is_byte_oriented ? TextureFormat::RGB24  : TextureFormat::RGBFloat;
+    case 4:  return format_is_byte_oriented ? TextureFormat::RGBA32 : TextureFormat::RGBAFloat;
     default: return std::nullopt;
     }
 }
[…]
