### 2025-05-19 0f034ef
Add more build/development/release information to the documentation (#1022 #1017)
