### 2025-05-19 4a908fd
Fix minor maths typos in SDF docs etc.
