+++ libopensimcreator/Utils/TPS3D.h
@@ -37,6 +37,10 @@ namespace osc
         friend bool operator==(const TPSCoefficientSolverInputs3D&, const TPSCoefficientSolverInputs3D&) = default;
 
         std::vector<LandmarkPair3D<T>> landmarks;
+        bool applyAffineTranslation = true;
+        bool applyAffineScale = true;
+        bool applyAffineRotation = true;
+        bool applyNonAffineWarp = true;
     };
 
     std::ostream& operator<<(std::ostream&, const TPSCoefficientSolverInputs3D<float>&);
