### 2025-05-21 8fe2aa4
Squashed 'third_party/lunasvg/' changes from 22a0e3a7ce8..49c8cb49ac8
49c8cb49ac8 Release v3.3.0
77e564cecba Bump plutovg version to 1.1.0
63080a73afa Generate `lunasvg.pc` from CMake for pkg-config support
12d285b542c Update CMakeLists.txt
0671fc2fd7b Merge pull request #219 from hatch01/master
57f859fb936 Fix: Expand clip bounds by 1 pixel to account for rendering overshoot
a13fddc712e Fix: Round float-to-fixed conversion to avoid truncation errors
93422b6d46a Enforce valid (0 < size < 32K) constraint on surface dimensions #143 #221 #222
a0408e1d11d Clip line to `clip_box` using Cyrus-Beck b
[…]
