+++ CMakeLists.txt
@@ -1,45 +1,152 @@
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 3.15)
 
-project(lunasvg VERSION 2.4.1 LANGUAGES CXX C)
+set(LUNASVG_VERSION_MAJOR 3)
+set(LUNASVG_VERSION_MINOR 3)
+set(LUNASVG_VERSION_MICRO 0)
 
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_C_STANDARD 11)
+project(lunasvg LANGUAGES CXX VERSION ${LUNASVG_VERSION_MAJOR}.${LUNASVG_VERSION_MINOR}.${LUNASVG_VERSION_MICRO})
 
-option(BUILD_SHARED_LIBS "Build as a shared library" ON)
-option(LUNASVG_BUILD_EXAMPLES "Build example(s)" OFF)
+option(USE_SYSTEM_PLUTOVG "Use system plutovg library" OFF)
 
-add_library(lunasvg)
+if(USE_SYSTEM_PLUTOVG)
+    find_package(plutovg 1.0.0 QUIET)
+    if(NOT plutovg_FOUND)
+        message(WARNING "Could not find: plutovg>=1.0.0. Falling back to plutovg submodule.")
+    endif()
+endif()
 
-add_subdirectory(include)
-add_subdirectory(source)
-add_subdirectory(3rdparty/plutovg)
+if(NOT plutovg_FOUND)
+    message(STATUS "Using plutovg submodule.")
+    add_subdirectory(plutovg)
+endif()
+
+set(lunasvg_sources
+    source/lunasvg.cpp
+    source/graphics.cpp
+    source/svgelement.cpp
+    source/svggeometryelement.cpp
+    source/svglayoutstate.cpp
+    source/svgpaintelement.cpp
+    source/svgparser.cpp
+    source/svgproperty.cpp
+    source/svgrenderstate.cpp
+    source/svgtextelement.cpp
+)
+
+set(lunasvg_headers
+    include/lunasvg.h
+    source/graphics.h
+    source/svgelement.h
+    source/svggeometryelement.h
+    source/svglayoutstate.h
+    source/svgpaintelement.h
+    source/svgparserutils.h
+    source/svgproperty.h
+    source/svgrenderstate.h
+    source/svgtextelement.h
+)
+
+add_library(lunasvg ${lunasvg_sources} ${lunasvg_headers})
+add_library(lunasvg::lunasvg ALIAS lunasvg)
+
+target_link_libraries(lunasvg PRIVATE plutovg::plutovg)
+
+set_target_properties(lunasvg PROPERTIES
+    SOVERSION ${LUNASVG_VERSION_MAJOR}
+    CXX_VISIBILITY_PRESET hidden
+    CXX_STANDARD_REQUIRED ON
+    CXX_STANDARD 17
+)
+
+target_include_directories(lunasvg PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+    ${CMAKE_CURRENT_SOURCE_DIR}/source
+)
+
+target_include_directories(lunasvg PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:include/lunasvg>
+)
 
 target_compile_definitions(lunasvg PRIVATE LUNASVG_BUILD)
[…]
