### 2025-05-21 760437e
Make necessary changes to update lunasvg to 3.3.0
