### 2025-05-21 535d61a
Squashed 'third_party/SDL/' changes from f303ccca64d..8d604353a53
8d604353a53 renderer: Always use the output size when updating the main view
8410e11ecbb wikiheaders: Man pages should escape apostrophe chars.
9f8161bf4af docs: improve man page generation
c038d6f7f84 Updated to version 3.2.14 for release
2fff37fffce Revert "x11: Send key events for dead keys consumed by the IME"
6feb86be711 Revert "x11: Filter out duplicate key presses when an IME is active"
ed22220bc6f Fixed crash if out of memory in the Vulkan GPU driver
219cb1a59df joystick: Fix linker error when building without virtual jo
[…]
