+++ src/video/x11/SDL_x11video.c
@@ -39,7 +39,6 @@
 #include "SDL_x11messagebox.h"
 #include "SDL_x11shape.h"
 #include "SDL_x11xsync.h"
-#include "SDL_x11xtest.h"
 
 #ifdef SDL_VIDEO_OPENGL_EGL
 #include "SDL_x11opengles.h"
@@ -280,7 +279,7 @@ static SDL_VideoDevice *X11_CreateDevice(void)
      * This is otherwise not wanted, as it can break fullscreen window positioning on multi-monitor configurations.
      */
     if (!X11_CheckCurrentDesktop("openbox")) {
-        device->device_caps |= VIDEO_DEVICE_CAPS_SENDS_DISPLAY_CHANGES;
+        device->device_caps |= VIDEO_DEVICE_CAPS_SENDS_FULLSCREEN_DIMENSIONS;
     }
 
     data->is_xwayland = X11_IsXWayland(x11_display);
@@ -444,17 +443,13 @@ static bool X11_VideoInit(SDL_VideoDevice *_this)
 
 #ifdef SDL_VIDEO_DRIVER_X11_XFIXES
     X11_InitXfixes(_this);
-#endif
+#endif // SDL_VIDEO_DRIVER_X11_XFIXES
 
     X11_InitXsettings(_this);
 
 #ifdef SDL_VIDEO_DRIVER_X11_XSYNC
     X11_InitXsync(_this);
-#endif
-
-#ifdef SDL_VIDEO_DRIVER_X11_XTEST
-    X11_InitXTest(_this);
-#endif
+#endif /* SDL_VIDEO_DRIVER_X11_XSYNC */
 
 #ifndef X_HAVE_UTF8_STRING
 #warning X server does not support UTF8_STRING, a feature introduced in 2000! This is likely to become a hard error in a future libSDL3.
