### 2025-05-21 f9e1f3d
Squashed 'third_party/stb/' changes from 5c205738c19..802cd454f25
802cd454f25 Create SECURITY.md
2dfbe86bef8 update stb_image_resize2 & readme
8cfb1605c02 resize 2.14 - fix off edge nans leaking in on some resize types
f0569113c93 update stb_image_resize2.h
adc1672878b Resize 2.13 - fixed a bug when using input callbacks, turned off simd for tiny-c, fixed some variables that should have been static, fixes a bug when calculating temp memory with resizes that exceed 2GB of temp memory (very large resizes).

git-subtree-dir: third_party/stb
git-subtree-split: 802cd454f25469d3123e678af41364153c132
[…]
