### 2025-05-21 8d6651b
Mention library upgrades in CHANGELOG.md
