### 2025-05-21 487bd2f
Make file filters default to something sensible
