### 2025-05-22 e207fb2
Clean up socket search bar
