+++ CHANGELOG.md
@@ -6,6 +6,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Upcoming Release]
 
+- The search bar in the "Add Component" dialog was cleaned up and now matches similar
+  search bars in other dialogs.
 - The mesh warping workflow now has a 'swap source <--> destination' button, to make it
   possible to see what the inverse of a TPS warp looks like.
 - File dialog filters now default to filtering typically-supported file extensions for
