### 2025-05-22 e2c33b4
Refactor global ui::context functions into a ui::Context class and make font getters explicit
