### 2025-05-22 37e5a81
Harden SVG reader implementation and add some basic tests for it
