### 2025-05-22 53c3139
Add+rollout ui::draw_vertical_spacer
