### 2025-05-22 5e935a2
Fix crash when spawning dialogs from within the 'Add Component' dialog (#1040)
