### 2025-05-22 2700914
Make minor UI cleanups/tweaks
