+++ libopensimcreator/UI/ModelEditor/ReassignSocketPopup.cpp
@@ -165,6 +165,7 @@ public:
 
         ui::draw_text("%s's new connectee:", socket->getName().c_str());
 
+        ui::set_next_item_width(ui::get_content_region_available().x);
         DrawSearchBar(m_EditedParams.search);
 
         ui::begin_child_panel(
