### 2025-05-23 #8006 @Kevin-Mattheus-Moerman
@modenaxe @romainmartinez @demotu @lambdaloop @cvhammond @elen4 @alcantarar @felixchenier @pariterre @clnsmith  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled "OpenSim Creator: A Graphical User Interface for Building OpenSim Musculoskeletal Models".

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/ComputationalBiomechanicsLab/opensim-creator) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08006/joss.08006/10.21105.joss.08006.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
