### 2025-05-25 6ec4422
Clean up resources dir
