+++ resources/OpenSimCreator/icons/README.md
@@ -1,12 +1,0 @@
-# `icons/`: Custom Iconset for OpenSim Creator
-
-These are custom icons that were specifically designed for OpenSim Creator
-by Adam Kewley.
-
-All icons in this directory are licensed under a
-[Creative Commons Attribution 4.0 International License][cc-by].
-
-[![CC BY 4.0][cc-by-image]][cc-by]
-
-[cc-by]: http://creativecommons.org/licenses/by/4.0/
-[cc-by-image]: https://i.creativecommons.org/l/by/4.0/88x31.png
