### 2025-05-26 #8006 @cvhammond
I am available to review this paper. The goal timeline of 2-3 weeks is ok for me but an extended timeline could be more convenient.
