### 2025-05-26 c3201d1
Add search bar to 'Add' context menu
