+++ CHANGELOG.md
@@ -6,6 +6,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Upcoming Release]
 
+- A search bar was added to the `Add` context menu, enabling users to search through
+  all available components.
 - The alignment of property editors was adjusted such that they all align on the
   left side (previously: double editors were indented slightly).
 - Fixed a crash that occured when opening a property editor that spawns an external
