### 2025-05-26 5ae6d0a
Clean up property editors
