+++ libopensimcreator/UI/Shared/PropertiesPanel.cpp
@@ -39,6 +39,7 @@ namespace
         }
 
         ui::set_num_columns(2);
+        ui::align_text_to_frame_padding();  // ensure it aligns with the button in the next column
         ui::draw_text("actions");
         ui::same_line();
         ui::draw_help_marker("Shows a menu containing extra actions that can be performed on this component.\n\nYou can also access the same menu by right-clicking the component in the 3D viewer, bottom status bar, or navigator panel.");
@@ -84,6 +85,7 @@ namespace
             ui::set_num_columns(2);
 
             ui::draw_separator();
+            ui::align_text_to_frame_padding();  // ensure it aligns with the next column
             ui::draw_text("name");
             ui::same_line();
             ui::draw_help_marker("The name of the component", "The component's name can be important. It can be used when components want to refer to eachover. E.g. a joint will name the two frames it attaches to.");
