### 2025-05-26 6d411ad
Fix a TODO in NavigatorPanel
