### 2025-05-26 5038c0b
Make center of masses look nicer (#575)
