+++ libopensimcreator/UI/MeshImporter/ChooseElLayer.h
@@ -311,7 +311,7 @@ namespace osc::mi
         void drawCancelButton()
         {
             ui::push_style_var(ui::StyleVar::FramePadding, {10.0f, 10.0f});
-            ui::push_style_color(ui::ColorVar::Button, Color::half_grey());
+            ui::push_style_color(ui::ColorVar::Button, Color::dark_grey());
 
             const CStringView text = OSC_ICON_ARROW_LEFT " Cancel (ESC)";
             const Vec2 margin = {25.0f, 35.0f};
