+++ libopensimcreator/UI/MeshWarper/MeshWarpingTabNavigatorPanel.h
@@ -151,7 +151,7 @@ namespace osc
             const Vec2 direction = normalize(dest.origin - src.origin);
             const Vec2 start = src.origin  + (src.radius  + Vec2{pad, 0.0f})*direction;
             const Vec2 end   = dest.origin - (dest.radius + Vec2{pad, 0.0f})*direction;
-            const Color color = Color::half_grey();
+            const Color color = Color::dark_grey();
             ui::get_panel_draw_list().add_line(start, end, color);
 
             // draw triangle on end of connecting line to form an arrow
@@ -223,7 +223,7 @@ namespace osc
                     return m_State->getUnpairedLandmarkColor();
                 }
             }
-            return Color::half_grey();
+            return Color::dark_grey();
         }
 
         ui::TableFlags getTableFlags() const
