### 2025-05-26 5b7a676
Fix typo in CHANGELOG.md
