### 2025-05-26 bc310ba
Fix issues found by clang-tidy (#1042)
