+++ libopensimcreator/Utils/OpenSimHelpers.cpp
@@ -1620,14 +1620,14 @@ OpenSim::Component& osc::AddComponentToAppropriateSet(OpenSim::Model& m, std::un
 OpenSim::ModelComponent& osc::AddModelComponent(OpenSim::Model& model, std::unique_ptr<OpenSim::ModelComponent>&& p)
 {
     OpenSim::ModelComponent& rv = *p;
-    model.addModelComponent(p.release());
+    model.addModelComponent(std::move(p).release());
     return rv;
 }
 
 OpenSim::Component& osc::AddComponent(OpenSim::Component& c, std::unique_ptr<OpenSim::Component>&& p)
 {
     OpenSim::Component& rv = *p;
-    c.addComponent(p.release());
+    c.addComponent(std::move(p).release());
     return rv;
 }
 
