### 2025-05-26 b89223c
Fix edge-case code that triggers bug in clang-tidy (#1042)
