### 2025-05-26 4a4fbc7
Make osc main function canonicalize relative paths (helps with OpenSim race from #1036)
