### 2025-05-26 afa6eb1
Fix minor typo in docs
