### 2025-05-26 8da240a
Update CHANGELOG.md for 0.5.23 (#1042)
