### 2025-05-27 5a729fa
Upate CITATION.cff, README.md, and codemeta.json for 0.5.23 (#1042)
