### 2025-05-27 32e948b
Add button for setting window size to 1920x1080 (used for screenshots)
