### 2025-05-27 f841b3c
Add ability to save screenshots in a JPEG format
