### 2025-05-28 51457ab
Clean up CMakeLists.txt and third_party README.md
