### 2025-05-28 9f39859
Drop custom bundle_sources script
