+++ libopensimcreator/UI/MeshWarper/MeshWarpingTabInputMeshPanel.h
@@ -481,7 +481,7 @@ namespace osc
             {
                 if (ui::draw_menu_item("Mesh File"))
                 {
-                    ActionLoadMeshFile(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
+                    ActionPromptUserToLoadMeshFile(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
                 }
                 if (ui::begin_menu("Generated Mesh")) {
                     drawGeneratedMeshOptions(SolidGeometries{});
@@ -489,12 +489,12 @@ namespace osc
                 }
                 if (ui::draw_menu_item("Landmarks from CSV"))
                 {
-                    ActionLoadLandmarksFromCSV(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
+                    ActionPromptUserToLoadLandmarksFromCSV(m_State->getUndoableSharedPtr(), m_DocumentIdentifier);
                 }
                 if (m_DocumentIdentifier == TPSDocumentInputIdentifier::Source &&
                     ui::draw_menu_item("Non-Participating Landmarks from CSV"))
                 {
-                    ActionLoadNonParticipatingLandmarksFromCSV(m_State->getUndoableSharedPtr());
+                    ActionPromptUserToLoadNonParticipatingLandmarksFromCSV(m_State->getUndoableSharedPtr());
                 }
                 ui::end_popup();
             }
@@ -522,33 +522,33 @@ namespace osc
             {
                 if (ui::draw_menu_item("Mesh to OBJ"))
                 {
-                    ActionTrySaveMeshToObjFile(m_State->getScratchMesh(m_DocumentIdentifier), ObjWriterFlag::Default);
+                    ActionPromptUserToSaveMeshToObjFile(m_State->getScratchMesh(m_DocumentIdentifier), ObjWriterFlag::Default);
                 }
                 if (ui::draw_menu_item("Mesh to OBJ (no normals)"))
                 {
-                    ActionTrySaveMeshToObjFile(m_State->getScratchMesh(m_DocumentIdentifier), ObjWriterFlag::NoWriteNormals);
+                    ActionPromptUserToSaveMeshToObjFile(m_State->getScratchMesh(m_DocumentIdentifier), ObjWriterFlag::NoWriteNormals);
                 }
                 if (ui::draw_menu_item("Mesh to STL"))
                 {
-                    ActionTrySaveMeshToStlFile(m_State->getScratchMesh(m_DocumentIdentifier));
+                    ActionPromptUserToMeshToStlFile(m_State->getScratchMesh(m_DocumentIdentifier));
                 }
                 if (ui::draw_menu_item("Landmarks to CSV"))
                 {
-                    ActionSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier);
+                    ActionPromptUserToSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier);
                 }
                 if (ui::draw_menu_item("Landmark Positions to CSV"))
                 {
-                    ActionSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier, LandmarkCSVFlags::NoHeader | LandmarkCSVFlags::NoNames);
+                    ActionPromptUserToSaveLandmarksToCSV(m_State->getScratch(), m_DocumentIdentifier, LandmarkCSVFlags::NoHeader | LandmarkCSVFlags::NoNames);
                 }
                 if (m_DocumentIdentifier == TPSDocumentInputIdentifier::Source)
                 {
                     if (ui::draw_menu_item("Non-Participating Landmarks to CSV"))
                     {
-                        ActionSaveNonParticipatingLandmarksToCSV(m_State->getScratch());
[…]
