+++ libopensimcreator/Documents/MeshWarper/UndoableTPSDocumentActions.h
@@ -9,13 +9,16 @@
 #include <liboscar/Maths/Vec3.h>
 #include <liboscar/Utils/UID.h>
 
+#include <iosfwd>
 #include <memory>
+#include <span>
 #include <string_view>
 #include <unordered_set>
 
 namespace osc { class Mesh; }
-namespace osc { struct TPSDocument; }
 namespace osc { class TPSResultCache; }
+namespace osc { struct TPSDocument; }
+namespace osc { struct TPSDocumentLandmarkPair; }
 
 // user-enactable actions
 namespace osc
@@ -70,31 +73,34 @@ namespace osc
     void ActionLoadMesh(UndoableTPSDocument&, const Mesh&, TPSDocumentInputIdentifier);
 
     // prompts the user to browse for an input mesh and assigns it to the document
-    void ActionLoadMeshFile(const std::shared_ptr<UndoableTPSDocument>&, TPSDocumentInputIdentifier);
+    void ActionPromptUserToLoadMeshFile(const std::shared_ptr<UndoableTPSDocument>&, TPSDocumentInputIdentifier);
 
     // loads landmarks from a CSV file into source/destination slot of the document
-    void ActionLoadLandmarksFromCSV(const std::shared_ptr<UndoableTPSDocument>&, TPSDocumentInputIdentifier);
+    void ActionPromptUserToLoadLandmarksFromCSV(const std::shared_ptr<UndoableTPSDocument>&, TPSDocumentInputIdentifier);
 
     // loads non-participating landmarks from a CSV file into the source input
-    void ActionLoadNonParticipatingLandmarksFromCSV(const std::shared_ptr<UndoableTPSDocument>&);
+    void ActionPromptUserToLoadNonParticipatingLandmarksFromCSV(const std::shared_ptr<UndoableTPSDocument>&);
 
     // saves all source/destination landmarks to a CSV file (matches loading)
-    void ActionSaveLandmarksToCSV(const TPSDocument&, TPSDocumentInputIdentifier, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
+    void ActionPromptUserToSaveLandmarksToCSV(const TPSDocument&, TPSDocumentInputIdentifier, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
+
+    // writes all source/destination landmark pairs with a location to the output stream in a CSV format.
+    void ActionWriteLandmarksAsCSV(std::span<const TPSDocumentLandmarkPair>, TPSDocumentInputIdentifier, lm::LandmarkCSVFlags, std::ostream&);
 
     // saves non-participating landmarks to a CSV file (matches loading)
-    void ActionSaveNonParticipatingLandmarksToCSV(const TPSDocument&, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
+    void ActionPromptUserToSaveNonParticipatingLandmarksToCSV(const TPSDocument&, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
 
     // saves all pairable landmarks in the TPS document to a user-specified CSV file
-    void ActionSavePairedLandmarksToCSV(const TPSDocument&, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
+    void ActionPromptUserToSavePairedLandmarksToCSV(const TPSDocument&, lm::LandmarkCSVFlags = lm::LandmarkCSVFlags::None);
 
     // prompts the user to save the mesh to an obj file
-    void ActionTrySaveMeshToObjFile(const Mesh&, ObjWriterFlags);
+    void ActionPromptUserToSaveMeshToObjFile(const Mesh&, ObjWriterFlags);
 
[…]

[…]
