### 2025-05-28 3cede82
Add App::prompt_user_to_select_directory_async
