+++ liboscar/Platform/App.cpp
@@ -718,15 +718,15 @@ namespace
         std::function<void()> callback_;
     };
 
-    // State that's stored in the sdl3 callback.
-    struct SDL3CallbackState final {
+    // State that's stored in the sdl3 callback when using a file dialog.
+    struct SDL3DialogCallbackState final {
 
         // This free function is what SDL calls with `SDL3CallbackState` when the user is
         // finished with the dialog.
         static void sdl3_compatible_callback(void* userdata, const char* const* filelist, int)
         {
             // Unpack callback state.
-            const std::unique_ptr<SDL3CallbackState> state{static_cast<SDL3CallbackState*>(userdata)};
+            const std::unique_ptr<SDL3DialogCallbackState> state{static_cast<SDL3DialogCallbackState*>(userdata)};
 
             // If there's an error, emit a `FileDialogResponse` that contains the error.
             if (not filelist) {
@@ -754,9 +754,9 @@ namespace
         }
 
         // Constructs the callback state that's stored in SDL3's dialog system.
-        explicit SDL3CallbackState(
+        explicit SDL3DialogCallbackState(
             std::function<void(FileDialogResponse&&)>&& callback_,
-            std::span<const FileDialogFilter> filters_) :
+            std::span<const FileDialogFilter> filters_ = {}) :
             caller_callback{std::move(callback_)},
             caller_filters(filters_.begin(), filters_.end())
         {
@@ -1050,9 +1050,9 @@ public:
         bool allow_many)
     {
         // Setup `SDL_ShowOpenFileDialog` arguments.
-        auto sdl3_callback_state = std::make_unique<SDL3CallbackState>(std::move(callback), filters);
-        const SDL_DialogFileFilter* sdl3_filters_ptr = sdl3_callback_state->sdl3_filters.data();
-        const auto sdl3_num_filters = static_cast<int>(sdl3_callback_state->sdl3_filters.size());
+        auto dialog_callback_state = std::make_unique<SDL3DialogCallbackState>(std::move(callback), filters);
+        const SDL_DialogFileFilter* sdl3_filters_ptr = dialog_callback_state->sdl3_filters.data();
+        const auto sdl3_num_filters = static_cast<int>(dialog_callback_state->sdl3_filters.size());
         std::string default_location;
         if (initial_directory_to_show) {
             default_location = initial_directory_to_show->string();
@@ -1063,8 +1063,8 @@ public:
 
         // Call into SDL3's dialog implementation.
         SDL_ShowOpenFileDialog(
-            SDL3CallbackState::sdl3_compatible_callback,
-            sdl3_callback_state.release(),
+            SDL3DialogCallbackState::sdl3_compatible_callback,
+            dialog_callback_state.release(),
             main_window_.get(),  // make it modal in the main window
             sdl3_filters_ptr,
             sdl3_num_filters,
@@ -1077,13 +1077,38 @@ public:
         request_redraw();
     }
 
+    void prompt_user_to_select_directory_async(
+        std::function<void(FileDialogResponse&&)> callback,
+        std::optional<std::filesystem::path> initial_directory_to_show,
+        bool allow_many)
+    {
+        // Setup `SDL_ShowOpenFolderDialog` arguments.
+        auto dialog_callback_state = std::make_unique<SDL3DialogCallbackState>(std::move(callback));
+        std::string default_location;
+        if (initial_directory_to_show) {
+            default_location = initial_directory_to_show->string();
+        }
+        else if (const auto fallback = get_initial_directory_to_show_fallback()) {
+            default_location = fallback->string();
+        }
+
+        // Call into SDL3's dialog implementation.
+        SDL_ShowOpenFolderDialog(
[…]
