### 2025-05-28 17c63af
Reformat UndoableTPSDocumentActions.tests.cpp
