### 2025-05-28 5638015
Remove hard package dependency on xdg-desktop-portal-gtk (not strictly required to boot)
