### 2025-05-29 a14dc60
Remove liboscar dependency on OpenSimCreator font/icon/config files
