+++ liboscar/UI/oscimgui.cpp
@@ -370,9 +370,9 @@ namespace
         }
     }
 
-    struct OscarImguiBackendData final {
+    struct UiGraphicsContextData final {
 
-        OscarImguiBackendData()
+        UiGraphicsContextData()
         {
             ui_material.set_transparent(true);
             ui_material.set_cull_mode(CullMode::Off);
@@ -390,10 +390,10 @@ namespace
 
     // Backend data stored in io.BackendRendererUserData to allow support for multiple Dear ImGui contexts
     // It is STRONGLY preferred that you use docking branch with multi-viewports (== single Dear ImGui context + multiple windows) instead of multiple Dear ImGui contexts.
-    OscarImguiBackendData* get_graphics_backend_data()
+    UiGraphicsContextData* get_graphics_backend_data()
     {
         if (ImGui::GetCurrentContext()) {
-            return static_cast<OscarImguiBackendData*>(ImGui::GetIO().BackendRendererUserData);
+            return static_cast<UiGraphicsContextData*>(ImGui::GetIO().BackendRendererUserData);
         }
         else {
             return nullptr;
@@ -419,7 +419,7 @@ namespace
     }
 
     void render_draw_command(
-        OscarImguiBackendData& bd,
+        UiGraphicsContextData& bd,
         const ImDrawData& draw_data,
         const ImDrawList&,
         Mesh& mesh,
@@ -478,7 +478,7 @@ namespace
     }
 
     void render_drawlist(
-        OscarImguiBackendData& bd,
+        UiGraphicsContextData& bd,
         const ImDrawData& draw_data,
         ImDrawList& draw_list,
         RenderTexture* maybe_target)
@@ -522,7 +522,7 @@ namespace
     template<SameAsAnyOf<Texture2D, RenderTexture> Texture>
     ImTextureID allocate_texture_for_current_frame(const Texture& texture)
     {
-        OscarImguiBackendData* bd = get_graphics_backend_data();
+        UiGraphicsContextData* bd = get_graphics_backend_data();
         OSC_ASSERT(bd != nullptr && "no oscar ImGui renderer backend was available to shutdown - this is a developer error");
         const UID texture_uid = bd->textures_allocated_this_frame.try_emplace(UID{}, texture).first->first;
         return to_imgui_texture_id(texture_uid);
@@ -548,14 +548,14 @@ namespace
         OSC_ASSERT(io.BackendRendererUserData == nullptr && "an oscar ImGui renderer backend is already initialized - this is a developer error (double-initialization)");
 
         // init backend data
-        io.BackendRendererUserData = static_cast<void*>(new OscarImguiBackendData{});
+        io.BackendRendererUserData = static_cast<void*>(new UiGraphicsContextData{});
         io.BackendRendererName = "imgui_impl_osc";
         io.BackendFlags |= ImGuiBackendFlags_RendererHasVtxOffset;
     }
 
     void graphics_backend_shutdown()
     {
-        OscarImguiBackendData* bd = get_graphics_backend_data();
+        UiGraphicsContextData* bd = get_graphics_backend_data();
         OSC_ASSERT(bd != nullptr && "no oscar ImGui renderer backend was available to shutdown - this is a developer error (double-free)");
 
         // shutdown platform interface
@@ -572,7 +572,7 @@ namespace
     {
         // `ImGui_ImplOpenGL3_CreateDeviceObjects` is now part of constructing `OscarImguiBackendData`
 
-        OscarImguiBackendData* bd = get_graphics_backend_data();
+        UiGraphicsContextData* bd = get_graphics_backend_data();
         OSC_ASSERT(bd != nullptr && "no oscar ImGui renderer backend was available - this is a developer error");
[…]
