+++ liboscar/Platform/IconCodepoints.h
@@ -1,4 +1,4 @@
-// This was generated by OpenSim Creator's compile_svgs_to_ttf.py script
+// This was generated by `compile_svgs_to_ttf.py` script
 
 #define OSC_ICON_MIN 0xf002
 #define OSC_ICON_MAX 0xf84c
