### 2025-05-30 d90af7f
Remove now-non-existent oscar/ resource dir from Packaging.cmake files
