+++ osc/Windows/Packaging.cmake
@@ -33,7 +33,6 @@ install(
 install(
     DIRECTORY
         "${PROJECT_SOURCE_DIR}/resources/OpenSimCreator"
-        "${PROJECT_SOURCE_DIR}/resources/oscar"
         "$<$<BOOL:${OSC_BUNDLE_OSCAR_DEMOS}>:${PROJECT_SOURCE_DIR}/resources/oscar_demos>"
     DESTINATION
         "resources/"
