### 2025-06-03 11ed53c
Clarify Texture2D encoding and coordinate system, fix related issues (#1044)
