+++ liboscar/Graphics/Texture2D.h
@@ -16,10 +16,10 @@
 
 namespace osc
 {
-    // A 2D texture that can be rendered by the graphics backend.
+    // Represents a 2D image array that can be read by `Shader`s.
     class Texture2D final {
     public:
-        // Default-constructs a single-pixel texture as a placeholder
+        // Default constructs a single-pixel texture.
         Texture2D() :
             Texture2D{Vec2i{1, 1}}
         {}
@@ -37,28 +37,33 @@ namespace osc
 
         // Returns the dimensions of the texture in device-independent pixels.
         //
-        // These dimensions should be used when compositing the texture in a
-        // user interface.
+        // - These dimensions should be used when compositing the texture into a
+        //   high DPI user interface.
         //
-        // The return value is equivalent to `texture.dimensions() / texture.device_pixel_ratio()`.
+        // - Returns the equivalent of `texture.dimensions() / texture.device_pixel_ratio()`.
         Vec2 device_independent_dimensions() const;
 
         // Returns the ratio of the resolution of the texture in physical pixels
-        // to the resolution of it in device-independent pixels.
+        // to the resolution of it in device-independent pixels. This is useful
+        // when compositing the texture into high DPI user interfaces that are built
+        // around a device-independent pixel scaling architecture.
         float device_pixel_ratio() const;
 
         // Sets the device-to-pixel ratio for the texture, which has the effect
         // of scaling the `device_independent_dimensions()` of the texture.
         void set_device_pixel_ratio(float);
 
-        // Returns the format of the underlying pixel data.
+        // Returns the storage format of the underlying pixel data.
         TextureFormat texture_format() const;
 
         // Returns the color space of the texture.
         ColorSpace color_space() const;
 
-        TextureWrapMode wrap_mode() const;  // same as wrap_mode_u
-        void set_wrap_mode(TextureWrapMode);  // sets all axes
+        // Returns the equivalent of `wrap_mode_u`.
+        TextureWrapMode wrap_mode() const;
+
+        // Sets all wrap axes (`u`, `v`, and `w`) to the specified `TextureWrapMode`.
+        void set_wrap_mode(TextureWrapMode);
         TextureWrapMode wrap_mode_u() const;
         void set_wrap_mode_u(TextureWrapMode);
         TextureWrapMode wrap_mode_v() const;
@@ -69,19 +74,51 @@ namespace osc
         TextureFilterMode filter_mode() const;
         void set_filter_mode(TextureFilterMode);
 
-        // - must contain pixels row-by-row
-        // - the size of the span must equal the width*height of the texture
-        // - may internally convert the provided `Color` structs into the format
-        //   of the texture, so don't expect `pixels` to necessarily return
-        //   exactly the same values as provided
+        // Returns the pixels, parsed into a `Color` (i.e. HDR sRGB RGBA) format, where:
+        //
+        // - Pixels are returned row-by-row, where:
+        //   - The first pixel corresponds to the lower-left corner of the image.
+        //   - Subsequent pixels progress left-to-right through the remaining pixels in the
+        //     lowest row of the image, and then in successively higher rows of the image.
+        //   - The final pixel corresponds to the upper-right corner of the image.
+        //   - Note: this right-handed coordinate system matches samplers in GLSL shaders. That
+        //     is, a texture/uv coordinate of `(0, 0)` sampled in a shader would sample the
+        //     bottom-left pixel of the texture in GLSL.
+        // - The returned pixels are parsed from the underlying `TextureFormat` storage. If
+        //   the storage format has fewer components than a `Color` (RGBA), the missing
+        //   components default to `0.0f` - apart from alpha, which defaults to `1.0f`.
         std::vector<Color> pixels() const;
+
+        // Assigns the given pixels to the texture.
+        //
+        // - Pixels should be provided row-by-row, where:
[…]
