### 2025-06-03 799a42c
Fix incorrect resolution in ScreenshotTab 'change' (resolution) button
