+++ liboscar/UI/Tabs/ScreenshotTab.cpp
@@ -128,7 +128,7 @@ public:
                 Vec2 s = App::get().main_window_dimensions();
                 ui::draw_text("%f %f", s.x, s.y);
                 if (ui::draw_button("change")) {
-                    App::upd().try_async_set_main_window_dimensions({1980.0f, 1080.0f});
+                    App::upd().try_async_set_main_window_dimensions({1920.0f, 1080.0f});
                 }
             }
 
