### 2025-06-03 eef7cd9
Fix bad osc::Rect initialization in osc::ui::draw_image
