+++ liboscar/UI/oscimgui.h
@@ -698,7 +698,7 @@ namespace osc::ui
     void draw_image(
         const Texture2D& texture,
         std::optional<Vec2> dimensions = std::nullopt,
-        const Rect& region_uv_coordinates = Rect({0.0f, 0.0f}, {1.0f, 1.0f})
+        const Rect& region_uv_coordinates = Rect{{0.0f, 0.0f}, {1.0f, 1.0f}}
     );
     void draw_image(
         const RenderTexture&
