### 2025-06-03 e94451e
Flesh out model warper docs slightly
