### 2025-06-03 14f76fe
Unify 'Add' menus, make it possible to add any component to any other one
