+++ libopensimcreator/UI/Shared/ModelAddMenuItems.cpp
@@ -3,24 +3,34 @@
 #include <libopensimcreator/ComponentRegistry/ComponentRegistry.h>
 #include <libopensimcreator/ComponentRegistry/StaticComponentRegistries.h>
 #include <libopensimcreator/Documents/Model/IModelStatePair.h>
-#include <libopensimcreator/UI/Shared/BasicWidgets.h>
+#include <libopensimcreator/Documents/Model/UndoableModelActions.h>
 #include <libopensimcreator/UI/ModelEditor/AddBodyPopup.h>
 #include <libopensimcreator/UI/ModelEditor/AddComponentPopup.h>
+#include <libopensimcreator/UI/ModelEditor/Select1PFPopup.h>
+#include <libopensimcreator/UI/ModelEditor/SelectComponentPopup.h>
+#include <libopensimcreator/UI/ModelEditor/SelectGeometryPopup.h>
+#include <libopensimcreator/UI/Shared/BasicWidgets.h>
 #include <libopensimcreator/Utils/OpenSimHelpers.h>
 
 #include <liboscar/Platform/App.h>
 #include <liboscar/Platform/Widget.h>
 #include <liboscar/Platform/WidgetPrivate.h>
+#include <liboscar/Shims/Cpp23/ranges.h>
 #include <liboscar/UI/Events/OpenPopupEvent.h>
 #include <liboscar/UI/oscimgui.h>
 #include <liboscar/Utils/CStringView.h>
 #include <liboscar/Utils/LifetimedPtr.h>
 #include <liboscar/Utils/StringHelpers.h>
 #include <OpenSim/Common/Component.h>
+#include <OpenSim/Common/ComponentPath.h>
 #include <OpenSim/Simulation/Control/Controller.h>
 #include <OpenSim/Simulation/Model/Actuator.h>
 #include <OpenSim/Simulation/Model/ContactGeometry.h>
+#include <OpenSim/Simulation/Model/ContactGeometry.h>
 #include <OpenSim/Simulation/Model/Force.h>
+#include <OpenSim/Simulation/Model/HuntCrossleyForce.h>
+#include <OpenSim/Simulation/Model/Model.h>
+#include <OpenSim/Simulation/Model/PathActuator.h>
 #include <OpenSim/Simulation/Model/Probe.h>
 #include <OpenSim/Simulation/SimbodyEngine/Constraint.h>
 #include <OpenSim/Simulation/SimbodyEngine/Joint.h>
@@ -41,6 +51,11 @@ public:
         m_Model{std::move(uum_)}
     {}
 
+    void setTargetParentComponent(const OpenSim::ComponentPath& path)
+    {
+        m_MaybeTargetParentComponent = path;
+    }
+
     void onDraw()
     {
         ui::push_id(this);
@@ -60,6 +75,8 @@ public:
             drawSearchResultsOrNoResults();
         }
 
+        drawTargetComponentSpecializedAdders();
+
         if (disabled) {
             ui::end_disabled();
         }
@@ -68,6 +85,170 @@ public:
     }
 
 private:
+    void drawTargetComponentSpecializedAdders()
+    {
+        const OpenSim::Component* component = FindComponent(m_Model->getModel(), m_MaybeTargetParentComponent);
+        if (not component) {
+            return;
+        }
+
+        if (const auto* joint = dynamic_cast<const OpenSim::Joint*>(component)) {
+            ui::draw_separator();
+            drawSpecializedContextualActions(*joint);
+        }
+        else if (const auto* hcf = dynamic_cast<const OpenSim::HuntCrossleyForce*>(component)) {
+            ui::draw_separator();
+            drawSpecializedContextualActions(*hcf);
+        }
[…]
