### 2025-06-03 10a61ec
Add more copy actions to the component context menu
