+++ .github/workflows/continuous-integration-workflow.yml
@@ -110,6 +110,6 @@ jobs:
 
     - uses: actions/upload-artifact@v4
       with:
-        name: OpenSimCreator_Deb.zip
+        name: opensimcreator-ubuntu22-arm64.zip
         path: build/*.deb
         if-no-files-found: error
