### 2025-06-04 177c645
Move IconCodepoints.h from liboscar to libopensimcreator so it isn't strongly coupled
