+++ libopensimcreator/UI/ModelEditor/ModelEditorMainMenu.cpp
@@ -3,6 +3,7 @@
 #include <libopensimcreator/Documents/Model/Environment.h>
 #include <libopensimcreator/Documents/Model/UndoableModelActions.h>
 #include <libopensimcreator/Documents/Model/UndoableModelStatePair.h>
+#include <libopensimcreator/Platform/IconCodepoints.h>
 #include <libopensimcreator/UI/ModelEditor/ExportPointsPopup.h>
 #include <libopensimcreator/UI/ModelEditor/ModelMusclePlotPanel.h>
 #include <libopensimcreator/UI/PerformanceAnalyzerTab.h>
@@ -13,7 +14,6 @@
 #include <libopensimcreator/Utils/OpenSimHelpers.h>
 
 #include <liboscar/Platform/App.h>
-#include <liboscar/Platform/IconCodepoints.h>
 #include <liboscar/Platform/WidgetPrivate.h>
 #include <liboscar/UI/Events/OpenPopupEvent.h>
 #include <liboscar/UI/Events/OpenTabEvent.h>
