### 2025-06-04 3ca68e3
Move mdspan shim into libopensimcreator (liboscar doesn't use it)
