+++ libopensimcreator/Utils/TPS3D.h
@@ -1,10 +1,10 @@
 #pragma once
 
+#include <libopensimcreator/Shims/Cpp23/mdspan.h>
 #include <libopensimcreator/Utils/LandmarkPair3D.h>
 
 #include <liboscar/Graphics/Mesh.h>
 #include <liboscar/Maths/Vec3.h>
-#include <liboscar/Shims/Cpp23/mdspan.h>
 
 #include <concepts>
 #include <iosfwd>
