### 2025-06-04 ebce6fa
Remove mention of OpenSim from some liboscar READMEs
