### 2025-06-05 e5125cd
Squashed 'third_party/imgui/' changes from da3baddf9d2..4806a1924ff
REVERT: da3baddf9d2 Patch dockspace viewport label for backwards compat with opensim-creator imgui.ini configs
REVERT: 198718bc3e7 Add CMakeLists.txt that's suitable for opensim-creator

git-subtree-dir: third_party/imgui
git-subtree-split: 4806a1924ff6181180bf5e4b8b79ab4394118875
