### 2025-06-05 7ff7c7f
Mention panel location breakage in CHANGELOG.md
