### 2025-06-05 7d3020a
Update imgui_base_config.ini for compatibility with newer imgui hashing method etc.
