+++ resources/OpenSimCreator/imgui_base_config.ini
@@ -1,618 +1,0 @@
-[Window][DockSpaceViewport_11111111]
-Pos=0,77
-Size=1920,903
-Collapsed=0
-
-[Window][Debug##Default]
-Pos=507,337
-Size=400,400
-Collapsed=0
-
-[Window][Splash screen]
-Pos=640,317
-Size=640,512
-Collapsed=0
-
-[Window][Loading message]
-Pos=775,331
-Size=682,67
-Collapsed=0
-
-[Window][render]
-Pos=0,48
-Size=1920,811
-Collapsed=0
-DockId=0x00000034,0
-
-[Window][Hierarchy]
-Pos=0,48
-Size=323,549
-Collapsed=0
-DockId=0x0000001B,0
-
-[Window][Muscles]
-Pos=1523,19
-Size=397,558
-Collapsed=0
-DockId=0x00000020,0
-
-[Window][Outputs]
-Pos=852,46
-Size=1068,1011
-Collapsed=0
-DockId=0x00000010,0
-
-[Window][Utils]
-Pos=1523,19
-Size=397,558
-Collapsed=0
-DockId=0x00000020,2
-
-[Window][Moment Arms]
-Pos=1523,19
-Size=397,558
-Collapsed=0
-DockId=0x00000020,3
-
-[Window][Selection]
-Pos=0,507
-Size=292,419
-Collapsed=0
-DockId=0x00000020,0
-
-[Window][Coordinates]
-Pos=0,77
-Size=270,484
-Collapsed=0
-DockId=0x00000020,1
-
-[Window][Simulate]
-Pos=1523,19
-Size=397,558
-Collapsed=0
-DockId=0x00000020,5
-
-[Window][UI]
-Pos=0,19
-Size=397,1117
-Collapsed=0
-DockId=0x00000020,5
-
-[Window][Error loading]
-Pos=60,60
-Size=1699,570
-Collapsed=0
-
-[Window][render2]
-Pos=364,19
-Size=1157,1117
-Collapsed=0
-DockId=0x00000034,1
-
-[Window][Edit Props]
-Pos=0,620
-Size=353,517
-Collapsed=0
-DockId=0x00000020,1
-
-[Window][Snapshots]
-Pos=1523,0
-Size=397,568
-Collapsed=0
-DockId=0x00000020,1
-
-[Window][Actions]
-Pos=355,832
-Size=1565,27
-Collapsed=0
-DockId=0x00000008,0
-
-[Window][fatal exception]
-Pos=79,62
-Size=1647,360
-Collapsed=0
-
-[Window][Dear ImGui Demo]
-Pos=650,45
-Size=550,680
-Collapsed=0
-
-[Window][##render]
-Size=1920,1136
-Collapsed=0
-DockId=0x00000020,0
-
-[Window][render1]
-Pos=364,19
-Size=1157,1117
-Collapsed=0
-DockId=0x00000034,0
-
-[Window][editor]
-Pos=60,60
-Size=61,54
-Collapsed=0
-
-[Window][main panel]
-Pos=325,115
-Size=114,48
-Collapsed=0
-
-[Window][root]
-Pos=133,293
-Size=374,296
-Collapsed=0
-
-[Window][Stacked 1]
-Pos=755,482
-Size=409,172
-Collapsed=0
-
-[Window][Delete?]
-Pos=804,509
-Size=311,118
-Collapsed=0
-
-[Window][attach geometry]
-Pos=825,343
-Size=270,449
-Collapsed=0
-
-[Window][Add Freejoint]
-Pos=471,365
-Size=96,314
-Collapsed=0
-
-[Window][Add PinJoint]
-Pos=704,382
-Size=512,371
-Collapsed=0
-
[…]
