### 2025-06-05 1cd3ced
Rearrange model warping documentation assets slightly
