### 2025-06-05 7a01f68
Refactor custom font API to overwrite both fonts at the same time
