### 2025-06-05 6a379e5
Flesh out model warping documentation a little more
