+++ docs/source/the-model-warper.rst
@@ -153,33 +153,134 @@ Add a Mesh Warping Step
 The model warper is designed around applying scaling steps to the source model one-by-one
 to produce the result model. :numref:`model-warper-after-loading-model` shows the most
 trivial case of this process, which is to apply no scaling steps and produce a result
-model that's identical to the source model. The essence of building a model warping
-procedure is to incrementally add the scaling steps you need.
+model that's identical to the source model.
 
-The first step is to apply the subject's (target) femoral torsion to the femur bone
-mesh in the model. There are external tools available online to do this (e.g.
-`this one <https://simtk.org/projects/bone_deformity>`_) but, for this walkthrough, we
-will use the Thin-Plate Spline technique, as described in :doc:`the-mesh-warper`.
+The essence of building a model warping procedure is to incrementally add the scaling
+steps you need. The first step is to apply the subject's (target) femoral torsion to
+the femur bone mesh in the model. There are external tools available online to do
+this (e.g. `this one <https://simtk.org/projects/bone_deformity>`_) but, for this
+walkthrough, we will use the Thin-Plate Spline technique, as described in
+:doc:`the-mesh-warper`, because the model warper has in-built support for it.
 
-**TODO**: walk through adding a Thin-Plate Spline scaling step for a mesh in the model. Mention any gotchas w.r.t. where the data should be stored, how it should be stored, etc.
+.. note::
 
+  In preparation for the mesh warping step, we have already established a TPS warp
+  *from* the source femur mesh in the model *to* a subject-specific femur mesh by
+  pairing landmarks between them in :doc:`the-mesh-warper`. Here's a screenshot of
+  how that looked:
+
+  .. figure:: _static/the-model-warper/mesh-warper-showing-basic-TPS-warp-of-femur.jpeg
+    :width: 60%
+
+    Screenshot of how landmarks were paired between the source mesh and the destination one
+    in the mesh warper. Rotation and translation (i.e. reorientation) of the mesh was removed
+    from the TPS warp using the appropriate checkboxes to correct for subject/scanner
+    orientation. Destination data was pre-scaled by 0.001 to account for a difference in
+    units between the mesh files (meters vs. millimeters).
+
+  For the pruposes of model warping, all you need to know is that the TPS warping technique
+  requires a ``.landmarks.csv`` for the "source" and a ``.landmarks.csv`` for the destination.
+  Where the landmarks in those files come from is up to you (:doc:`the-mesh-warper` is one way).
+  The model warping implementation uses pairs of landmarks from those files to warp, scale, reorient,
+  and translate the applicable mesh, station, or muscle point.
+
+The model warper's TPS-based mesh scaling step requires two sequences of landmarks. The model
+download zip (**TODO**) includes a ``Geometry/`` directory that contains ``femur_r.landmarks.csv``
+and ``subject_femur_r.landmarks.csv``, which represent ``femur_r.vtp``\'s landmarks and landmarks
+``subject_femur_r.stl``\'s landmarks respectively.
+
+To add a scaling step in the model warper UI, click the appropriate button and add a "Apply Thin-Plate
+Spline (TPS) to Meshes" step (pictured in :numref:`model-warper-apply-tps-to-meshes-button`).
+
+.. _model-warper-apply-tps-to-meshes-button:
+.. figure:: _static/the-model-warper/apply-thin-plate-spline-to-meshes-scaling-step-button.jpeg
+    :width: 60%
+
+    The "Add Scaling Step" button in the model warper UI opens a menu where you can select
[…]
