### 2025-06-10 6913c21
Add more content to model warper documentation
