+++ docs/source/the-model-warper.rst
@@ -246,19 +246,132 @@ in the appropriate input boxes. Once you do that, you should end up with somethi
 .. figure:: _static/the-model-warper/after-applying-tps-mesh-warp.jpeg
     :width: 60%
 
-    The model warping UI after applying the mesh warping step. The warped mesh is shorter
+    The model after applying the mesh warping step. The warped mesh is shorter
     and slightly twisted when compared to the source mesh. Warping the joint frames, muscle
     points, and wrap geometry is handled later in this walkthrough. An easy way to see what
     a scaling step is doing is to toggle the ``enabled`` button on the step.
 
+
 Add a Frame Warping Step
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
-**TODO**: Explain why ``StationDefinedFrame``\s are relevant for this step.
+Warping the femur mesh only warps the mesh data while keeping the rest of the model the
+same. This means the model now looks wrong (the femur mesh is too small compared to the
+rest of the model). To fix that, we need to scale the remaining components.
 
-**TODO**: explain adding the scaling step, and any unusual-looking behavior (e.g. of
-muscle points or bodies snapping around because they are temporarily in a new coordinate
-system?)
+The next component we recommend scaling is the joint frame of the knee. The reason why
+is because it will help move the downstream components (the wrap cylinder and lower-leg
+muscle points) into a better location. As for *how* to scale the knee frame, it makes
+sense to use the TPS technique, given the femur was already scaled that way.
+
+.. note::
+
+  The TPS technique operates on points in space, not frames, which typically combine a
+  reorientation and a translation.
+
+  The way we work around that problem is by ensuring that frames in the source model use
+  ``StationDefinedFrame``\s, as described in :doc:`station-defined-frames`, rather than
+  ``PhysicalOffsetFrame``\s, which encode an orientation.
+
+  The difference between the two frame definitions is subtle, but significant: ``StationDefinedFrame``\s
+  are entirely defined in terms of points in space, which can be warped via the TPS technique,
+  whereas ``PhysicalOffsetFrame``\s include a not-TPS-warpable reorientation.
+
+  If you have a model that uses ``PhysicalOffsetFrame``\s, then the TPS technique can only
+  reliably warp the translation part of the frame. It can *approximately* warp the
+  orientation, but this relies on reprojection/orthogonalization techniques, which
+  require tweaking and aren't anywhere near as robust as using a ``StationDefinedFrame``\.
+
+In this model, the knee joint (``knee_r``) is defined as joining ``femur_r_knee_frame``
+(a ``StationDefinedFrame``) to the lower leg (``lower_leg_r``). Therefore, to warp the
+femur's side of ``knee_r``'s joint definition we should warp the stations that are used
+to define ``femur_r_knee_frame`` (listed in :numref:`stations-to-warp-for-knee-definition`).
+
+.. _stations-to-warp-for-knee-definition:
+.. list-table:: Stations that are used to define ``femur_r_knee_frame`` in the model. These should be warped via the TPS technique.
+   :header-rows: 1
+
+   * - .. centered:: Absolute Path in Model
+   * - .. centered:: ``/bodyset/femur_r/femur_r_head``
+   * - .. centered:: ``/bodyset/femur_r/femur_r_med_condyl``
+   * - .. centered:: ``/bodyset/femur_r/femur_r_lat_condyl``
+   * - .. centered:: ``/bodyset/femur_r/femur_r_condyls_midpoint``
+
+Concretely, we start by adding a "Apply Thin-Plate Spline (TPS) to Stations" scaling step to the
+model by clicking the appropriate menu item (:numref:`model-warper-apply-tps-to-stations-button`):
+
[…]
