### 2025-06-11 368dc3e
Revert imgui to old hashing/ID method (backwards compat with 0.5.23)
