+++ resources/OpenSimCreator/imgui_base_config.ini
@@ -1,262 +1,618 @@
+[Window][DockSpaceViewport_11111111]
+Pos=0,77
+Size=1920,903
+Collapsed=0
+
 [Window][Debug##Default]
-Pos=60,81
+Pos=507,337
 Size=400,400
 Collapsed=0
 
 [Window][Splash screen]
-Pos=640,401
+Pos=640,317
 Size=640,512
 Collapsed=0
 
-[Window][WindowOverViewport_11111111]
-Pos=0,77
-Size=1920,951
+[Window][Loading message]
+Pos=775,331
+Size=682,67
 Collapsed=0
 
-[Window][Source Mesh]
-Pos=0,83
-Size=615,1065
+[Window][render]
+Pos=0,48
+Size=1920,811
 Collapsed=0
-DockId=0x00000003,0
+DockId=0x00000034,0
 
-[Window][Destination Mesh]
-Pos=617,83
-Size=691,1065
-Collapsed=0
-DockId=0x00000004,0
-
-[Window][Result]
-Pos=1310,83
-Size=610,1065
-Collapsed=0
-DockId=0x0000002C,0
-
-[Window][Log]
-Pos=285,823
-Size=1635,205
-Collapsed=0
-DockId=0x00000006,0
-
-[Window][Control Panel]
-Pos=0,83
-Size=413,1094
-Collapsed=0
-DockId=0x00000007,0
-
-[Window][Source Model]
-Pos=415,83
-Size=703,887
-Collapsed=0
-DockId=0x0000002C,0
-
-[Window][Result Model]
-Pos=1120,83
-Size=800,887
-Collapsed=0
-DockId=0x0000000A,0
-
-[Window][Navigator]
-Pos=0,77
-Size=283,511
-Collapsed=0
-DockId=0x0000000D,0
-
-[Window][viewer0]
-Pos=285,77
-Size=1635,744
-Collapsed=0
-DockId=0x0000002C,0
-
-[Window][Coordinates]
-Pos=0,77
-Size=283,511
-Collapsed=0
-DockId=0x0000000D,1
-
-[Window][Properties]
-Pos=0,590
-Size=283,438
-Collapsed=0
-DockId=0x0000000E,0
-
-[Window][navigator]
-Pos=0,46
-Size=280,1131
-Collapsed=0
-DockId=0x00000029,0
-
-[Window][wizard_3dViewer]
-Pos=282,46
-Size=1638,1131
-Collapsed=0
-DockId=0x0000002C,0
-
-[Window][Loading Message]
-Pos=672,506
-Size=576,56
-Collapsed=0
-
-[Window][Output Watches]
-Pos=1520,77
-Size=400,532
-Collapsed=0
-DockId=0x00000011,0
-
-[Window][Simulation Details]
-Pos=1507,615
-Size=413,533
-Collapsed=0
-DockId=0x00000012,0
-
-[Window][viewer1]
-Pos=1354,77
-Size=566,442
-Collapsed=0
-DockId=0x00000019,0
-
-[Window][muscleplot0]
-Pos=1542,77
-Size=378,537
-Collapsed=0
-DockId=0x00000023,0
-
-[Window][viewer2]
-Pos=1354,521
-Size=566,420
+[Window][Hierarchy]
+Pos=0,48
+Size=323,549
 Collapsed=0
 DockId=0x0000001B,0
 
+[Window][Muscles]
+Pos=1523,19
+Size=397,558
+Collapsed=0
+DockId=0x00000020,0
+
+[Window][Outputs]
+Pos=852,46
+Size=1068,1011
+Collapsed=0
+DockId=0x00000010,0
+
+[Window][Utils]
+Pos=1523,19
[…]
