### 2025-06-11 747352a
Flesh out model warper documentation a little more
