+++ libopensimcreator/UI/ModelWarper/ModelWarperTab.cpp
@@ -2328,6 +2328,7 @@ namespace
                 if (ui::draw_button(OSC_ICON_PLAY " Export Warped Model")) {
                     m_State->exportWarpedModelToModelEditor();
                 }
+                App::upd().add_frame_annotation("Export Warped Model Button", ui::get_last_drawn_item_screen_rect());
                 ui::pop_style_color();
                 ui::same_line(0.0f, 1.0f);
                 if (ui::draw_button(OSC_ICON_COG)) {
