### 2025-06-12 f48d49f
Reword 'viewport' to 'main_window' or 'workspace' where appropriate
