+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLParallaxMappingTab.cpp
@@ -100,7 +100,7 @@ public:
             graphics::draw(cube_mesh_, light_transform_, light_cube_material_, camera_);
         }
 
-        camera_.set_pixel_rect(ui::get_main_viewport_workspace_screenspace_rect());
+        camera_.set_pixel_rect(ui::get_main_window_workspace_screenspace_rect());
         camera_.render_to_screen();
 
         ui::begin_panel("controls");
