+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLShadowMappingTab.cpp
@@ -96,8 +96,8 @@ public:
 private:
     void draw_3d_scene()
     {
-        const Rect viewport_screen_space_rect = ui::get_main_viewport_workspace_screenspace_rect();
-        const Vec2 top_left = top_left_rh(viewport_screen_space_rect);
+        const Rect workspace_screenspace_rect = ui::get_main_window_workspace_screenspace_rect();
+        const Vec2 top_left = top_left_rh(workspace_screenspace_rect);
         constexpr float depth_overlay_size = 200.0f;
 
         render_shadows_to_depth_texture();
@@ -111,7 +111,7 @@ private:
         scene_material_.set("uShadowMapTexture", depth_texture_);
 
         draw_meshes_with_material(scene_material_);
-        camera_.set_pixel_rect(viewport_screen_space_rect);
+        camera_.set_pixel_rect(workspace_screenspace_rect);
         camera_.render_to_screen();
         camera_.set_pixel_rect(std::nullopt);
         graphics::blit_to_screen(depth_texture_, Rect{top_left - Vec2{0.0f, depth_overlay_size}, top_left + Vec2{depth_overlay_size, 0.0f}});
