### 2025-06-12 23f0b00
Clean out ImGui multi-viewport, make main_window_ methods clearer
Clean up oscimgui.cpp a little
