+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLParallaxMappingTab.cpp
@@ -84,7 +84,7 @@ public:
         camera_.on_draw();
 
         // clear screen and ensure camera has correct pixel rect
-        App::upd().clear_screen(Color::dark_grey());
+        App::upd().clear_main_window(Color::dark_grey());
 
         // draw normal-mapped quad
         {
