+++ libopensimcreator/UI/SplashTab.cpp
@@ -261,7 +261,7 @@ private:
             auto tab = std::make_unique<mi::MeshImporterTab>(parent());
             App::post_event<OpenTabEvent>(*parent(), std::move(tab));
         }
-        App::upd().add_frame_annotation("SplashTab/ImportMeshesMenuItem", ui::get_last_drawn_item_screen_rect());
+        App::upd().add_main_window_frame_annotation("SplashTab/ImportMeshesMenuItem", ui::get_last_drawn_item_screen_rect());
         if (ui::draw_menu_item(OSC_ICON_BOOK " Open Documentation")) {
             open_url_in_os_default_web_browser(OpenSimCreatorApp::get().docs_url());
         }
@@ -282,13 +282,13 @@ private:
             auto tab = std::make_unique<MeshWarpingTab>(parent());
             App::post_event<OpenTabEvent>(*parent(), std::move(tab));
         }
-        App::upd().add_frame_annotation("SplashTab/MeshWarpingMenuItem", ui::get_last_drawn_item_screen_rect());
+        App::upd().add_main_window_frame_annotation("SplashTab/MeshWarpingMenuItem", ui::get_last_drawn_item_screen_rect());
 
         if (ui::draw_menu_item(OSC_ICON_MAGIC " Model Warping (" OSC_ICON_MAGIC " experimental)")) {
             auto tab = std::make_unique<ModelWarperTab>(parent());
             App::post_event<OpenTabEvent>(*parent(), std::move(tab));
         }
-        App::upd().add_frame_annotation("SplashTab/ModelWarpingMenuItem", ui::get_last_drawn_item_screen_rect());
+        App::upd().add_main_window_frame_annotation("SplashTab/ModelWarpingMenuItem", ui::get_last_drawn_item_screen_rect());
     }
 
     void drawRecentlyOpenedFilesMenuSectionContent(int& imguiID)
