### 2025-06-13 62c9224
Unify osc::Screen into osc::Widget and refactor many uses of 'screen'
