+++ liboscar-demos/learnopengl/PBR/LOGLPBRSpecularIrradianceTexturedTab.cpp
@@ -236,7 +236,7 @@ public:
 
     void on_draw()
     {
-        const Rect workspace_screenspace_rect = ui::get_main_window_workspace_screenspace_rect();
+        const Rect workspace_screenspace_rect = ui::get_main_window_workspace_screen_space_rect();
         const float device_pixel_ratio = App::get().main_window_device_pixel_ratio();
         const Vec2 workspace_pixel_dimensions = device_pixel_ratio * dimensions_of(workspace_screenspace_rect);
 
@@ -247,7 +247,7 @@ public:
         camera_.on_draw();
         draw_3d_render();
         draw_background();
-        graphics::blit_to_screen(output_render_, workspace_screenspace_rect);
+        graphics::blit_to_main_window(output_render_, workspace_screenspace_rect);
         perf_panel_.on_draw();
     }
 
