+++ liboscar-demos/CustomWidgetsTab.cpp
@@ -56,7 +56,7 @@ namespace
 
         const float title_height = ui::get_text_line_height_in_current_panel();
 
-        const Vec2 p = ui::get_cursor_screen_pos();
+        const Vec2 p = ui::get_cursor_ui_pos();
         const Vec2 bb(ui::get_column_width(), ui::get_frame_height());
         ui::push_style_var(ui::StyleVar::ButtonTextAlign, {0.0f, 0.0f});
         ui::push_id(label);
