+++ libopensimcreator/UI/LoadingTab.cpp
@@ -101,7 +101,7 @@ public:
 
     void onDraw()
     {
-        const Rect workspaceUIRect = ui::get_main_window_workspace_uiscreenspace_rect();
+        const Rect workspaceUIRect = ui::get_main_window_workspace_ui_rect();
         const Vec2 workspaceDims = dimensions_of(workspaceUIRect);
         const Vec2 menuDimsGuess = {0.3f * workspaceDims.x, 6.0f * ui::get_font_base_size()};
 
