+++ liboscar/Maths/MathsImplementation.cpp
@@ -646,21 +646,21 @@ Vec3 osc::PolarPerspectiveCamera::position() const
     return PolarToCartesian(focus_point, radius, theta, phi);
 }
 
-Vec2 osc::PolarPerspectiveCamera::project_onto_screen_rect(
+Vec2 osc::PolarPerspectiveCamera::project_onto_viewport(
     const Vec3& worldspace_location,
-    const Rect& screen_rect) const
+    const Rect& viewport_rect) const
 {
-    return osc::project_onto_screen_rect(
+    return osc::project_onto_viewport_rect(
         worldspace_location,
         view_matrix(),
-        projection_matrix(aspect_ratio_of(screen_rect)),
-        screen_rect
+        projection_matrix(aspect_ratio_of(viewport_rect)),
+        viewport_rect
     );
 }
 
 Line osc::PolarPerspectiveCamera::unproject_topleft_pos_to_world_ray(Vec2 pos, Vec2 dimensions) const
 {
-    return perspective_unproject_topleft_screen_pos_to_world_ray(
+    return perspective_unproject_topleft_normalized_pos_to_world(
         pos / dimensions,
         this->position(),
         view_matrix(),
@@ -1007,32 +1007,32 @@ EulerAngles osc::extract_eulers_xyz(const Quat& quaternion)
     return extract_eulers_xyz(mat4_cast(quaternion));
 }
 
-Vec2 osc::topleft_relative_pos_to_ndc_point(Vec2 relative_pos)
+Vec2 osc::topleft_normalized_point_to_ndc(Vec2 normalized_point)
 {
-    relative_pos.y = 1.0f - relative_pos.y;
-    return 2.0f*relative_pos - 1.0f;
+    normalized_point.y = 1.0f - normalized_point.y;
+    return 2.0f*normalized_point - 1.0f;
 }
 
-Vec2 osc::ndc_point_to_topleft_relative_pos(Vec2 ndc_pos)
+Vec2 osc::ndc_point_to_topleft_normalized(Vec2 ndc_point)
 {
-    ndc_pos = (ndc_pos + 1.0f) * 0.5f;
-    ndc_pos.y = 1.0f - ndc_pos.y;
-    return ndc_pos;
+    ndc_point = (ndc_point + 1.0f) * 0.5f;
+    ndc_point.y = 1.0f - ndc_point.y;
+    return ndc_point;
 }
 
-Vec4 osc::topleft_relative_pos_to_ndc_cube(Vec2 relative_pos)
+Vec4 osc::topleft_normalized_point_to_ndc_cube(Vec2 normalized_point)
 {
-    return {topleft_relative_pos_to_ndc_point(relative_pos), -1.0f, 1.0f};
+    return {topleft_normalized_point_to_ndc(normalized_point), -1.0f, 1.0f};
 }
 
-Line osc::perspective_unproject_topleft_screen_pos_to_world_ray(
-    Vec2 relative_pos,
+Line osc::perspective_unproject_topleft_normalized_pos_to_world(
+    Vec2 normalized_point,
     Vec3 camera_worldspace_origin,
     const Mat4& camera_view_matrix,
     const Mat4& camera_proj_matrix)
 {
     // position of point, as if it were on the front of the 3D NDC cube
-    const Vec4 line_origin_ndc = topleft_relative_pos_to_ndc_cube(relative_pos);
+    const Vec4 line_origin_ndc = topleft_normalized_point_to_ndc_cube(normalized_point);
 
     Vec4 line_origin_view = inverse(camera_proj_matrix) * line_origin_ndc;
     line_origin_view /= line_origin_view.w;  // perspective divide
[…]
