### 2025-06-13 f6c8be5
Unify coordinate space naming (separate words)
