### 2025-06-13 8566bef
Ensure ui:: methods clearly document/express their coordinate system(s).
