+++ liboscar-demos/CustomWidgetsTab.cpp
@@ -68,8 +68,8 @@ namespace
 
         ui::pop_id();
         ui::pop_style_var();
-        const Vec2 button_top_left = ui::get_item_topleft();
-        const Vec2 button_bottom_right = ui::get_item_bottomright();
+        const Vec2 button_top_left = ui::get_item_top_left_ui_pos();
+        const Vec2 button_bottom_right = ui::get_item_bottom_right_ui_pos();
 
         draw_widget_title(label, p);
 
