+++ liboscar/UI/oscimgui.cpp
@@ -2139,27 +2139,27 @@ Vec2 osc::ui::get_content_region_available()
     return ImGui::GetContentRegionAvail();
 }
 
-Vec2 osc::ui::get_cursor_start_pos()
+Vec2 osc::ui::get_cursor_start_panel_pos()
 {
     return ImGui::GetCursorStartPos();
 }
 
-Vec2 osc::ui::get_cursor_pos()
+Vec2 osc::ui::get_cursor_panel_pos()
 {
     return ImGui::GetCursorPos();
 }
 
-void osc::ui::set_cursor_pos(Vec2 pos)
+void osc::ui::set_cursor_panel_pos(Vec2 pos)
 {
     ImGui::SetCursorPos(pos);
 }
 
-float osc::ui::get_cursor_pos_x()
+float osc::ui::get_cursor_panel_pos_x()
 {
     return ImGui::GetCursorPosX();
 }
 
-void osc::ui::set_cursor_pos_x(float local_x)
+void osc::ui::set_cursor_panel_pos_x(float local_x)
 {
     ImGui::SetCursorPosX(local_x);
 }
@@ -2174,7 +2174,7 @@ void osc::ui::set_cursor_ui_pos(Vec2 pos)
     ImGui::SetCursorScreenPos(pos);
 }
 
-void osc::ui::set_next_panel_pos(Vec2 pos, Conditional conditional, Vec2 pivot)
+void osc::ui::set_next_panel_ui_pos(Vec2 pos, Conditional conditional, Vec2 pivot)
 {
     ImGui::SetNextWindowPos(pos, to<ImGuiCond>(conditional), pivot);
 }
@@ -2349,7 +2349,7 @@ bool osc::ui::wants_keyboard()
     return ImGui::GetIO().WantCaptureKeyboard;
 }
 
-void osc::ui::push_style_var(StyleVar var, const Vec2& pos)
+void osc::ui::push_style_var(StyleVar var, Vec2 pos)
 {
     ImGui::PushStyleVar(to<ImGuiStyleVar>(var), pos);
 }
@@ -2449,12 +2449,12 @@ bool osc::ui::is_item_deactivated_after_edit()
     return ImGui::IsItemDeactivatedAfterEdit();
 }
 
-Vec2 osc::ui::get_item_topleft()
+Vec2 osc::ui::get_item_top_left_ui_pos()
 {
     return ImGui::GetItemRectMin();
 }
 
-Vec2 osc::ui::get_item_bottomright()
+Vec2 osc::ui::get_item_bottom_right_ui_pos()
 {
     return ImGui::GetItemRectMax();
 }
@@ -2562,44 +2562,44 @@ Vec2 osc::ui::get_panel_size()
     return ImGui::GetWindowSize();
 }
 
-void osc::ui::DrawListAPI::add_rect(const Rect& rect, const Color& color, float rounding, float thickness)
+void osc::ui::DrawListAPI::add_rect(const Rect& ui_rect, const Color& color, float rounding, float thickness)
 {
-    impl_get_drawlist().AddRect(rect.p1, rect.p2, to_ImU32(color), rounding, 0, thickness);
+    impl_get_drawlist().AddRect(ui_rect.p1, ui_rect.p2, to_ImU32(color), rounding, 0, thickness);
 }
 
-void osc::ui::DrawListAPI::add_rect_filled(const Rect& rect, const Color& color, float rounding)
+void osc::ui::DrawListAPI::add_rect_filled(const Rect& ui_rect, const Color& color, float rounding)
 {
-    impl_get_drawlist().AddRectFilled(rect.p1, rect.p2, to_ImU32(color), rounding);
[…]
