### 2025-06-13 92e032b
Clean up some naming (subspace, rect type for screenshots)
