+++ liboscar/Graphics/Camera.h
@@ -113,9 +113,9 @@ namespace osc
         //
         // the usefulness of scissor testing is that it can be used to:
         //
-        // - limit running an expensive fragment shader to a smaller subspace
+        // - limit running an expensive fragment shader to a smaller region
         // - only draw sub-parts of a scene without having to recompute transforms or render it differently etc.
-        // - only clear + draw to a smaller subspace of the output (e.g. writing to subsections of a larger UI)
+        // - only clear + draw to a smaller region of the output (e.g. writing to subsections of a larger UI)
         std::optional<Rect> scissor_rect() const;
         void set_scissor_rect(std::optional<Rect>);
 
