### 2025-06-14 #8284 @Kevin-Mattheus-Moerman
@modenaxe yes, you are listed as a reviewer. You should be able to call `@editorialbot generate my checklist` to get started.
