### 2025-06-16 4f850fa
Make screenshots use screen space, rather than ui space, and clean up associated functions
