+++ liboscar/Platform/ScreenshotAnnotation.h
@@ -9,16 +9,18 @@ namespace osc
 {
     class ScreenshotAnnotation final {
     public:
-        ScreenshotAnnotation(std::string label, Rect rect) :
+        explicit ScreenshotAnnotation(
+            std::string label,
+            const Rect& screenspace_rect) :
             label_{std::move(label)},
-            rect_{rect}
+            rect_{screenspace_rect}
         {}
 
         const std::string& label() const { return label_; }
 
-        // Returns the bounding rectangle of the annotation in device-independent pixels.
+        // Returns the bounding rectangle of the annotation in screenspace in
+        // device-independent pixels.
         const Rect& rect() const { return rect_; }
-
     private:
         std::string label_;
         Rect rect_;
