### 2025-06-16 76cafe9
Make App::set_main_window_unicode_input_rect use screen space
