### 2025-06-16 7039a17
Define mouse position getter/event in screen space (previously: ui space)
