+++ liboscar/Platform/Events/WindowEventType.h
@@ -2,6 +2,7 @@
 
 namespace osc
 {
+    // Represents the type of window change associated with a `WindowEvent`.
     enum class WindowEventType {
         GainedMouseFocus,
         LostMouseFocus,
