+++ liboscar/Platform/Events/EventType.h
@@ -2,6 +2,8 @@
 
 namespace osc
 {
+    // Represents an enumeration of known-at-compile-time event types
+    // that can be cheaply queried (compared to downcasting an `Event`).
     enum class EventType {
         DropFile,
         KeyDown,
