### 2025-06-16 7ddeb7b
Make graphics::blit_to_main_window destination region optional.
