### 2025-06-16 643185d
Clean up some ui workspace query uses.
