+++ liboscar-demos/learnopengl/Guest/LOGLCSMTab.cpp
@@ -289,7 +289,7 @@ private:
         for (size_t i = 1; i < c_normalized_cascade_planes.size(); ++i) {
             const auto [near, far] = user_camera_.clipping_planes();
             const Vec4 view_pos = {0.0f, 0.0f, -lerp(near, far, c_normalized_cascade_planes[i]), 1.0f};
-            const Mat4 proj = user_camera_.projection_matrix(aspect_ratio_of(ui::get_main_window_workspace_screen_space_rect()));
+            const Mat4 proj = user_camera_.projection_matrix(ui::get_main_window_workspace_aspect_ratio());
             const Vec4 proj_pos = (proj * view_pos);
             ends.push_back(proj_pos.z);
         }
