### 2025-06-16 0d7bb0e
Make SVG icons and splash screen images respect device_pixel_ratio
