+++ libopensimcreator/UI/Shared/ModelViewerPanel.cpp
@@ -207,7 +207,8 @@ namespace
 
         std::shared_ptr<IconCache> m_IconCache = App::singleton<IconCache>(
             App::resource_loader().with_prefix("OpenSimCreator/icons/"),
-            ui::get_font_base_size()/128.0f
+            ui::get_font_base_size()/128.0f,
+            App::get().highest_device_pixel_ratio()
         );
         std::string panel_name_;
         ModelSelectionGizmo m_Gizmo;
