+++ liboscar/UI/IconCache.cpp
@@ -20,9 +20,11 @@ class osc::IconCache::Impl final {
 public:
     explicit Impl(
         ResourceLoader& loader_prefixed_at_dir_containing_svgs,
-        float vertical_scale)
+        float vertical_scale,
+        float device_pixel_ratio)
     {
         OSC_ASSERT(vertical_scale > 0.0f && "icon cache's vertical scale must be a positive number");
+        OSC_ASSERT(device_pixel_ratio > 0.0f && "icon cache's device pixel ratio must be a positive number");
 
         const auto it = loader_prefixed_at_dir_containing_svgs.iterate_directory(".");
 
@@ -32,9 +34,9 @@ public:
             if (p.has_extension(".svg")) {
                 Texture2D texture = load_texture2D_from_svg(
                     loader_prefixed_at_dir_containing_svgs.open(p),
-                    vertical_scale
+                    vertical_scale,
+                    device_pixel_ratio
                 );
-                texture.set_filter_mode(TextureFilterMode::Nearest);
 
                 icons_by_name_.try_emplace(
                     p.stem(),
@@ -62,8 +64,8 @@ private:
 };
 
 
-osc::IconCache::IconCache(ResourceLoader loader_prefixed_at_dir_containing_svgs, float vertical_scale) :
-    impl_{std::make_unique<Impl>(loader_prefixed_at_dir_containing_svgs, vertical_scale)}
+osc::IconCache::IconCache(ResourceLoader loader_prefixed_at_dir_containing_svgs, float vertical_scale, float device_pixel_ratio) :
+    impl_{std::make_unique<Impl>(loader_prefixed_at_dir_containing_svgs, vertical_scale, device_pixel_ratio)}
 {}
 osc::IconCache::IconCache(IconCache&&) noexcept = default;
 osc::IconCache& osc::IconCache::operator=(IconCache&&) noexcept = default;
