### 2025-06-17 3c91e3f
Fix cubemap learnopengl demo addressing
