+++ liboscar/Maths/AABBFunctions.h
@@ -164,7 +164,7 @@ namespace osc
         return rv;
     }
 
-    // returns a `Rect` in normalized device coordinate (NDC) space that loosely
+    // returns a `Rect` in normalized device coordinate-like (NDC-like) space that loosely
     // bounds the world-space-located `aabb`
     //
     // if the AABB does not lie within the NDC clipping bounds (i.e. between (-1, -1)
