### 2025-06-17 07cc301
Rename Texture2D dimensions -> pixel_dimensions to match other APIs
