### 2025-06-17 1854b4e
Refactor RenderTexture and related APIs to use pixel_dimensions.
