+++ liboscar/Graphics/ColorRenderBufferParams.h
@@ -10,7 +10,7 @@ namespace osc
     struct ColorRenderBufferParams final {
         friend bool operator==(const ColorRenderBufferParams&, const ColorRenderBufferParams&) = default;
 
-        Vec2i dimensions = {1, 1};
+        Vec2i pixel_dimensions = {1, 1};
         TextureDimensionality dimensionality = TextureDimensionality::Tex2D;
         AntiAliasingLevel anti_aliasing_level = AntiAliasingLevel{1};
         ColorRenderBufferFormat format = ColorRenderBufferFormat::Default;
