### 2025-06-17 c63800d
Ensure oscimgui uses the render target's device pixel ratio.
