### 2025-06-17 09f4d21
Clarify more dimensions vs. pixel_dimensions uses.
