### 2025-06-17 6544fba
Rename virtual_pixel_dimensions -> dimensions for consistency with other APIs.
