+++ libopensimcreator/UI/MeshImporter/MeshImporterSharedState.h
@@ -433,7 +433,7 @@ namespace osc::mi
 
             // setup rendering params
             SceneRendererParams p;
-            p.virtual_pixel_dimensions = dimensions_of(get3DSceneRect());
+            p.dimensions = dimensions_of(get3DSceneRect());
             p.device_pixel_ratio = app.settings().get_value<float>("graphics/render_scale", 1.0f) * app.main_window_device_pixel_ratio();
             p.antialiasing_level = app.anti_aliasing_level();
             p.draw_rims = true;
@@ -441,7 +441,7 @@ namespace osc::mi
             p.near_clipping_plane = m_3DSceneCamera.znear;
             p.far_clipping_plane = m_3DSceneCamera.zfar;
             p.view_matrix = m_3DSceneCamera.view_matrix();
-            p.projection_matrix = m_3DSceneCamera.projection_matrix(aspect_ratio_of(p.virtual_pixel_dimensions));
+            p.projection_matrix = m_3DSceneCamera.projection_matrix(aspect_ratio_of(p.dimensions));
             p.view_pos = m_3DSceneCamera.position();
             p.light_direction = recommended_light_direction(m_3DSceneCamera);
             p.light_color = Color::white();
