+++ libopensimcreator/UI/SplashTab.cpp
@@ -195,7 +195,7 @@ private:
         ui::pop_style_var();
 
         SceneRendererParams params{m_LastSceneRendererParams};
-        params.virtual_pixel_dimensions = dimensions_of(workspaceUIRect);
+        params.dimensions = dimensions_of(workspaceUIRect);
         params.device_pixel_ratio = App::settings().get_value<float>("graphics/render_scale", 1.0f) * App::get().main_window_device_pixel_ratio(),
         params.antialiasing_level = App::get().anti_aliasing_level();
         params.projection_matrix = m_Camera.projection_matrix(aspect_ratio_of(workspaceUIRect));
