+++ liboscar/Graphics/Scene/SceneHelpers.h
@@ -136,8 +136,8 @@ namespace osc
     SceneRendererParams calc_standard_dark_scene_render_params(
         const PolarPerspectiveCamera&,
         AntiAliasingLevel,
-        Vec2 render_virtual_pixel_dimensions,
-        float render_device_pixel_ratio
+        Vec2 dimensions,
+        float device_pixel_ratio
     );
 
     // returns a triangle BVH for the given triangle mesh, or an empty BVH if the mesh is non-triangular or empty
