### 2025-06-17 b0b452c
Fix test suite for latest refactors
