+++ liboscar/Graphics/ColorRenderBufferParams.tests.cpp
@@ -13,10 +13,10 @@ TEST(ColorRenderBufferParams, can_default_construct)
     [[maybe_unused]] const ColorRenderBufferParams default_constructed;
 }
 
-TEST(ColorRenderBufferParams, default_constructed_has_1x1_dimensions)
+TEST(ColorRenderBufferParams, default_constructed_has_1x1_pixel_dimensions)
 {
     const ColorRenderBufferParams default_constructed;
-    ASSERT_EQ(default_constructed.dimensions, Vec2i(1, 1));
+    ASSERT_EQ(default_constructed.pixel_dimensions, Vec2i(1, 1));
 }
 
 TEST(ColorRenderBufferParams, default_constructed_has_1x_AntiAliasingLevel)
@@ -32,20 +32,20 @@ TEST(ColorRenderBufferParams, can_compare_for_equality)
     ASSERT_EQ(lhs, rhs) << "i.e. it's value equality";
 }
 
-TEST(ColorRenderBufferParams, changing_dimensions_changes_equality)
+TEST(ColorRenderBufferParams, changing_pixel_dimensions_changes_equality)
 {
     const ColorRenderBufferParams lhs;
-    const ColorRenderBufferParams rhs{.dimensions = {2, 2}};
+    const ColorRenderBufferParams rhs{.pixel_dimensions = {2, 2}};
     ASSERT_NE(lhs, rhs);
 }
 
 TEST(ColorRenderBufferParams, can_be_initialized_with_designated_initializer)
 {
     const ColorRenderBufferParams params = {
-        .dimensions = {2, 2},
+        .pixel_dimensions = {2, 2},
         .format = ColorRenderBufferFormat::R8G8B8A8_UNORM,
     };
 
-    ASSERT_EQ(params.dimensions, Vec2i(2, 2));
+    ASSERT_EQ(params.pixel_dimensions, Vec2i(2, 2));
     ASSERT_EQ(params.format, ColorRenderBufferFormat::R8G8B8A8_UNORM);
 }
