+++ liboscar/Graphics/RenderTexture.tests.cpp
@@ -14,7 +14,7 @@ using namespace osc;
 TEST(RenderTexture, default_constructor_creates_1x1_default_texture)
 {
     const RenderTexture render_texture;
-    ASSERT_EQ(render_texture.dimensions(), Vec2i(1, 1));
+    ASSERT_EQ(render_texture.dimensions(), Vec2(1.0f, 1.0f));
     ASSERT_EQ(render_texture.depth_stencil_format(), DepthStencilRenderBufferFormat::Default);
     ASSERT_EQ(render_texture.color_format(), ColorRenderBufferFormat::Default);
     ASSERT_EQ(render_texture.anti_aliasing_level(), AntiAliasingLevel{1});
@@ -81,11 +81,11 @@ TEST(RenderTexture, reformat_throws_if_given_CubeDimensionality_and_anti_aliasin
     ASSERT_ANY_THROW(RenderTexture().reformat(render_texture_params));
 }
 
-TEST(RenderTexture, throws_if_given_non_square_dimensions_but_Cube_dimensionality)
+TEST(RenderTexture, throws_if_given_non_square_pixel_dimensions_but_Cube_dimensionality)
 {
     // permitted
     const RenderTextureParams render_texture_params = {
-        .dimensions = {1, 2},
+        .pixel_dimensions = {1, 2},
         .dimensionality = TextureDimensionality::Cube,
     };
 
@@ -93,20 +93,20 @@ TEST(RenderTexture, throws_if_given_non_square_dimensions_but_Cube_dimensionalit
     ASSERT_ANY_THROW(const RenderTexture render_texture(render_texture_params));
 }
 
-TEST(RenderTexture, set_dimensionality_throws_if_set_on_RenderTexture_with_non_square_dimensions)
+TEST(RenderTexture, set_dimensionality_throws_if_set_on_RenderTexture_with_non_square_pixel_dimensions)
 {
     RenderTexture render_texture;
-    render_texture.set_dimensions({1, 2});  // not square
+    render_texture.set_pixel_dimensions({1, 2});  // not square
 
     ASSERT_ANY_THROW(render_texture.set_dimensionality(TextureDimensionality::Cube));
 }
 
-TEST(RenderTexture, set_dimensions_throws_if_set_on_RenderTexture_with_cube_dimensionality)
+TEST(RenderTexture, set_pixel_dimensions_throws_if_set_on_RenderTexture_with_cube_dimensionality)
 {
     RenderTexture render_texture;
     render_texture.set_dimensionality(TextureDimensionality::Cube);
 
-    ASSERT_ANY_THROW(render_texture.set_dimensions({1, 2}));
+    ASSERT_ANY_THROW(render_texture.set_pixel_dimensions({1, 2}));
 }
 
 TEST(RenderTexture, set_dimension_changes_equality)
@@ -121,11 +121,11 @@ TEST(RenderTexture, set_dimension_changes_equality)
     ASSERT_NE(texture_a, texture_b);
 }
 
-TEST(RenderTexture, can_be_constructed_from_dimensions_vector)
+TEST(RenderTexture, can_be_constructed_from_pixel_dimensions_vector)
 {
-    const Vec2i dimensions = {12, 12};
-    const RenderTexture render_texture{{.dimensions = dimensions}};
-    ASSERT_EQ(render_texture.dimensions(), dimensions);
+    const Vec2i pixel_dimensions = {12, 12};
+    const RenderTexture render_texture{{.pixel_dimensions = pixel_dimensions}};
+    ASSERT_EQ(render_texture.pixel_dimensions(), pixel_dimensions);
 }
 
[…]
