### 2025-06-17 2cf2240
Remove unused header in LatheGeometry.h
