### 2025-06-17 1ec9d9c
Port open_{file,url}* os functions to SDL3
