### 2025-06-18 7e598bf
Drop backtrace handling code from os.h
