+++ liboscar/Platform/os.cpp
@@ -1,9 +1,6 @@
 #include "os.h"
 
 #include <liboscar/Platform/Log.h>
-#include <liboscar/Platform/LogLevel.h>
-#include <liboscar/Platform/LogSink.h>
-#include <liboscar/Utils/Assertions.h>
 #include <liboscar/Utils/ScopeExit.h>
 #include <liboscar/Utils/StringHelpers.h>
 #include <liboscar/Utils/SynchronizedValue.h>
@@ -16,16 +13,11 @@
 
 #include <algorithm>
 #include <array>
-#include <bit>
 #include <cerrno>
 #include <cstddef>
-#include <ctime>
 #include <fstream>
-#include <iostream>
 #include <memory>
 #include <random>
-#include <ranges>
-#include <span>
 #include <sstream>
 #include <stdexcept>
 #include <string>
@@ -36,10 +28,6 @@ namespace rgs = std::ranges;
 
 namespace
 {
-    // care: this is necessary because segfault crash handlers don't appear to
-    // be able to have data passed to them
-    constinit SynchronizedValue<std::optional<std::filesystem::path>> g_crash_dump_dir;  // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
-
     std::filesystem::path convert_SDL_filepath_to_std_filepath(CStringView method_name, const char* p)
     {
         // nullptr disallowed
@@ -188,7 +176,7 @@ std::pair<std::fstream, std::filesystem::path> osc::mkstemp(std::string_view suf
         // TODO: remove these `pragma`s once the codebase is upgraded to C++23, because it has `std::ios_base::noreplace` support
 #pragma warning(push)
 #pragma warning(suppress : 4996)
-        if (auto fd = std::fopen(attempt_path.string().c_str(), "w+x"); fd != nullptr) {  // TODO: replace with `std::ios_base::noreplace` in C++23
+        if (auto fd = std::fopen(attempt_path.string().c_str(), "w+x"); fd != nullptr) {  // TODO: replace with `std::ios_base::noreplace` in C++23 (waiting on XCode16)
             std::fclose(fd);
             return {std::fstream{attempt_path, std::ios_base::in | std::ios_base::out | std::ios_base::binary}, std::move(attempt_path)};
         }
@@ -200,24 +188,15 @@ std::pair<std::fstream, std::filesystem::path> osc::mkstemp(std::string_view suf
 
 #ifdef __LINUX__
 
-#ifndef _GNU_SOURCE
-#define _GNU_SOURCE
-#endif
-#ifndef __USE_GNU
-#define __USE_GNU
-#endif
+#include <liboscar/Platform/Log.h>  // osc::log_warn
 
-#include <csignal>
-#include <cstdio>
-#include <cstdlib>
+#include <ctime>                    // std::time_t, std::tm
+#include <array>                    // std::array
+#include <string>                   // std::string
+#include <type_traits>              // std::is_same_v
 
-#include <execinfo.h>
-#include <ucontext.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-
-using osc::log_error;
+#include <string.h>                 // strerror_r
+#include <time.h>                   // gmtime_r
 
 namespace
 {
@@ -232,135 +211,30 @@ namespace
     {
         std::array<char, 1024> buffer{};
 
-        auto* maybeErr = strerror_r(errnum, buffer.data(), buffer.size());
-        if (std::is_same_v<int, decltype(maybeErr)> && !maybeErr)
-        {
-            log_warn("a call to strerror_r failed with error code %i", maybeErr);
+        [[maybe_unused]] auto* maybeErr = strerror_r(errnum, buffer.data(), buffer.size());
[…]
