### 2025-06-18 47f3935
Replace errno_to_string_threadsafe with ostream exceptions (portable)
