### 2025-06-18 c6744c6
Refactor calendar time calculation to use SDL, remove remaining OS-dependent code in os.cpp
