### 2025-06-18 2d0b205
Update continuous-integration-workflow.yml
Fix arch typo in ubuntu GitHub Actions artifact upload
