### 2025-06-19 0bf636a
#ifdef guard std::tm fields that aren't available in Windows
