### 2025-06-19 c6f659c
Remove environment variable manipulators from os (available in std::)
