+++ liboscar/Platform/os.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <ctime>
 #include <filesystem>
 #include <fstream>
 #include <string_view>
@@ -10,9 +9,6 @@
 // os: where all the icky OS/distro/filesystem-specific stuff is hidden
 namespace osc
 {
-    // returns local system time as a calendar date and time broken down into its components.
-    std::tm system_calendar_time();
-
     // returns the full path to the currently-executing application
     //
     // care: can be slow: downstream callers should cache it
