+++ liboscar/Formats/STL.cpp
@@ -8,14 +8,15 @@
 #include <liboscar/Platform/os.h>
 #include <liboscar/Utils/Algorithms.h>
 #include <liboscar/Utils/Assertions.h>
+#include <liboscar/Utils/ChronoHelpers.h>
 #include <liboscar/Utils/ObjectRepresentation.h>
 #include <liboscar/Strings.h>
 
 #include <algorithm>
+#include <chrono>
 #include <cstddef>
 #include <cstdint>
 #include <functional>
-#include <iomanip>
 #include <ostream>
 #include <limits>
 #include <span>
@@ -30,7 +31,7 @@ namespace
     std::string calc_header_text(const StlMetadata& metadata)
     {
         std::stringstream ss;
-        ss << "created " << std::put_time(&metadata.creation_time, "%Y-%m-%d %H:%M:%S") << " by " << metadata.authoring_tool;
+        ss << "created " << to_iso8601_timestamp(metadata.creation_time) << " by " << metadata.authoring_tool;
         return std::move(ss).str();
     }
 
@@ -108,7 +109,7 @@ osc::StlMetadata::StlMetadata(
     std::string_view authoring_tool_) :
 
     authoring_tool{authoring_tool_},
-    creation_time{system_calendar_time()}
+    creation_time{std::chrono::current_zone(), std::chrono::time_point_cast<std::chrono::seconds>(std::chrono::system_clock::now())}
 {}
 
 void osc::write_as_stl(
