### 2025-06-19 99033bf
Replace std::getenv usage with SDL_getenv_unsafe
