### 2025-06-19 861bab7
Fix bad destination_screen_rect forwarding
