+++ liboscar/Utils/PerfClock.h
@@ -4,5 +4,7 @@
 
 namespace osc
 {
-    using PerfClock = std::chrono::high_resolution_clock;
+    // A `std::chrono`-compatible clock that is steady (i.e. always monotonically
+    // increases) and has a high-resolution enough for performance measurements.
+    using PerfClock = std::chrono::steady_clock;
 }

[…]
