+++ liboscar/Graphics/Mesh.tests.cpp
@@ -29,7 +29,6 @@
 #include <cstdint>
 #include <functional>
 #include <iterator>
-#include <random>
 #include <ranges>
 #include <span>
 #include <sstream>
