+++ liboscar/testing/TestRenderer.cpp
@@ -24,7 +24,6 @@
 #include <array>
 #include <cstdint>
 #include <memory>
-#include <random>
 #include <span>
 #include <sstream>
 #include <string>
