### 2025-06-19 64e4291
Make workflow badge point to workflow rather than icon (openjournals/joss-reviews#8284)
