### 2025-06-19 ComputationalBiomechanicsLab/opensim-creator#1049 @adamkewley
Add "Change/Edit Geometry" contextual action to `OpenSim::Geometry`
Suggested during https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976828623

The implementation code will need to ensure that the geometry can be changed in-place (e.g. that deletion is possible, and so on).
