### 2025-06-19 ComputationalBiomechanicsLab/opensim-creator#1050 @adamkewley
Fix mesh importer crashes when trying to import a non-osim file
Mentioned during https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976828623

The mesh importer shouldn't allow non-osim files, or should at least wrap the loading step in a `try...catch` or similar to ensure that IO errors don't cause the whole process to die.
