### 2025-06-19 ComputationalBiomechanicsLab/opensim-creator#1051 @adamkewley
Add a play button (scrubber) to the Preview Experimental Data workflow
Mentioned during https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976828623

Effectively, port the simulation scrubber to the preview experimental data tab.
