### 2025-06-19 ComputationalBiomechanicsLab/opensim-creator#1052 @adamkewley
Make dialog filters in preview experimental data tab more restrictive (or add a tooltip)
Mentioned during https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976828623

User reported that it's unclear what "raw data" or "model trajectory/state" are in the preview experimental data UI.
