### 2025-06-19 bc03c62
Add "model_editor/monitor_osim_changes" flag to osc.toml (#1000)
