+++ CHANGELOG.md
@@ -5,6 +5,10 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Upcoming Release]
 
+- The `osc.toml` configuration file now supports a `model_editor/monitor_osim_changes`
+  boolean option, which can be used to explicitly tell the OpenSim model editor
+  whether or not to auto-reload the file when it changes on disk (defaults to
+  `true`, #1000).
 - SVG icons and banners now rasterize in high DPI mode when rendering to a
   high DPI monitor.
 - Keyboard navigation between elements in the UI is now easier and supports (e.g.)
