+++ osc/osc_dev_config.toml.in
@@ -12,7 +12,7 @@
 # INSTALL configuration file.
 
 resources = "@CMAKE_CURRENT_SOURCE_DIR@/../resources"
-docs_url = "@CMAKE_CURRENT_SOURCE_DIR@/../docs/build/index.html"
+docs_url = "file://@CMAKE_CURRENT_SOURCE_DIR@/../docs/build/index.html"
 
 # change this if you need OSC to print more information to the log
 # log_level = "trace"  # trace < debug < info < warn < err < critical
@@ -23,3 +23,9 @@ docs_url = "@CMAKE_CURRENT_SOURCE_DIR@/../docs/build/index.html"
 # a scale factor that is multiplied by the number of physical pixels that comprise
 # each 3D element in the UI (e.g. you can use this to downscale HighDPI viewports)
 # graphics.render_scale = 1.0
+
+# [model_editor]
+#
+# # tells the OpenSim model editor interface to enable/disable auto-reloading osim
+# # files if the associated .osim file changes (#1000)
+# monitor_osim_changes = false
