### 2025-06-19 13e8c79
Make lightning icon clickable in simulation tab (#777).
