### 2025-06-19 61f40a3
Make Graphics/ tests follow same convention as other tests (#763)
