### 2025-06-19 1088e7b
Make 3D hittest prioritize subcomponents (#592).
