### 2025-06-19 a2ab45b
Clean up collision detection code (#592).
