### 2025-06-19 764548d
Fix collision detection code (whoops)
