### 2025-06-19 f83a9da
Make PathPoint have Marker icon
