### 2025-06-20 ComputationalBiomechanicsLab/opensim-creator#1056 @adamkewley
Add contribution link/section to README.md
Mentioned here https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976767980

There should be a little section in the README that mentions/guides the user to the contribution guide. It can still potentially be in the documentation but it should probably be documented better.

> It would be nice to have a small section about contribution rules in the Readme, or at least a link to the appropriate section of the documentation: potential contributors are likely to primarily search for this information on the code base rather than on a website.
