### 2025-06-20 ComputationalBiomechanicsLab/opensim-creator#1057 @adamkewley
Add tooltips and filter input files in the preview experimental data workflow
Mentioned here: https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-2976828623

> In the same UI, it is not necessary clear what "raw data" and "model trajectory/state" files mean, so it would be great to filter trc/mot/sto/osim files accordingly
