+++ libopensimcreator/UI/Shared/ImportStationsFromCSVPopup.cpp
@@ -9,6 +9,7 @@
 #include <liboscar/Graphics/Color.h>
 #include <liboscar/Formats/CSV.h>
 #include <liboscar/Platform/App.h>
+#include <liboscar/Platform/FileDialogFilter.h>
 #include <liboscar/Platform/os.h>
 #include <liboscar/UI/Popups/Popup.h>
 #include <liboscar/UI/Popups/PopupPrivate.h>
