+++ libopensimcreator/Graphics/SimTKMeshLoader.cpp
@@ -3,14 +3,12 @@
 #include <libopensimcreator/Utils/SimTKConverters.h>
 
 #include <liboscar/Graphics/Mesh.h>
-#include <liboscar/Graphics/MeshTopology.h>
-#include <liboscar/Graphics/VertexAttribute.h>
-#include <liboscar/Graphics/VertexAttributeFormat.h>
 #include <liboscar/Graphics/VertexFormat.h>
 #include <liboscar/Maths/MathHelpers.h>
 #include <liboscar/Maths/Triangle.h>
 #include <liboscar/Maths/TriangleFunctions.h>
 #include <liboscar/Maths/Vec3.h>
+#include <liboscar/Platform/FileDialogFilter.h>
 #include <liboscar/Utils/Assertions.h>
 #include <SimTKcommon/internal/DecorativeGeometry.h>
 #include <SimTKcommon/internal/PolygonalMesh.h>
