### 2025-06-20 fab7b65
Fix mesh importer import malformed osim crash (thanks @davidpagnon, #1050).
