### 2025-06-20 2264cc8
Make documentation buttons etc. open docs.opensimcreator.com (#1048).
