+++ libopensimcreator/Platform/OpenSimCreatorApp.cpp
@@ -179,13 +179,3 @@ TabRegistry& osc::OpenSimCreatorApp::upd_tab_registry()
 {
     return *singleton<TabRegistry>();
 }
-
-std::string osc::OpenSimCreatorApp::docs_url() const
-{
-    if (const auto runtime_url = settings().find_value("docs_url")) {
-        return to<std::string>(*runtime_url);
-    }
-    else {
-        return "https://docs.opensimcreator.com";
-    }
-}
