+++ osc/osc_config.h.in
@@ -28,10 +28,15 @@
 
 // char[]
 //
-// the internet location of the OSC repo. Used by in-app links, shortcuts, etc.
-// to the repo so that users can just click and go, rather than having to separately
-// browse to the repo
+// a URL for OSC's source code repo. Used by in-app links, shortcuts, etc. so that
+// users can just click and go, rather than having to separately browse to it.
 #define OSC_REPO_URL "@OSC_REPO_URL@"
+
+// char[]
+//
+// a URL for the OSC's documentation Used by in-app links, shortcuts, etc. so that
+// users can just click and go, rather than having to separately browse to it.
+#define OSC_DOCS_URL "@OSC_DOCS_URL@"
 
 // char[]
 //
