### 2025-06-20 7fea4f1
Refactor UI context to ensure all UI state has correct lifetime (#1037).
