+++ liboscar/UI/oscimgui.cpp
@@ -93,14 +93,12 @@
 #include <variant>
 #include <vector>
 
-#include "liboscar/Platform/Log.h"
-#include "liboscar/Utils/StringHelpers.h"
-
 namespace plot = osc::ui::plot;
 namespace rgs = std::ranges;
 using namespace osc::literals;
 using namespace osc;
 
+// A converter that converts `ImGuiMouseCursor` to oscar's `CursorShape`.
 template<>
 struct osc::Converter<ImGuiMouseCursor, CursorShape> final {
     CursorShape operator()(ImGuiMouseCursor cursor) const
@@ -125,6 +123,7 @@ struct osc::Converter<ImGuiMouseCursor, CursorShape> final {
     }
 };
 
+// A converter that converts oscar's `Key` enum to an `ImGuiKey`.
 template<>
 struct osc::Converter<Key, ImGuiKey> final {
     ImGuiKey operator()(Key key) const
@@ -260,8 +259,13 @@ static_assert(osc::ui::gizmo_annotation_offset() == ImGuizmo::AnnotationOffset()
 
 namespace
 {
+    // The default size of the font in the UI
     constexpr float c_default_base_font_device_independent_pixel_size = 15.0f;
 
+    // The default threshold, in device-independent pixels that a user has to drag their mouse to trigger a drag
+    constexpr float c_default_drag_threshold = 5.0f;
+
+    // Vertex shader that's used to rasterize the UI.
     constexpr std::string_view c_ui_vertex_shader_src = R"(
         #version 330 core
 
@@ -282,6 +286,7 @@ namespace
         }
     )";
 
+    // Fragment shader that's used to rasterize the UI.
     constexpr std::string_view c_ui_fragment_shader_src = R"(
         #version 330 core
 
@@ -297,6 +302,87 @@ namespace
             Out_Color = Frag_Color * texture(uTexture, Frag_UV.st);
         }
     )";
+
+    // Returns a lookup table that maps sRGB color bytes to linear-space color bytes
+    std::array<uint8_t, 256> create_srgb_to_linear_lut()
+    {
+        std::array<uint8_t, 256> rv{};
+        for (size_t i = 0; i < 256; ++i) {
+            const auto ldr_color = Unorm8{static_cast<uint8_t>(i)};
+            const float hdr_color = ldr_color.normalized_value();
+            const float linear_hdr_color = to_linear_colorspace(hdr_color);
+            rv[i] = Unorm8{linear_hdr_color}.raw_value();
+        }
+        return rv;
+    }
+
+    // The internal backend data associated with one UI context.
+    //
+    // Must be constructed BEFORE the ImGui context is initialized and
+    // destroyed AFTER the ImGui context is shutdown, because ImGui has
+    // a few fields/structs that specify things like "this pointer must
+    // outlive the ImGui context".
+    struct OscarUIBackendData final {
+
+        explicit OscarUIBackendData(
+            CopyOnUpdPtr<ui::ContextConfiguration::Impl> config,
+            WindowID window_id) :
+
+            CallerConfig{std::move(config)},
+            Window{window_id}
+        {
+            ui_material.set_transparent(true);
+            ui_material.set_cull_mode(CullMode::Off);
+            ui_material.set_depth_tested(false);
+            ui_material.set_wireframe(false);
+        }
+
+        CopyOnUpdPtr<ui::ContextConfiguration::Impl>     CallerConfig;
+
+        WindowID                                         Window;
+        WindowID                                         ImeWindow;  // important: used for UI's textual inputs (e.g. `ImGui::InputText`)
+        std::string                                      ClipboardText;
[…]
