### 2025-06-20 f803648
Fix memory leak in osc::App::highest_device_pixel_ratio
