### 2025-06-20 eb96003
Fix lint warnings found by clang-tidy
