+++ liboscar/Platform/os.cpp
@@ -73,7 +73,7 @@ std::tm osc::system_calendar_time()
     rv.tm_year = dt.year - 1900;
     rv.tm_wday = dt.day_of_week;
     rv.tm_yday = doy;
-    rv.tm_isdst = dt.utc_offset > 0;
+    rv.tm_isdst = dt.utc_offset > 0 ? 1 : 0;
 #ifdef __USE_MISC
     rv.tm_gmtoff = dt.utc_offset;
     // rv.tm_zone  // TODO
