### 2025-06-20 7a26c13
Remove unused namespace declaration
