### 2025-06-20 13d22a4
Fix std::bit_width Ubuntu22 defect regression
