### 2025-06-23 a3166bb
Clean up CHANGELOG.md for 0.5.24 (#1060)
