### 2025-06-23 b9cf803
Remove unnecessary step from release process documentation (#1060)
