+++ docs/source/the-release-process.rst
@@ -20,7 +20,6 @@ Creator, it's usually copied into a GitHub issue:
 
     - [ ] Create an issue called something like `Release XX.xx.pp`
     - [ ] Copy this checklist into it
-    - [ ] Bump OSC's version number in `CMakeLists.txt` (`project`) to the desired version
     - [ ] Use `scripts/build_linux_debugging.sh` to clean-build a debug (+ libASAN)
           version of OSC on Ubuntu 24 (debugging os).
     - [ ] Ensure the `clang-tidy` lints and test suite passes with the debug build
