### 2025-06-23 de950c2
Credit @mrrezaie for reporting #1000
