### 2025-06-23 39b40b2
Bump version to 0.5.25 (#1060)
