### 2025-06-23 fea9e57
Update CITATION.cff, README.md, and codemeta.json for 0.5.24 (#1060)
