### 2025-06-23 a6bdc68
Add little fixups to release process notes (#1060)
