### 2025-06-23 bce99d4
Shorten some documentation page names
