+++ docs/source/make-a-lower-leg.rst
@@ -1,0 +1,611 @@
+Make a Lower Leg
+================
+
+.. warning::
+
+    This tutorial is new ⭐, and uses ``StationDefinedFrame``\s, which require OpenSim >= v4.5.1.
+    The content of this tutorial should be valid long-term, but we are waiting for OpenSim GUI
+    v4.6 to be released before we remove any "experimental" labelling. We also anticipate
+    adding some handy tooling around re-socketing existing joints and defining ``StationDefinedFrame``\s.
+
+In this tutorial, we will be making a model of a lower leg using OpenSim Creator:
+
+.. figure:: _static/make-a-lower-leg/final-model-screenshot.jpeg
+    :width: 60%
+
+    The model created by this tutorial. It contains three bodies, two joints, three muscles,
+    and a wrap cylinder. This covers the basics of using OpenSim Creator's model editor
+    to build a model with biological components (see :doc:`make-a-bouncing-block` for
+    a more mechanical example).
+
+This tutorial will primarily use the model editor workflow to build a model from scratch
+that contains some of the steps/components necessary to build a human model. In
+essence, the content here will be similar to that in :doc:`make-a-bouncing-block`, but
+with a focus on using landmark data, :doc:`station-defined-frames`, and wrap surfaces
+to build a model of a biological system.
+
+
+Prerequisites
+-------------
+
+This tutorial assumes you:
+
+- Have a basic working knowledge of OpenSim, which is covered in :doc:`make-a-pendulum`
+  and :doc:`make-a-bouncing-block`.
+- (*optional*) The modelling process will also include adding a ``StationDefinedFrame`` to
+  the model. The details of how they work is explained in :doc:`station-defined-frames`.
+- (*optional*) The building process uses externally-provided landmarks from CSV files (e.g.
+  ``femur_r.landmarks.csv``). If you would like to know how to manually place landmarks
+  on a mesh, we recommend reading through :doc:`the-mesh-importer`.
+
+
+Topics Covered by this Tutorial
+-------------------------------
+
+* Creating an OpenSim model by adding bodies and joints.
+* Adding ``StationDefinedFrame``\s to the model in order to define anatomically
+  representative joint frames.
+* Adding a muscle to the model.
+* Adding a wrap surface to the model and associating muscles to that surface.
+
+
+.. _make-a-lower-leg-resources-link:
+
+Download Resources
+------------------
+
+In order to follow this tutorial, you will need to download the associated
+resources (:download:`download here <_static/the-model-warper/make-a-lower-leg_resources.zip>`)
+and unzip them on your computer.
+
+
+Create a New Model
+------------------
+
+Create a new model, as described in :doc:`make-a-pendulum` (:ref:`create-new-model`).
+
+
+Add a ``pelvis`` Body
+---------------------
+
+Add a pelvis body. For this model, use the following parameters:
+
+.. figure:: _static/make-a-lower-leg/add-pelvis-body.jpeg
+    :width: 60%
+
+    Create a body called ``pelvis``. The mass and inertia can be handled later.
+    ``pelvis`` should directly (no offset frames) be joined to ``ground`` with
+    a ``FreeJoint`` called ``pelvis_to_ground``. Pelvis meshes are attached in
+    the next step.
+
+Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
+:ref:`create-the-foot`.
+
+
[…]
