+++ docs/source/the-model-warper.rst
@@ -9,23 +9,23 @@ The Model Warper
 
     The model warper has been in development since 2023 and has undergone several
     redesigns from its initial version (an extended form of :ref:`the-mesh-warper`)
-    to what it is now (a way of combining various scaling steps in a linear pipeline).
-    This is because because non-linear "model warping" combines a variety of scaling
-    algorithms in a model-specific way - our goal is to provide clear UI tooling
-    that makes combining those algorithms easier.
+    to what it is now. This is because because non-linear "model warping" combines
+    a variety of scaling algorithms in a model-specific way - our goal is to
+    provide clear UI tooling that makes combining various scaling algorithms into a
+    single pipeline easier.
 
     We invite you to try the model warper out and get a feel for how it might be
     useful. Maybe it's already useful enough for you to use it in something
-    serious (some researchers already have 🎉).
+    serious.
 
 In this tutorial, we will be using the model warper to create a warping pipeline that
-can be used to warp an entire OpenSim model using experimental measurements, such as
-CT scans and weight measurements. The benefit of the model warper is that it lets you
+can be used to warp an entire OpenSim model using subject-specific data, such as CT
+scans and weight measurements. The benefit of the model warper is that it lets you
 combine various, potentially non-uniform, scaling steps into a single warping pipeline
 that's standard, introspectible, and reusable.
 
 .. _model-warper-ui:
-.. figure:: _static/the-model-warper/model-warper.jpg
+.. figure:: _static/the-model-warper/model-warper.jpeg
     :width: 60%
 
     The model warping UI. This tutorial goes through top-level model warping concepts
@@ -48,16 +48,21 @@ Prerequisites
   pairing landmarks between two corresponding meshes as inputs for the TPS technique.
 
 * **Familiarity with StationDefinedFrames**. The walkthrough in this tutorial uses 
-  ``StationDefinedFrame``\s so that non-linear TPS scaling steps correctly recompute
-  the source model's joint frames. The :doc:`station-defined-frames` documentation
-  outlines what ``StationDefinedFrame``\s are and how to add them to models.
+  a model that contains ``StationDefinedFrame``\s so that non-linear TPS scaling steps
+  correctly recompute the model's joint frames. The :doc:`station-defined-frames`
+  documentation outlines what ``StationDefinedFrame``\s are and how to add them to
+  models.
+
+* **Familiarity with the previous tutorial**: The model created in :doc:`make-a-lower-leg`
+  is the source model in this one, so going through it will provide valuable context about
+  the source model's design.
 
 
 Topics Covered by this Tutorial
 -------------------------------
 
 * A technical overview of how the model warper works
-* A concrete walkthrough of warping a simple model
+* A concrete walkthrough of warping the model from :doc:`make-a-lower-leg`
 * An explanation of how model warping behavior can be customized
 
 
@@ -75,11 +80,11 @@ OpenSim Creator provides a workflow for designing and executing a model warping
 procedure, pictured in :numref:`model-warper-overview-screenshot`. The workflow
 UI is designed to provide visual feedback about each scaling step, so that you
 can incrementally build a warping procedure one scaling step at a time. The model
-warping procedure can then be saved to a standard XML file so that it can be reused
-and modified.
+warping procedure (scaling document) can then be saved to a standard XML file so
+that it can be reused and modified.
 
 .. _model-warper-overview-screenshot:
-.. figure:: _static/the-model-warper/model-warper.jpg
[…]
