+++ docs/source/esmac-2025.rst
@@ -1,0 +1,61 @@
+ESMAC 2025
+==========
+
+The OpenSim Creator team will present a model building and warping workshop at
+the European Society for Movement Analysis in Adults and Children (ESMAC). This
+documentation is a guidebook for that workshop.
+
+
+📅 When/Where is it?
+--------------------
+
+The workshop shall be held at the ESMAC conference on the afternoon of the 10th
+September 2025. The exact time/room is still to-be-decided (TBD).
+
+
+👩‍💻 Prepratory Steps
+----------------------
+
+We recommend attendants prepare for the workshop in the following ways:
+
+- **Download and install OpenSim Creator**. Most of the workshop's content
+  will focus on OpenSim Creator's features for model building and warping.
+- **Download and install OpenSim**. The workshop may also involve using
+  parts of OpenSim to help build or analyze OpenSim models, so you should
+  also install OpenSim.
+- **Familiarize Yourself with the basics of OpenSim/modelling**. We
+  will try to keep the content simple, but it will be much easier to follow
+  if you have already played around with some.
+
+🗺️ Content Overview
+-------------------
+
+TODO: explanation of the various phases of the workshop, maybe a little
+planning diagram so that people can see where they are as we progress
+through the topics.
+
+
+📚 Workshop Materials
+---------------------
+
+The materials for the workshop are published openly and will be hosted
+as part of OpenSim Creator's documentation after the workshop has ended. This
+means you can come back here if you need a refresher.
+
+Other educators and workshop organizers are invited to use these materials
+however they please.
+
+TODO: FIRST PRESENTATION PPT
+
+TODO: MID PRESENTATION PPT
+
+TODO: CLOSING PRESENTATION PPT
+
+
+⚡ Post-Workshop Actions
+-------------------------
+
+- TODO: ask for feedback about the software, general design directions, etc.
+- TODO: ask for feedback via a questionnaire?
+- TODO: mention our upcoming project, maybe someone's interested in
+  collaborating on it?
