### 2025-06-24 #8006 @modenaxe
@davidpagnon I know but I cannot subscribe for some reason! My status revers to unsubscribed automatically!
