### 2025-06-24 6339875
Flesh out ESMAC workshop page with stubbed sections etc.
