+++ docs/source/esmac-2025.rst
@@ -1,61 +1,150 @@
-ESMAC 2025
-==========
+ESMAC 2025 Workshop
+===================
 
 The OpenSim Creator team will present a model building and warping workshop at
 the European Society for Movement Analysis in Adults and Children (ESMAC). This
 documentation is a guidebook for that workshop.
 
 
-📅 When/Where is it?
---------------------
+📅 When/Where
+-------------
 
 The workshop shall be held at the ESMAC conference on the afternoon of the 10th
-September 2025. The exact time/room is still to-be-decided (TBD).
+September 2025. The exact time and room is still to-be-decided (TBD).
 
 
-👩‍💻 Prepratory Steps
-----------------------
+🧑🏽‍🔬 Organizers and Expert Helpers
+-------------------------------------
+
+- Adam Kewley
+- Ajay Seth
+- TBD (Judith/mmreazie?)
+
+
+👩‍💻 Before You Arrive
+-----------------------
 
 We recommend attendants prepare for the workshop in the following ways:
 
-- **Download and install OpenSim Creator**. Most of the workshop's content
-  will focus on OpenSim Creator's features for model building and warping.
-- **Download and install OpenSim**. The workshop may also involve using
-  parts of OpenSim to help build or analyze OpenSim models, so you should
-  also install OpenSim.
-- **Familiarize Yourself with the basics of OpenSim/modelling**. We
-  will try to keep the content simple, but it will be much easier to follow
-  if you have already played around with some.
+- **Download and install OpenSim Creator**. The work sessions will focus
+  on using OpenSim Creator's features for model building and warping.
+  The :ref:`installation` page outlines the procedure.
+- **(optional) Download and install OpenSim**. To answer attendants' specific
+  modelling questions, team members may recommend OpenSim. It can be downloaded
+  from the `OpenSim SimTK Page`_.
+- **(optional) Familiarize Yourself with the basics of OpenSim/modelling**. The
+  work sessions may be easier to follow if you have already gone through an early
+  tutorial such as :doc:`make-a-pendulum`.
 
-🗺️ Content Overview
--------------------
+**If you are already working on an OpenSim model**, bring it along. The workshop
+is structured such that work sessions' walkthroughs and exercises can *usually* be
+completed with any model. The work sessions are an ideal time to ask specific
+questions about your model.
 
-TODO: explanation of the various phases of the workshop, maybe a little
-planning diagram so that people can see where they are as we progress
-through the topics.
+
+🗺️ Workshop Overview
+--------------------
+
+The workshop will last around four hours, with a thirty minute break in the
+middle. There will be three presentations at the start, middle, and end of
+the workshop.
+
+**TODO**: rough planning diagram
+
+The first half of the workshop will be focused on model building. The work
+session will include feature demonstrations by the team, specific model
+building walkthroughs (e.g. building a basic model), and a chance for attendants
+to explore OpenSim Creator or work on their own model with help from the team.
+
+The second half of the workshop will focus on model scaling/warping. We
+will introduce our latest beta workflow, :doc:`the-model-warper`, and explain
+how it can be used to create subject-specific models. The work session will be
+a chance for attendants to give this workflow a try and provide us with feedback
+about it - this time can also be used to continue exploring OpenSim Creator and
+build models.
 
 
 📚 Workshop Materials
 ---------------------
 
-The materials for the workshop are published openly and will be hosted
[…]
