### 2025-06-24 b0513bd
Flesh out ESMAC workshop page with stubbed sections etc.
