### 2025-06-24 cc61189
Flesh out ESMAC workshop page with stubbed sections etc.
