### 2025-06-25 7fcde7d
Add rough schedule, add bullet points outlining presentation requirements
