### 2025-06-26 291c535
Fix MacOS keybindings not re-mapping ctrl to command (#1069)
