+++ liboscar/Platform/Events/KeyEvent.h
@@ -4,6 +4,7 @@
 #include <liboscar/Platform/KeyCombination.h>
 #include <liboscar/Platform/KeyModifier.h>
 #include <liboscar/Platform/Key.h>
+#include <liboscar/Platform/PhysicalKeyModifier.h>
 
 namespace osc
 {
@@ -25,6 +26,7 @@ namespace osc
         KeyModifiers modifiers() const { return combination_.modifiers(); }
         Key key() const { return combination_.key(); }
         bool has_modifier(KeyModifier modifier) const { return static_cast<bool>(modifier & combination_.modifiers()); }
+        bool has_modifier(PhysicalKeyModifier modifier) const { return static_cast<bool>(modifier & to<PhysicalKeyModifiers>(combination_.modifiers())); }
     private:
         explicit KeyEvent(EventType event_type, KeyCombination combination) :
             Event{event_type},
