### 2025-06-27 ec68648
Make minor adjustments to guidebook page
