### 2025-06-27 905e100
Remove sphinx-toolbox dependency
