+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 Sphinx==6.2.1
 sphinx-book-theme==1.0.1
 Jinja2==3.1.3
+
