### 2025-06-27 9a4bbe7
Remove sphinx-toolbox dependency
