+++ docs/requirements.txt
@@ -1,4 +1,4 @@
 Sphinx==6.2.1
 sphinx-book-theme==1.0.1
-sphinx-toolbox==3.8.1
 Jinja2==3.1.3
+
