### 2025-06-27 b6a7aec
Add more notes to docs
