### 2025-06-27 ff46141
Change osc::BoxGeometryParams to take dimensions/segments via a Vec3
