+++ liboscar/Graphics/Geometries/BoxGeometry.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <liboscar/Graphics/Mesh.h>
+#include <liboscar/Maths/Vec3.h>
 #include <liboscar/Utils/CStringView.h>
 
 #include <cstddef>
@@ -10,12 +11,8 @@ namespace osc
     struct BoxGeometryParams final {
         friend bool operator==(const BoxGeometryParams&, const BoxGeometryParams&) = default;
 
-        float width = 1.0f;
-        float height = 1.0f;
-        float depth = 1.0f;
-        size_t num_width_segments = 1;
-        size_t num_height_segments = 1;
-        size_t num_depth_segments = 1;
+        Vec3 dimensions = {1.0f, 1.0f, 1.0f};
+        Vec3uz num_segments = {1, 1, 1};
     };
 
     class BoxGeometry : public Mesh {
