+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLSSAOTab.cpp
@@ -282,7 +282,7 @@ private:
     MouseCapturingCamera camera_ = create_camera_that_matches_learnopengl();
 
     Mesh sphere_mesh_ = SphereGeometry{{.num_width_segments = 32, .num_height_segments = 32}};
-    Mesh cube_mesh_ = BoxGeometry{{.width = 2.0f, .height = 2.0f, .depth = 2.0f}};
+    Mesh cube_mesh_ = BoxGeometry{{.dimensions = Vec3{2.0f}}};
     Mesh quad_mesh_ = PlaneGeometry{{.width = 2.0f, .height = 2.0f}};
 
     // rendering state
