### 2025-06-27 8a93fa9
Refactor PlaneGeometry to use dimensions/num_segments (same as CubeGeometry)
