+++ liboscar/Graphics/Scene/SceneCache.cpp
@@ -258,7 +258,7 @@ private:
     Mesh uncapped_cylinder = CylinderGeometry{{.height = 2.0f, .num_radial_segments = 16, .open_ended = true}};
     Mesh cube = BoxGeometry{{.dimensions = Vec3{2.0f}}};
     Mesh cone = ConeGeometry{{.radius = 1.0f, .height = 2.0f, .num_radial_segments = 16}};
-    Mesh floor = PlaneGeometry{{.width = 2.0f, .height = 2.0f}};
+    Mesh floor = PlaneGeometry{{.dimensions = Vec2{2.0f}}};
     Mesh grid100x100 = GridGeometry{{.num_divisions = 1000}};
     Mesh cube_wireframe = AABBGeometry{};
     Mesh y_line = generate_y_to_y_line_mesh();
