+++ liboscar/Graphics/GraphicsImplementation.cpp
@@ -6694,7 +6694,7 @@ private:
     }};
 
     // a generic quad mesh: two triangles covering NDC @ Z=0
-    Mesh quad_mesh_ = PlaneGeometry{{.width = 2.0f, .height = 2.0f}};
+    Mesh quad_mesh_ = PlaneGeometry{{.dimensions = Vec2{2.0f}}};
 
     // storage for instance data
     std::vector<float> instance_cpu_buffer_;
