### 2025-06-27 c21e9a3
Fix Geometries.tests.cpp
