### 2025-06-28 7ad7e45
Squashed 'third_party/libosim/opensim-core/' changes from c0650141fb..131a3bf6eb
131a3bf6eb Add Component::removeComponent
d14a3aac55 Add OPENSIM_USE_STATIC_LIBRARIES
28e14e6a76 Merge pull request #4106 from opensim-org/wrapping_tests_cleanup
b1cfb679b2 Fix includes for testWrapObject and testGeometryPathWrapping
84074eaba1 Move WrapObject tests to new file; use better name for GeometryPath wrapping tests; move testFunctionBasedPath to OpenSim/Simulation/Test
0673b771c0 Revert include structure in testGeometryPath
8d76498bf4 Consolidate wrapping tests, remove deprecated wrapping tests, and rem
[…]
