### 2025-06-29 #8284 @cvhammond
Hi Adam,

Thank you for the contributions and work that you have put into OpenSim Creator. Based on the Review Checklist, I have the following comments based on the items I have not yet checked off:

## General checks

-  Regarding data sharing, I do not feel it is appropriate to include the pelvis MRI example in the Mesh Warper tutorial if the data used in the tutorial is not available to download and run on a users local computer. Especially considering this submission to JOSS, a closed-data example does not feel appropriate. I would recommend modifying that example to focus on another set of data that is freely downloadable.

- For reproducibility, of the two example use cases that you mention in the paper, the first one does not contain a reference to the results of the work to create a subject-specific model of the human shoulder. Could you add a reference (paper, preprint, SimTK repo, etc) that allows people to interact with the example use case? Similarly, I looked through the thesis cited for the second example and was not able to find information about accessing the results for reproducibility purposes.

## Functionality

- Installation: for the Windows version of the installer, the documentation says to search 'OpenSim Creator' yet nothing appears when doing so. Instead 'opensimcreator' must be searched for Windows to find the software after installation.

- Functionality: I would recommend reorganizing the 'statement of need' and 'design and implementation' sections a bit to allow for a more explicit outline of the features of OpenSim Creator. In the last paragraph of the statement of need section, you roundaboutly indicate some functionality by comparing OpenSim Creator to CAD software, yet it is still ambiguous what functionality is included. Similarly, in bullet point two in the design and implementation section you mention some functionality in parentheses that had not been previously introduced and is not expanded upon later.

- For performance, you mention one of the key attributes of the software is high performance, but do not indicate if or how that performance is measured. OpenSim Creator does feel fast and responsive while using it, but JOSS specifically asks if the performance claims have been confirmed.

## Documentation

- For your statement of need, at the end of the first paragraph you indicate two challenges with creating biomechanical models, but OpenSim Creator is not sufficiently introduced to help the reader understand how it was designed to address these two specific challenges. While the example in paragraph two of the statement of need section is illustrative, it similarly does not answer the question 'what problem is the software designed to solve?'.

- As I touched on in the reproduceability section, readers are not able to access the data for the example use-cases.

- I mentioned above in the functionality section, the core functionality is not sufficiently documented in the paper. It should be clear to the reader what the most important features of OpenSim Creator are and how they can fulfill the users needs.

- I noticed that your automated test builds OpenSim Creator on multiple devices to verify compilation/build after a new commit. How do you verify that the addition of a new feature does not break other previously implemented features?

## Software paper

- Summary: I think the summary could be made more clear to help a non-specialist reader understand the purpose of OpenSim Creator better. Also a pedantic item: in the first sentence of the summary you state OpenSim models previously required manual editing of XML files, but many researchers, including myself, edit OpenSim models programmatically using the MATLAB and/or Python wrappers of the OpenSim API.

- Statement of need: I have mentioned this item above.

- State of the field: It would be good to mention the core functionality in the OpenSim GUI (scale tool, RRA?, etc) as well as mention some existing tools that researchers use to make subject-specific models even if they are script-based rather than UI-based.

- Quality of writing: I am hoping the statement of need section can be modified so readers feel the issues with the current state of the field and understand the unique, novel value of OpenSim Creator.

- References: please include additional information about example use-case number one.
