### 2025-07-08 0f0cdc9
Normalize mouse wheel scrolling delta across all OSes
