### 2025-07-08 15b6234
Explicitly set OSX_SYSROOT to 14.5
