### 2025-07-08 f7c72a2
Put GitHub CI commands into separate ci_ scripts
