+++ scripts/ci_build_ubuntu22.sh
@@ -1,0 +1,15 @@
+#!/usr/bin/env bash
+#
+# Performs an end-to-end CI build of OpenSim Creator. This is what build
+# agents should run if they want to build release amd64 binaries of OpenSim
+# Creator on Ubuntu 22.04
+
+set -xeuo pipefail
+
+# Hide window creation in CI, because CI runners typically do not have
+# a desktop environment.
+export OSC_INTERNAL_HIDE_WINDOW="1"
+
+# Run buildscript under virtual desktop with `xvfb-run` (for UI tests)
+CC=gcc-12 CXX=g++-12 OSC_BUILD_CONCURRENCY=$(nproc) xvfb-run ./scripts/build_ubuntu.sh
+
