+++ scripts/ci_setup_ubuntu22.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+sudo apt update
+sudo ./scripts/setup_ubuntu.sh
+sudo apt-get install -y gcc-12 g++-12 xvfb
