+++ .github/workflows/continuous-integration-workflow.yml
@@ -1,32 +1,21 @@
 on: [push]
 
 env:
-  # Bake a build ID into the binary so that there's a way of looking up
-  # the build logs for a build.
+  # Bake a GitHub build ID into the binary so that there's a way of looking
+  # up the build logs for a build via GitHub Actions' UI/API.
   OSC_BUILD_ID: "GITHUB_${{ github.run_id }}.${{ github.run_number }}"
-
-  # Hide window creation in CI, because CI runners typically do not have
-  # a desktop environment.
-  OSC_INTERNAL_HIDE_WINDOW: "1"
 
 jobs:
 
-  windows10_msvc_amd64:
-    name: Windows 2022 (amd64, MSVC)
+  windows_amd64:
+    name: Windows (amd64)
     runs-on: windows-2022
     steps:
 
     - uses: actions/checkout@v4
 
     - name: build
-      run: |
-        cd $env:GITHUB_WORKSPACE
-
-        # --system-version is necessary with GitHub's `windows-2022` runner because the
-        # default SDK doesn't include `gameinput.h` correctly (see: https://github.com/actions/runner-images/issues/10980)
-        #
-        # the specified SDK supports Windows 10, version 1507 or higher (developer.microsoft.com/en-us/windows/downloads/windows-sdk/)
-        python scripts\\build_windows.py --system-version=10.0.26100.0 || exit /b
+      run: Set-Location $env:GITHUB_WORKSPACE -ErrorAction Stop; .\scripts\ci_build_windows.ps1
 
     - uses: actions/upload-artifact@v4
       with:
@@ -34,26 +23,15 @@ jobs:
         path: build/*windows-amd64.exe
         if-no-files-found: error
 
-  macos14_xcode15-4_targeting14-5_amd64:
-    name: MacOS 14 (amd64, XCode 15.4, targeting 14.5)
+  macos_amd64:
+    name: MacOS (amd64)
     runs-on: macos-14
     steps:
 
     - uses: actions/checkout@v4
 
     - name: run build script
-      run: |
-        cd $GITHUB_WORKSPACE
-
-        # configure buildscript
-        export OSC_CMAKE_CONFIG_EXTRA="-DCMAKE_OSX_ARCHITECTURES=x86_64 -DCMAKE_OSX_DEPLOYMENT_TARGET=14.5 -DCMAKE_OSX_SYSROOT=/Library/Developer/CommandLineTools/SDKs/MacOSX14.5.sdk/"
-
-        # run buildscript
-        ./scripts/build_mac.sh
-
-        # run after-build OS-specific test scripts
-        ./scripts/macos_check-dependencies.py build/osc/osc
-        ./scripts/macos_check-sdk.py 14.5 build/osc/osc
+      run: cd $GITHUB_WORKSPACE && ./scripts/ci_build_mac-amd64.sh
 
     - uses: actions/upload-artifact@v4
       with:
@@ -61,26 +39,14 @@ jobs:
         path: build/*.dmg
         if-no-files-found: error
 
-  macos14_xcode15-4_targeting14-5_arm64:
-    name: MacOS 14 (arm64, XCode 15.4, targeting 14.5)
+  macos_arm64:
+    name: MacOS (arm64)
     runs-on: macos-14
     steps:
-
     - uses: actions/checkout@v4
 
     - name: run build script
-      run: |
-        cd $GITHUB_WORKSPACE
-
-        # configure buildscript
[…]
