### 2025-07-09 8a3aae2
Clean up comments, remove GitHub mentions in certain places
