### 2025-07-09 f680b30
Add FORCE when setting OSC_BUILD_ID via an environment variable
