### 2025-07-09 dc951bc
De-duplicate ci_build_mac-amd64.sh and ci_build_mac-arm64.sh into a single script
