### 2025-07-09 58b283e
Refactor ImGuizmo::SetRect to accept an osc::Rect
