### 2025-07-09 3c2e233
Refactor enums in ImGuizmo to more modern enum class style etc.
