+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -69,6 +69,7 @@ namespace
     constexpr float ZPI = 3.14159265358979323846f;
     constexpr float DEG2RAD = (ZPI / 180.f);
     constexpr float screenRotateSize = 0.06f;
+
     // scale a bit so translate axis do not touch when in universal
     constexpr float rotationDisplayFactor = 1.2f;
     const char* translationInfoMask[] = {
@@ -107,30 +108,49 @@ namespace
     constexpr int halfCircleSegmentCount = 64;
     constexpr float snapTension = 0.5f;
 
+    enum COLOR {
+        DIRECTION_X,      // directionColor[0]
+        DIRECTION_Y,      // directionColor[1]
+        DIRECTION_Z,      // directionColor[2]
+        PLANE_X,          // planeColor[0]
+        PLANE_Y,          // planeColor[1]
+        PLANE_Z,          // planeColor[2]
+        SELECTION,        // selectionColor
+        INACTIVE,         // inactiveColor
+        TRANSLATION_LINE, // translationLineColor
+        SCALE_LINE,
+        ROTATION_USING_BORDER,
+        ROTATION_USING_FILL,
+        HATCHED_AXIS_LINES,
+        TEXT,
+        TEXT_SHADOW,
+        COUNT
+    };
+
     constexpr ImGuiID blank_id()
     {
         return std::numeric_limits<ImGuiID>::max();
     }
 
-    OPERATION operator&(OPERATION lhs, OPERATION rhs)
+    Operation operator&(Operation lhs, Operation rhs)
     {
-        return static_cast<OPERATION>(static_cast<int>(lhs) & static_cast<int>(rhs));
+        return static_cast<Operation>(static_cast<int>(lhs) & static_cast<int>(rhs));
     }
 
-    bool operator!=(OPERATION lhs, int rhs)
+    bool operator!=(Operation lhs, int rhs)
     {
         return static_cast<int>(lhs) != rhs;
     }
 
-    bool Intersects(OPERATION lhs, OPERATION rhs)
+    bool Intersects(Operation lhs, Operation rhs)
     {
-        return (lhs & rhs) != 0;
+        return (std::to_underlying(lhs) & std::to_underlying(rhs)) != 0;
     }
 
     // True if lhs contains rhs
-    bool Contains(OPERATION lhs, OPERATION rhs)
+    bool Contains(Operation lhs, Operation rhs)
     {
-        return (lhs & rhs) == rhs;
+        return static_cast<Operation>(std::to_underlying(lhs) & std::to_underlying(rhs)) == rhs;
     }
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -602,10 +622,10 @@ namespace
    }
 
    // Matches MT_MOVE_AB order
-   constexpr OPERATION TRANSLATE_PLANS[3] = {
-       TRANSLATE_Y | TRANSLATE_Z,
-       TRANSLATE_X | TRANSLATE_Z,
-       TRANSLATE_X | TRANSLATE_Y
+   constexpr Operation TRANSLATE_PLANS[3] = {
+       Operation::TranslateY | Operation::TranslateZ,
+       Operation::TranslateX | Operation::TranslateZ,
+       Operation::TranslateX | Operation::TranslateY
    };
 
    struct Style {
@@ -668,7 +688,7 @@ namespace
       ImDrawList* mDrawList;
       Style mStyle;
 
-      MODE mMode;
+      Mode mMode;
       matrix_t mViewMat;
       matrix_t mProjectionMat;
       matrix_t mModel;
@@ -755,7 +775,7 @@ namespace
 
       ImVector<ImGuiID> mIDStack;
       ImGuiID mEditingID = blank_id();
[…]
