### 2025-07-10 7eaf517
Remove unused functions in ImGuizmo, refactor to use overload&
