+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -132,21 +132,6 @@ namespace
         return std::numeric_limits<ImGuiID>::max();
     }
 
-    Operation operator&(Operation lhs, Operation rhs)
-    {
-        return static_cast<Operation>(static_cast<int>(lhs) & static_cast<int>(rhs));
-    }
-
-    bool operator!=(Operation lhs, int rhs)
-    {
-        return static_cast<int>(lhs) != rhs;
-    }
-
-    bool Intersects(Operation lhs, Operation rhs)
-    {
-        return (std::to_underlying(lhs) & std::to_underlying(rhs)) != 0;
-    }
-
     // True if lhs contains rhs
     bool Contains(Operation lhs, Operation rhs)
     {
@@ -1144,7 +1129,7 @@ namespace
 
    void DrawRotationGizmo(Operation op, int type)
    {
-      if(!Intersects(op, Operation::Rotate))
+      if(not (op & Operation::Rotate))
       {
          return;
       }
@@ -1173,11 +1158,10 @@ namespace
 
       gCurrentContext->mRadiusSquareCenter = screenRotateSize * gCurrentContext->mHeight;
 
-      bool hasRSC = Intersects(op, Operation::RotateInScreen);
+      bool hasRSC = op & Operation::RotateInScreen;
       for (int axis = 0; axis < 3; axis++)
       {
-         if(!Intersects(op, Operation::RotateZ >> axis))
-         {
+         if (not (op & (Operation::RotateZ >> axis))) {
             continue;
          }
 
@@ -1264,8 +1248,7 @@ namespace
    {
       ImDrawList* drawList = gCurrentContext->mDrawList;
 
-      if(!Intersects(op, Operation::Scale))
-      {
+      if (not (op & Operation::Scale)) {
         return;
       }
 
@@ -1283,7 +1266,7 @@ namespace
 
       for (int i = 0; i < 3; i++)
       {
-         if(!Intersects(op, Operation::ScaleX << i))
+         if (not (op & (Operation::ScaleX << i)))
          {
             continue;
          }
@@ -1351,7 +1334,7 @@ namespace
    {
       ImDrawList* drawList = gCurrentContext->mDrawList;
 
-      if (!Intersects(op, Operation::ScaleU))
+      if (not (op & Operation::ScaleU))
       {
          return;
       }
@@ -1370,7 +1353,7 @@ namespace
 
       for (int i = 0; i < 3; i++)
       {
-         if (!Intersects(op, Operation::ScaleXU << i))
+         if (not (op & (Operation::ScaleXU << i)))
          {
             continue;
          }
@@ -1439,7 +1422,7 @@ namespace
          return;
       }
 
-      if(!Intersects(op, Operation::Translate))
+      if(not (op & Operation::Translate))
       {
          return;
       }
@@ -1461,7 +1444,7 @@ namespace
          if (!gCurrentContext->mbUsing || (gCurrentContext->mbUsing && type == MT_MOVE_X + i))
          {
             // draw axis
-            if (belowAxisLimit && Intersects(op, Operation::TranslateX << i))
+            if (belowAxisLimit && (op & (Operation::TranslateX << i)))
             {
                ImVec2 baseSSpace = worldToPos(dirAxis * 0.1f * gCurrentContext->mScreenFactor, gCurrentContext->mMVP);
                ImVec2 worldDirSSpace = worldToPos(dirAxis * gCurrentContext->mScreenFactor, gCurrentContext->mMVP);
@@ -1654,21 +1637,17 @@ namespace
             int type = MT_NONE;
             vec_t gizmoHitProportion;
 
-            if(Intersects(operation, Operation::Translate))
-            {
[…]
