### 2025-07-10 326364e
Refactor ImGuizmo to use more osc:: types
