+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -52,11 +52,17 @@
 #include "imgui_internal.h"
 #include "ImGuizmo.h"
 
+#include <liboscar/Maths/Mat4.h>
+#include <liboscar/Maths/MatFunctions.h>
 #include <liboscar/Maths/Rect.h>
 #include <liboscar/Maths/RectFunctions.h>
+#include <liboscar/Maths/Transform.h>
+#include <liboscar/Maths/TransformFunctions.h>
+#include <liboscar/Utils/ScopeExit.h>
 
 #include <functional>
 #include <limits>
+#include <optional>
 #include <vector>
 
 using namespace ImGuizmo;
@@ -1105,7 +1111,7 @@ namespace
       }
    }
 
-   void ComputeSnap(vec_t& value, const float* snap)
+   void ComputeSnap(vec_t& value, const osc::Vec3& snap)
    {
       for (int i = 0; i < 3; i++)
       {
@@ -1998,7 +2004,7 @@ namespace
       return type;
    }
 
-   bool HandleTranslation(float* matrix, float* deltaMatrix, Operation op, int& type, const float* snap)
+   bool HandleTranslation(float* matrix, float* deltaMatrix, Operation op, int& type, const std::optional<OperationSnappingSteps>& snap)
    {
       if (not (op & Operation::Translate) or type != MT_NONE) {
         return false;
@@ -2033,7 +2039,7 @@ namespace
          }
 
          // snap
-         if (snap)
+         if (snap and snap->position)
          {
             vec_t cumulativeDelta = gCurrentContext->mModel.v.position + delta - gCurrentContext->mMatrixOrigin;
             if (applyRotationLocaly)
@@ -2043,15 +2049,14 @@ namespace
                matrix_t modelSourceNormalizedInverse;
                modelSourceNormalizedInverse.Inverse(modelSourceNormalized);
                cumulativeDelta.TransformVector(modelSourceNormalizedInverse);
-               ComputeSnap(cumulativeDelta, snap);
+               ComputeSnap(cumulativeDelta, *snap->position);
                cumulativeDelta.TransformVector(modelSourceNormalized);
             }
             else
             {
-               ComputeSnap(cumulativeDelta, snap);
+               ComputeSnap(cumulativeDelta, *snap->position);
             }
             delta = gCurrentContext->mMatrixOrigin + cumulativeDelta - gCurrentContext->mModel.v.position;
-
          }
 
          if (delta != gCurrentContext->mTranslationLastDelta)
@@ -2120,7 +2125,7 @@ namespace
       return modified;
    }
 
-   bool HandleScale(float* matrix, float* deltaMatrix, Operation op, int& type, const float* snap)
+   bool HandleScale(float* matrix, float* deltaMatrix, Operation op, int& type, const std::optional<OperationSnappingSteps>& snap)
    {
       if ((not (op & Operation::Scale) and not (op & Operation::ScaleU))
           or type != MT_NONE
@@ -2196,10 +2201,8 @@ namespace
          }
 
          // snap
-         if (snap)
-         {
-            float scaleSnap[] = { snap[0], snap[0], snap[0] };
-            ComputeSnap(gCurrentContext->mScale, scaleSnap);
+         if (snap and snap->scale) {
+            ComputeSnap(gCurrentContext->mScale, *snap->scale);
          }
 
          // no 0 allowed
@@ -2245,7 +2248,7 @@ namespace
       return modified;
    }
 
-   bool HandleRotation(float* matrix, float* deltaMatrix, Operation op, int& type, const float* snap)
[…]
