### 2025-07-10 f1b16f7
Ensure ImGuizmo::Manipulate cannot throw when decomposing the delta matrix
