### 2025-07-10 42efaae
Fix keybinds not working in mesh importer (#1072)
