### 2025-07-10 94aad18
Remove now-unused DEG2RAD constant from ImGuizmo
