### 2025-07-10 f5f641f
Make it possible to edit models with not-equilibrateable muscles (#1070)
