### 2025-07-10 88b7302
Add 'Add Path Point' contextual action to OpenSim::GeometryPath (#1061)
