### 2025-07-10 66433d5
Make loading a TRC file containing invalid column headers a warning, rather than error (#1068)
