### 2025-07-10 a1ca25e
Harden 'Preview Experimental Data' workflow against crashes when loading data (#1068)
