### 2025-07-10 62c150c
Update CHANGELOG.md (#1070, #1068)
