### 2025-07-10 68d91c3
Add SceneDecorationFlag::NoSceneVolumeContribution (#1071)
