### 2025-07-10 22c9a83
Revert hacked-in removal of 'try' from speculative testing
