### 2025-07-11 c69884d
Move CHANGELOG.md entry to correct section (#1071)
