### 2025-07-11 1e430e8
Use /fp:contract when compiling with MSVC
