### 2025-07-11 412d937
Remove __LINUX__ compiler definition (unused)
