+++ scripts/ci_build_ubuntu22.sh
@@ -12,4 +12,3 @@ export OSC_INTERNAL_HIDE_WINDOW="1"
 
 # Run buildscript under virtual desktop with `xvfb-run` (for UI tests)
 CC=gcc-12 CXX=g++-12 OSC_BUILD_CONCURRENCY=$(nproc) xvfb-run ./scripts/build_ubuntu.sh
-
