### 2025-07-11 c6fc508
Remove LUNASVG_BUILD_STATIC (now handled by upstream lunasvg)
