+++ liboscar/CMakeLists.txt
@@ -55,8 +55,6 @@ target_compile_definitions(oscar
 PUBLIC
     $<$<BOOL:${OSC_FORCE_ASSERTS_ENABLED}>:OSC_FORCE_ASSERTS_ENABLED>
     $<$<BOOL:${OSC_RUNTIME_PERF_MEASUREMENTS_ENABLED}>:OSC_RUNTIME_PERF_MEASUREMENTS_ENABLED>
-PRIVATE
-    LUNASVG_BUILD_STATIC
 )
 target_include_directories(oscar
 PUBLIC
