### 2025-07-11 99efdf8
Default OSCDEPS_BUILD_ALWAYS to ON in CI build scripts
