+++ scripts/ci_build_ubuntu22.sh
@@ -10,5 +10,9 @@ set -xeuo pipefail
 # a desktop environment.
 export OSC_INTERNAL_HIDE_WINDOW="1"
 
+# Ensure dependencies are re-checked/re-built if the CI script is ran on
+# a potentially stale/cached workspace directory.
+export OSCDEPS_BUILD_ALWAYS=${OSCDEPS_BUILD_ALWAYS:-ON}
+
 # Run buildscript under virtual desktop with `xvfb-run` (for UI tests)
 CC=gcc-12 CXX=g++-12 OSC_BUILD_CONCURRENCY=$(nproc) xvfb-run ./scripts/build_ubuntu.sh
