### 2025-07-11 e755bf5
Squashed 'third_party/libosim/opensim-core/' changes from 131a3bf6eb..28e14e6a76
REVERT: 131a3bf6eb Add Component::removeComponent
REVERT: d14a3aac55 Add OPENSIM_USE_STATIC_LIBRARIES

git-subtree-dir: third_party/libosim/opensim-core
git-subtree-split: 28e14e6a76fd2a96cfac3bc36357127f3e53862f
