+++ OpenSim/Actuators/osimActuatorsDLL.h
@@ -38,12 +38,10 @@
 #define WIN32_LEAN_AND_MEAN
 #define NOMINMAX
 #include <windows.h>
-#if defined(OPENSIM_USE_STATIC_LIBRARIES)
-    #define OSIMACTUATORS_API
-#elif defined (OSIMACTUATORS_EXPORTS)
-    #define OSIMACTUATORS_API __declspec(dllexport)
+#ifdef OSIMACTUATORS_EXPORTS
+#define OSIMACTUATORS_API __declspec(dllexport)
 #else
-    #define OSIMACTUATORS_API __declspec(dllimport)
+#define OSIMACTUATORS_API __declspec(dllimport)
 #endif
 
 #endif // PLATFORM
