### 2025-07-11 919f438
Reduce in-tree mentions of GitHub actions (build is no longer dependent on it)
