### 2025-07-13 c03d8dd
Refactor osc::Rect to store an origin + dimensions instead of two corners
