+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLShadowMappingTab.cpp
@@ -97,7 +97,7 @@ private:
     void draw_3d_scene()
     {
         const Rect workspace_screen_space_rect = ui::get_main_window_workspace_screen_space_rect();
-        const Vec2 top_left = top_left_rh(workspace_screen_space_rect);
+        const Vec2 workspace_screen_space_top_left = workspace_screen_space_rect.ypu_top_left();
         constexpr float depth_overlay_size = 200.0f;
 
         render_shadows_to_depth_texture();
@@ -114,7 +114,7 @@ private:
         camera_.set_pixel_rect(workspace_screen_space_rect);
         camera_.render_to_main_window();
         camera_.set_pixel_rect(std::nullopt);
-        graphics::blit_to_main_window(depth_texture_, Rect{top_left - Vec2{0.0f, depth_overlay_size}, top_left + Vec2{depth_overlay_size, 0.0f}});
+        graphics::blit_to_main_window(depth_texture_, Rect{workspace_screen_space_top_left - Vec2{0.0f, depth_overlay_size}, workspace_screen_space_top_left + Vec2{depth_overlay_size, 0.0f}});
 
         scene_material_.unset("uShadowMapTexture");
     }

[…]
