### 2025-07-14 04388b8
Add left(), right(), ypd_top(), etc. to Rect and refactor uses
