### 2025-07-14 cee9e3d
Fix minor issues found by clang-tidy (#1075)
