### 2025-07-14 4cf94e0
Make minor cleanups/refactors for readability
