### 2025-07-14 ee748f5
Update CHANGELOG for 0.5.25 (#1075)
