### 2025-07-14 ComputationalBiomechanicsLab/opensim-creator#1077 @adamkewley
Change mesh warper tutorial to use a more accessible femur mesh file
Suggested during https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-3016669587

JOSS paper reviewer pointed out that the pelvis MRI example in the mesh warper documentation isn't reproducible.

Although the documentation does ask users to "bring their own meshes" it might just be an easy-enough fix to change those documentation pages to walk through with the OpenSim femur or similar.
