### 2025-07-14 ComputationalBiomechanicsLab/opensim-creator#1078 @adamkewley
Change mesh warper tutorial to use an accessible mesh file
Mentioned during JOSS review: https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-3016669587

> Regarding data sharing, I do not feel it is appropriate to include the pelvis MRI example in the Mesh Warper tutorial if the data used in the tutorial is not available to download and run on a users local computer. Especially considering this submission to JOSS, a closed-data example does not feel appropriate. I would recommend modifying that example to focus on another set of data that is freely downloadable.
