### 2025-07-14 ComputationalBiomechanicsLab/opensim-creator#1079 @adamkewley
Make "OpenSim Creator" work in Windows' start bar or update documentation
Found during JOSS review: https://github.com/openjournals/joss-reviews/issues/8284#issuecomment-3016669587

> Installation: for the Windows version of the installer, the documentation says to search 'OpenSim Creator' yet nothing appears when doing so. Instead 'opensimcreator' must be searched for Windows to find the software after installation.
