### 2025-07-14 #8284 @adamkewley
@Kevin-Mattheus-Moerman , based on @davidpagnon 's and @cvhammond 's reviews, below are TODOs that I need to work on before this can move forward.

# OpenSim Creator Codebase Issues

I have added concrete issues/changes that I collected from the reviewers to the OSC repository with the tag [joss-2025](https://github.com/ComputationalBiomechanicsLab/opensim-creator/issues?q=is%3Aissue%20state%3Aopen%20label%3Ajoss-2025). Most of the issues I've tagged this way are (imo) technical changes that shouldn't block the paper. However, it's fair to assert that the following issues do need to be fixed before moving forward:

- https://github.com/ComputationalBiomechanicsLab/opensim-creator/issues/1078

# Paper TODOs

These are things related to the paper's content that need to be changed:

- [ ] From @davidpagnon : "There is no "State-of-the-field" section, but the question is covered in the "Statement of needs". It will be up to the editor's point of view whether they want to stick to the demanded outline or not."
- [ ] From @cvhammond : "For reproducibility, of the two example use cases that you mention in the paper, the first one does not contain a reference to the results of the work to create a subject-specific model of the human shoulder."
- [ ] From @cvhammond : "I would recommend reorganizing the 'statement of need' and 'design and implementation' sections a bit to allow for a more explicit outline of the features of OpenSim Creator."
- [ ] From @cvhammond : "For performance, you mention one of the key attributes of the software is high performance, but do not indicate if or how that performance is measured. "
- [ ] From @cvhammond : "For your statement of need, at the end of the first paragraph you indicate two challenges with creating biomechanical models, but OpenSim Creator is not sufficiently introduced to help the reader understand how it was designed to address these two specific challenges."
- [ ] From @cvhammond : "the core functionality is not sufficiently documented in the paper."
- [ ] From @cvhammond : "I think the summary could be made more clear to help a non-specialist reader understand the purpose of OpenSim Creator better. "
- [ ] From @cvhammond : "State of the field: It would be good to mention the core functionality in the OpenSim GUI (scale tool, RRA?, etc) as well as mention some existing tools that researchers use to make subject-specific models even if they are script-based rather than UI-based."

---

I'll update this issue once I believe I have handled all of these points.
