### 2025-07-14 2b17bdc
Update CITATION.cff, README.md, and codemeta.json
