### 2025-07-15 #8284 @modenaxe
@adamkewley I will provide my comments by the end of today and some are related to the statement of need as well.
If you can wait to work on that until I post them you might save some time.
Sorry this is taking long!
