### 2025-07-15 2ada583
Remove sphinx extension usages
