### 2025-07-15 #8284 @modenaxe
@adamkewley  @Kevin-Mattheus-Moerman  I have reviewed the paper and the documentation and tested the software. Below are some of my thoughts.

## Software paper
The paper is well-written and overall well describes the software design and principles.
Before ticking off the review’s items, however, I have some comments, some of which have been suggested in various degrees also by other reviewers:
-	I feel that the state of the field is not very well represented in the paper. Various options to create OpenSim models are not listed. The main one that I am aware of is NMSBuilder  (https://ior-bic.github.io/software/nmsbuilder/, published at https://doi.org/10.1016/j.cmpb.2017.09.012) based on the ALBA library (https://github.com/IOR-BIC/ALBA, published at https://doi.org/10.1016/j.softx.2025.102188). Another, although I believe currently not maintained, is the MAP client (https://link.springer.com/chapter/10.1007/978-3-319-12057-7_21). OpenSim itself has some (limited) options for modifying some elements through the UI and more articulated ones through the APIs – both more sophisticated than XML editing.
-	I have little experience with OS Creator (OCC), and I find it overall a very nicely crafted software. I do not fully understand, however, its scope with respect to OpenSim. The statement of need focuses on researchers from different fields willing to create new musculoskeletal models that will be then used in OpenSim. OCC, however, has some simulations capabilities and to what extent these overlap with OpenSim it is not specified anywhere, e.g. can I specify muscle excitations? There is also overlap with some OpenSim functionalities like markers/force visualisation etc. Tutorial 6, for example, has nothing to do with creating models, so it is outside the current scope of the statement of need. In summary: can you clarify the software scope, i.e. what a user is meant to do in this software environment and what they are supposed to do in OpenSim in your workflow design? For example, it would be clumsy to use OCC to create a model and inspect some experimental data, but then jump to OpenSim to run inverse kinematics and then come back to OCC to see the results of that. Explicitly including an educational use in the statement of need, e.g. having students practicing building models and running basic simulations in the classroom as an additional target, would improve the consistency with the JOSS requirements in my opinion.
-	One of the main issues of other existing modelling options like NMSBuilder is that it is designed to “manually” build each model component separately, i.e. they do not work well if you want to apply the same procedure multiple times, e.g. create 30 different muscles to include in the model. For that kind of operation, the programmatic OpenSim APIs have always been more practical in my experience. Does OCC have some functionality to overcome this limitation? I am asking because I have noticed that most example consist of models including a limited number of bodies/muscles. More generally, can you realistically use OCC for building “large scale” models in reasonable operator time? Using NMSBuilder, for example, it takes hours for an expert user to create a full lower limb - you can see a typical procedure at the [supplementary materials](https://orda.shef.ac.uk/articles/dataset/Data_for_paper_Investigation_of_the_dependence_of_joint_contact_forces_on_musculotendon_parameters_using_a_codified_workflow_for_image-based_modelling_/5863422) of this paper that we published a while ago (link: https://doi.org/10.1016/j.jbiomech.2018.03.039). I am not suggesting to cite it, but using it as an example of existing, codified model creation workflow using another software. Could OCC streamline a workflow of that type or is it designed more for adapting/morphing existing models?
-	Software design: it is great to have rapid a clear philosophy of software prototyping, in your case quickly shipping new versions based on user feedback. This might be not ideal for users though, because they will have to manually update the software several times per year if they want to use the latest version. Have you considered including a mechanism for updating the software within the application? FEBio (https://febio.org/) for example has a “Check for updates” and “Update” option and distinguishes between developer versions and main releases.
-	The use-cases are nice examples, and work well as an high level description, but it would be nice to have some additional details for potential users familiar with OpenSim (which will be the majority of your users: excluding the use for teaching, I assume most researchers wanting to build OpenSim models already know how to use that software to generate simulations). In the shoulder example: using the real-time model processing, subject-specific joints and bodies were manually defined based on what criteria? Was the generic model used as a reference for joint definition? Were the moment arms of the generic model used for reference imported/available in OCC during model creation? Were complex components like ellipsoidal joint and scapular rhythm constraint also implemented in the new model using OCC (if I recall correctly, they were included in the generic model)? I am asking these questions to understand the level of complexity that can be achieved in a model with OCC.  I also think, as highlighted by another reviewer, that it would be ideal to have the shoulder model made available, so that a learning user has a reference endpoint when trying to generate a research-quality model.
-	Could the second example focusing on model simplification while preserving moment arms not also be performed in OpenSim, through the muscle editing functionalities of their UI/visualiser? I am trying to identify the specific advantage in using OCC for this task: is it the real-time feedback and visualisation of the moment arms?
-	Claiming quick model creation is certainly a major achievement and appeal of a software a scope like OCC, but the sentence at line 94-96 in my opinion needs more evidence. “Our users report significantly reduced amounts of time and effort” lets the reader wondering: time was reduced compared to which other “existing approach” to create OpenSim models (none was mentioned excepts XML editing)? What were these users trying to do? How complex was the model?
-	Similarly, “some researchers already have” used the Model Warper (from the documentation of the tool)…it would be nice to know who they are and what they have produced, so the reader can learn from that.
