### 2025-07-16 978103a
Make osc::Rect corners ctor explicit
