### 2025-07-16 fa205b6
Add Rect::from_corners factory function
