### 2025-07-16 27a7983
Refactor osc::Rect to require using explicit Rect::from_points factory function
