+++ liboscar/Maths/MathHelpers.tests.cpp
@@ -9,8 +9,8 @@ using namespace osc;
 
 TEST(ndc_rect_to_topleft_viewport_rect, works_with_basic_example)
 {
-    const Rect ndc_rect{Vec2{-0.5}, Vec2{+0.5}};
-    const Rect viewport_rect{Vec2{10.0f, 100.0f}, Vec2{1034.0f, 1200.0f}};
+    const Rect ndc_rect = Rect::from_corners(Vec2{-0.5}, Vec2{+0.5});
+    const Rect viewport_rect = Rect::from_corners({10.0f, 100.0f}, {1034.0f, 1200.0f});
 
     const Rect result = ndc_rect_to_topleft_viewport_rect(ndc_rect, viewport_rect);
 
