### 2025-07-17 3dc2cdb
Make 'OpenSim Creator' the search term for osc (previously: 'opensimcreator', #1079)
