+++ osc/CMakeLists.txt
@@ -165,12 +165,11 @@ target_link_options(osc PRIVATE
 
 # set shared/common CPack variables
 set(CPACK_PACKAGE_NAME "${OSC_PACKAGE_NAME}")
-set(CPACK_RESOURCE_FILE_LICENSE "${PROJECT_SOURCE_DIR}/LICENSE")
 set(CPACK_PACKAGE_VENDOR "${OSC_AUTHOR}")
 set(CPACK_PACKAGE_CONTACT "${OSC_AUTHOR} <${OSC_AUTHOR_EMAIL}>")
 set(CPACK_PACKAGE_HOMEPAGE_URL "${OSC_REPO_URL}")
 set(CPACK_PACKAGE_DESCRIPTION "${OSC_SOFTWARE_DESCRIPTION}")
-set(CPACK_PACKAGE_EXECUTABLES "osc;${OSC_PACKAGE_NAME}")
+set(CPACK_PACKAGE_EXECUTABLES "osc;${OSC_LONG_APPNAME}")
 
 # `include` platform-specific packaging instructions
 include(${OSC_PACKAGING_DIR}/Packaging.cmake)
