### 2025-07-17 3e3092a
Changed 'load raw data file' to 'load sto/mot/trc file' (#1057)
