+++ CHANGELOG.md
@@ -11,6 +11,9 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The "load raw data file" button in the preview experimental data workflow was
   changed to "load sto/mot/trc file" to clarify its purpose (reported by
   @davidpagnon, #1057).
+- A small contribution section was added to the README, which directs users to
+  the contribution section in the official documentation, which is also available
+  in-tree (reported by @davidpagnon, #1056).
 
 
 ## [0.5.25] - 2025/07/14
