### 2025-07-17 38400db
Change 'load model trajectory/states' wording for clarity (#1052)
