### 2025-07-18 5d8304b
Add link to top-level ESMAC2025 issue (#1081)
