### 2025-07-18 cefa3b1
Create GitHub Action that automatically uploads esmac2025 documentation (#1081)
