### 2025-07-18 2ed28c9
Change key permissions
