### 2025-07-18 71ef923
Add OneDrive links (temporarily) to ESMAC2025 workshop page (#1081)
