+++ docs/source/esmac-2025.rst
@@ -91,8 +91,7 @@ after the workshop has ended.
 1. Presentation: Model Building with OpenSim Creator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-**TODO**: OpenSim and Model Building Presentation. Target duration is 15-30
-minutes and should include:
+Approx 20 min., LINK: `Presentation, Model Building with OpenSim Creator`_
 
 1. Welcome
 2. Quick overview of the workshop plan (above)
@@ -113,7 +112,12 @@ minutes and should include:
 2. Work Session: Model Building
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-**TODO**: work session exercises, example models, walkthroughs, etc.
+Approx 90 min., LINK: `Work Session, Model Building`_
+
+TODO: provide a list of actions, exercies, etc. here in the
+documentation page. Work session exercises, example models,
+walkthroughs, etc. The 'presentation' for this section is just
+a backdrop we can throw up while people mill around etc.
 
 - Direct people to the tutorials or ESMAC-specific exercises. Maybe shape
   it such that morphometric (mesh) scaling is an available thing they can
@@ -134,14 +138,17 @@ minutes and should include:
 3. Break
 ^^^^^^^^
 
+30 min., LINK: `Backdrop, Break`_
+
 Chill out for 30 minutes. Stretch your legs, chat to other attendants, etc.
 
 
 4. Presentation: Model Warping with OpenSim Creator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-**TODO**: Model warping presentation. Target duration is 15-25 minutes and
-should include:
+Approx 20 min., LINK: `Presentation, Model Warping with OpenSim Creator`_
+
+Presentation should include:
 
 1. Introduction, what scaling is, where it's handy (obvious)
 2. Quick overview of existing model scaling in OpenSim
@@ -157,8 +164,9 @@ should include:
 5. Work Session: Model Warping
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-**TODO**: work session exercises, example models, walkthroughs, etc. Target
-duration is 1h30-1h45 and should include:
+Approx 90 min., LINK: `Work Session, Model Warping with OpenSim Creator`_
+
+**TODO**: work session exercises, example models, walkthroughs, etc.
 
 - Invite attendants to work through :doc:`the-model-warper`
 - Ideally, get people to warp the model made in the first half, walking through
@@ -176,6 +184,8 @@ duration is 1h30-1h45 and should include:
 6. Presentation: Closing Thoughts, Outlook
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Approx 10 min., LINK: `Presentation, Closing Thoughts`_
+
 **TODO**: Final closing presentation. Target duration is 10-20 minutes and should
 include:
 
@@ -190,7 +200,6 @@ include:
 ⚡ Post-Workshop Actions
 -------------------------
 
-- **TODO**: ask for feedback about the software, general design directions, etc.
 - **TODO**: ask for feedback via a questionnaire?
 - **TODO**: mention our upcoming project, maybe someone's interested in
   collaborating on it?
@@ -198,3 +207,9 @@ include:
[…]
