### 2025-07-21 3c940fb
Clean up ESMAC page (info is now in stubbed powerpoint docs #1081)
