+++ docs/source/esmac-2025.rst
@@ -19,7 +19,7 @@ Adults and Children (ESMAC). This documentation is a guidebook for that workshop
 -------------
 
 The workshop shall be held at the ESMAC conference on the afternoon of the 10th
-September 2025. The exact time and room is still to-be-decided (:red:`TBD`).
+September 2025. The exact time and room is still to-be-decided (:red:`TODO`).
 
 
 🧑🏽‍🔬 Organizers and Helpers
@@ -27,7 +27,7 @@ September 2025. The exact time and room is still to-be-decided (:red:`TBD`).
 
 - Adam Kewley
 - Ajay Seth
-- :red:`Other helpers TBD`
+- :red:`Other helpers TODO`
 
 
 👩‍💻 Before You Arrive
@@ -91,54 +91,34 @@ after the workshop has ended.
 1. Presentation: Model Building with OpenSim Creator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Approx 20 min., LINK: `Presentation, Model Building with OpenSim Creator`_
+LINK: `Presentation, Model Building with OpenSim Creator`_
 
-1. Welcome
-2. Quick overview of the workshop plan (above)
-3. Direct people to the documentation/guidebook pages, etc.
-4. Basic OpenSim ecosystem overview (e.g. briefly mentioning OpenCap, Add Bioinformatics, etc.)
-5. (cont.)
-6. Historical/development overview of OpenSim Creator (major feature releases, old screenshots, development methodology, etc.)
-7. (cont.)
-8. (cont.)
-9. Introduce model building concepts/challenges
-10. Introduce/start the work session. Introduce walkthroughs, exercises, etc. Give
-    the attendants something concrete they can do, even if they aren't already
-    working on a model.
-11. (cont.)
-12. Splash slide, call to action
+**TODO**: once everything is basically ready in OneDrive, publish a final(ish)
+version to files.opensimcreator.com, put a screenshot of the first slide here
+and link to the download.
 
 
 2. Work Session: Model Building
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Approx 90 min., LINK: `Work Session, Model Building`_
+LINK: `Work Session, Model Building`_
 
-TODO: provide a list of actions, exercies, etc. here in the
-documentation page. Work session exercises, example models,
-walkthroughs, etc. The 'presentation' for this section is just
-a backdrop we can throw up while people mill around etc.
+**TODO**: once everything is basically ready in OneDrive, publish a final(ish)
+version to files.opensimcreator.com, put a screenshot of the first slide here
+and link to the download.
 
-- Direct people to the tutorials or ESMAC-specific exercises. Maybe shape
-  it such that morphometric (mesh) scaling is an available thing they can
-  work on if they are mostly here for the 2nd half of the workshop - **ensure
-  a warp-able model is made**
-- On-whiteboard walkthroughs about specific features in OpenSim.
-- Staff/team milling around answering questions, and so on.
-- Show gallery page, which unofficially documents new buttons/features, etc.
-- Show live moment arm plotting
-- Edit meshes in Blender, update in OpenSim Creator?
-- Plot muscle curves, show how to edit them in external software?
-- Make a model with a ``StationDefinedFrame``
-
-- Maybe in parallel, or at the end, try to edit/mutate a bigger model with
-  motion data to be warpable in the knee or similar?
[…]
