### 2025-07-22 #8284 @adamkewley
Just FYI for the editor + reviewers (@Kevin-Mattheus-Moerman , @davidpagnon , @modenaxe , @cvhammond )

It's now the holiday and conference season, so I won't really be available to work through and actually implement solutions/changes for people's concerns until mid-August (earliest) or even early-September (after conferences).
