### 2025-07-23 652b231
Work session -> practical session
