+++ docs/source/esmac-2025.rst
@@ -35,7 +35,7 @@ September 2025. The exact time and room is still to-be-decided (:red:`TODO`).
 
 We recommend attendants prepare for the workshop in the following ways:
 
-- **Download and install OpenSim Creator**. The work sessions will focus
+- **Download and install OpenSim Creator**. The practical sessions will focus
   on using OpenSim Creator's features for model building and warping.
   The :ref:`installation` page outlines the procedure.
 - **(optional) Download and install OpenSim**. To answer attendants' specific
@@ -43,12 +43,12 @@ We recommend attendants prepare for the workshop in the following ways:
   can be downloaded from the `OpenSim SimTK Page`_. Setting up a scripting (API)
   environment is described in the `OpenSim Scripting Documentation`_.
 - **(optional) Familiarize Yourself with the basics of OpenSim/modelling**. The
-  work sessions may be easier to follow if you have already gone through an early
+  practical sessions may be easier to follow if you have already gone through an early
   tutorial, such as :doc:`make-a-pendulum`.
 
 **If you are already working on an OpenSim model**, bring it along. The workshop
-is structured such that work sessions' walkthroughs and exercises can *usually* be
-completed with any model. The work sessions are an ideal time to ask specific
+is structured such that practical sessions' walkthroughs and exercises can *usually* be
+completed with any model. The practicals are an ideal time to ask specific
 questions about creating your model.
 
 
@@ -63,17 +63,17 @@ the workshop.
     :width: 80%
 
     A (very 🙃) rough schedule for the workshop. The longest parts will be
-    work sessions in which we'll go through some walkthroughs, specific
+    practical sessions in which we'll go through some walkthroughs, specific
     features, and answer attendants' questions.
 
-The first half of the workshop will be focused on model building. The work
+The first half of the workshop will be focused on model building. The practical
 session will include feature demonstrations by the team, specific model
 building walkthroughs (e.g. building a basic model), and a chance for attendants
 to explore OpenSim Creator or work on their own model with help from the team.
 
 The second half of the workshop will focus on model scaling/warping. We
 will introduce our latest beta workflow, :doc:`the-model-warper`, and explain
-how it can be used to create subject-specific models. The work session will be
+how it can be used to create subject-specific models. The practical session will be
 a chance for attendants to give this workflow a try and provide us with feedback
 about it - this time can also be used to continue exploring OpenSim Creator and
 build models.
@@ -98,10 +98,10 @@ version to files.opensimcreator.com, put a screenshot of the first slide here
 and link to the download.
 
 
-2. Work Session: Model Building
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+2. Practical: Model Building
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-LINK: `Work Session, Model Building`_
+LINK: `Practical Session, Model Building`_
 
[…]
