### 2025-07-23 0400634
Refactor ImGuizmo towards oscar-style design
