+++ liboscar/UI/Detail/ImGuizmo.cpp
@@ -52,6 +52,7 @@
 #include "imgui_internal.h"
 #include "ImGuizmo.h"
 
+#include <liboscar/Maths/Constants.h>
 #include <liboscar/Maths/Mat4.h>
 #include <liboscar/Maths/MatFunctions.h>
 #include <liboscar/Maths/Rect.h>
@@ -66,13 +67,13 @@
 #include <vector>
 
 using namespace ImGuizmo;
+using namespace osc;
 
 namespace
 {
     struct Context;
-    constinit Context* gCurrentContext = nullptr;
+    constinit Context* g_CurrentContext = nullptr;
 
-    constexpr float ZPI = 3.14159265358979323846f;
     constexpr float screenRotateSize = 0.06f;
 
     // scale a bit so translate axis do not touch when in universal
@@ -770,26 +771,26 @@ namespace
       bool mAllowAxisFlip = false;
       float mGizmoSizeClipSpace = 0.1f;
 
-      inline ImGuiID GetCurrentID() { return mIDStack.back();}
+      inline ImGuiID GetCurrentID() const { return mIDStack.back();}
    };
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
    //
-   int GetMoveType(Operation op, vec_t* gizmoHitProportion);
-   int GetRotateType(Operation op);
-   int GetScaleType(Operation op);
+   int GetMoveType(Context&, Operation op, vec_t* gizmoHitProportion);
+   int GetRotateType(const Context&, Operation op);
+   int GetScaleType(Context&, Operation op);
 
-   ImU32 GetColorU32(int idx)
+   ImU32 GetColorU32(const Context& context, int idx)
    {
       IM_ASSERT(idx < COLOR::COUNT);
-      return ImGui::ColorConvertFloat4ToU32(gCurrentContext->mStyle.Colors[idx]);
+      return ImGui::ColorConvertFloat4ToU32(context.mStyle.Colors[idx]);
    }
 
    ImVec2 worldToPos(
        const vec_t& worldPos,
        const matrix_t& mat,
-       ImVec2 position = ImVec2(gCurrentContext->mX, gCurrentContext->mY),
-       ImVec2 size = ImVec2(gCurrentContext->mWidth, gCurrentContext->mHeight))
+       ImVec2 position,
+       ImVec2 size)
    {
       vec_t trans;
       trans.TransformPoint(worldPos, mat);
@@ -803,22 +804,38 @@ namespace
       return ImVec2(trans.x, trans.y);
    }
 
+   ImVec2 worldToPos(
+       const Context& context,
+       const vec_t& worldPos,
+       const matrix_t& mat)
+   {
+       return worldToPos(
+           worldPos,
+           mat,
+           ImVec2(context.mX, context.mY),
+           ImVec2(context.mWidth, context.mHeight)
+       );
+   }
+
    void ComputeCameraRay(
+       const matrix_t& viewMat,
+       const matrix_t& projectionMat,
+       bool reversed,
        vec_t& rayOrigin,
        vec_t& rayDir,
-       ImVec2 position = ImVec2(gCurrentContext->mX, gCurrentContext->mY),
-       ImVec2 size = ImVec2(gCurrentContext->mWidth, gCurrentContext->mHeight))
+       ImVec2 position,
+       ImVec2 size)
    {
       ImGuiIO& io = ImGui::GetIO();
 
       matrix_t mViewProjInverse;
-      mViewProjInverse.Inverse(gCurrentContext->mViewMat * gCurrentContext->mProjectionMat);
+      mViewProjInverse.Inverse(viewMat * projectionMat);
 
[…]
