### 2025-07-25 976b96f
Make main window content scale contribute to 'highest content scale' calculation
