### 2025-08-07 e1e9b1b
Clean up Rect methods etc.
