### 2025-08-07 e4c13c9
Make App default CursorShape to Arrow rather than Forbidden
