### 2025-08-08 692a987
Refactor file format APIs to group formats under a class.
