+++ liboscar/Formats/CSV.tests.cpp
@@ -6,30 +6,30 @@
 
 using namespace osc;
 
-TEST(read_csv_row, reading_an_empty_stream_returns_a_single_empty_column)
+TEST(CSV, read_row_reading_an_empty_stream_returns_a_single_empty_column)
 {
     std::istringstream input;
-    const std::optional<std::vector<std::string>> output = read_csv_row(input);
+    const std::optional<std::vector<std::string>> output = CSV::read_row(input);
 
     ASSERT_TRUE(output.has_value());
     ASSERT_EQ(output->size(), 1);
     ASSERT_EQ(output->at(0), "");
 }
 
-TEST(read_csv_row, reading_a_stream_containing_one_space_returns_a_single_column_containing_the_spaace)
+TEST(CSV, read_row_reading_a_stream_containing_one_space_returns_a_single_column_containing_the_spaace)
 {
     std::istringstream input{" "};
-    const std::optional<std::vector<std::string>> output = read_csv_row(input);
+    const std::optional<std::vector<std::string>> output = CSV::read_row(input);
 
     ASSERT_TRUE(output.has_value());
     ASSERT_EQ(output->size(), 1);
     ASSERT_EQ(output->at(0), " ");
 }
 
-TEST(read_csv_row, reading_a_stream_containing_just_two_commas_should_return_three_empty_columns)
+TEST(CSV, read_row_reading_a_stream_containing_just_two_commas_should_return_three_empty_columns)
 {
     std::istringstream input{",,"};
-    const  std::optional<std::vector<std::string>> output = read_csv_row(input);
+    const  std::optional<std::vector<std::string>> output = CSV::read_row(input);
 
     ASSERT_TRUE(output.has_value());
     ASSERT_EQ(output->size(), 3);
@@ -38,17 +38,17 @@ TEST(read_csv_row, reading_a_stream_containing_just_two_commas_should_return_thr
     }
 }
 
-TEST(read_csv_row, reading_a_stream_containing_standard_column_headers_returns_expected_output)
+TEST(CSV, read_row_reading_a_stream_containing_standard_column_headers_returns_expected_output)
 {
     std::istringstream input{"col1,col2,col3"};
     const std::vector<std::string> expected_output = {"col1", "col2", "col3"};
-    const std::optional<std::vector<std::string>> output = read_csv_row(input);
+    const std::optional<std::vector<std::string>> output = CSV::read_row(input);
 
     ASSERT_TRUE(output.has_value());
     ASSERT_EQ(*output, expected_output);
 }
 
-TEST(read_csv_row, reading_a_stream_containing_multiple_lines_returns_each_row_as_expected)
+TEST(CSV, read_row_reading_a_stream_containing_multiple_lines_returns_each_row_as_expected)
 {
     std::istringstream input{"col1,col2\n1,2\n,\n \n\n"};
     const std::vector<std::vector<std::string>> expected_outputs = {
@@ -60,16 +60,16 @@ TEST(read_csv_row, reading_a_stream_containing_multiple_lines_returns_each_row_a
     };
 
     for (const auto& expected_output : expected_outputs) {
-        const std::optional<std::vector<std::string>> columns = read_csv_row(input);
+        const std::optional<std::vector<std::string>> columns = CSV::read_row(input);
         ASSERT_TRUE(columns.has_value());
         ASSERT_EQ(*columns, expected_output);
     }
 }
 
-TEST(read_csv_row, reading_a_stream_containing_nested_quotes_works_as_expected_for_basic_example)
+TEST(CSV, read_row_reading_a_stream_containing_nested_quotes_works_as_expected_for_basic_example)
 {
[…]
