### 2025-08-08 7efc082
Rename Rect to Ray
