+++ liboscar/Maths/MathsImplementation.cpp
@@ -121,7 +121,7 @@ namespace
     bool bvh_for_each_ray_aabb_collisions_recursive(
         std::span<const BVHNode> nodes,
         std::span<const BVHPrim> prims,
-        const Line& ray,
+        const Ray& ray,
         ptrdiff_t node_index,
         const std::function<void(BVHCollision)>& callback)
     {
@@ -155,7 +155,7 @@ namespace
         std::span<const BVHPrim> prims,
         std::span<const Vec3> vertices,
         std::span<const TIndex> indices,
-        const Line& ray,
+        const Ray& ray,
         float& closest,
         ptrdiff_t node_index)
     {
@@ -255,7 +255,7 @@ namespace
         std::span<const BVHPrim> prims,
         std::span<const Vec3> vertices,
         std::span<const TIndex> indices,
-        const Line& ray)
+        const Ray& ray)
     {
         if (nodes.empty() or prims.empty() or indices.empty()) {
             return std::nullopt;
@@ -315,28 +315,28 @@ void osc::BVH::build_from_indexed_triangles(std::span<const Vec3> vertices, std:
 std::optional<BVHCollision> osc::BVH::closest_ray_indexed_triangle_collision(
     std::span<const Vec3> vertices,
     std::span<const uint16_t> indices,
-    const Line& line) const
+    const Ray& ray) const
 {
     return bvh_get_closest_ray_indexed_triangle_collision<uint16_t>(
         nodes_,
         prims_,
         vertices,
         indices,
-        line
+        ray
     );
 }
 
 std::optional<BVHCollision> osc::BVH::closest_ray_indexed_triangle_collision(
     std::span<const Vec3> vertices,
     std::span<const uint32_t> indices,
-    const Line& line) const
+    const Ray& ray) const
 {
     return bvh_get_closest_ray_indexed_triangle_collision<uint32_t>(
         nodes_,
         prims_,
         vertices,
         indices,
-        line
+        ray
     );
 }
 
@@ -363,7 +363,7 @@ void osc::BVH::build_from_aabbs(std::span<const AABB> aabbs)
 }
 
 void osc::BVH::for_each_ray_aabb_collision(
-    const Line& ray,
+    const Ray& ray,
     const std::function<void(BVHCollision)>& callback) const
 {
     if (nodes_.empty() or prims_.empty()) {
@@ -540,9 +540,9 @@ Mat4 osc::EulerPerspectiveCamera::projection_matrix(float aspect_ratio) const
 }
 
 
-std::ostream& osc::operator<<(std::ostream& out, const Line& line)
+std::ostream& osc::operator<<(std::ostream& out, const Ray& ray)
 {
-    return out << "Line(origin = " << line.origin << ", direction = " << line.direction << ')';
+    return out << "Ray(origin = " << ray.origin << ", direction = " << ray.direction << ')';
 }
 
 
@@ -658,7 +658,7 @@ Vec2 osc::PolarPerspectiveCamera::project_onto_viewport(
     );
 }
 
-Line osc::PolarPerspectiveCamera::unproject_topleft_pos_to_world_ray(Vec2 pos, Vec2 dimensions) const
+Ray osc::PolarPerspectiveCamera::unproject_topleft_pos_to_world_ray(Vec2 pos, Vec2 dimensions) const
 {
     return perspective_unproject_topleft_normalized_pos_to_world(
         pos / dimensions,
@@ -908,11 +908,11 @@ namespace
         return res;
     }
 
-    std::optional<RayCollision> find_collision_analytic(const Sphere& sphere, const Line& line)
+    std::optional<RayCollision> find_collision_analytic(const Sphere& sphere, const Ray& ray)
     {
[…]
