+++ libopensimcreator/Graphics/CachedModelRenderer.h
@@ -13,7 +13,7 @@ namespace osc { class IModelStatePair; }
 namespace osc { class Rect; }
 namespace osc { class RenderTexture; }
 namespace osc { class SceneCache; }
-namespace osc { struct Line; }
+namespace osc { struct Ray; }
 namespace osc { struct ModelRendererParams; }
 namespace osc { struct SceneDecoration; }
 
