### 2025-08-18 13bb15b
Rename Transform::position -> Transform::translation and remove many 'pos' shorthands
