### 2025-08-18 be08c57
Disambiguate 'location' and 'position'
