+++ liboscar/Maths/Ellipsoid.h
@@ -5,7 +5,7 @@
 
 namespace osc
 {
-    // an ellipsoid located + oriented at an origin location
+    // an ellipsoid located + oriented at an origin position
     struct Ellipsoid final {
         friend bool operator==(const Ellipsoid&, const Ellipsoid&) = default;
 
