### 2025-08-18 193942f
Replace osc::PointDirection with osc::Ray
