### 2025-08-18 8aedc78
Drop osc::UnitVec
