### 2025-08-18 da995f5
Update documentation for AABB, ClosedInterval, and Rect
