+++ liboscar/Maths/QuaternionFunctions.h
@@ -71,7 +71,7 @@ namespace osc
     }
 
     template<typename T>
-    Qua<T> quat_cast(const Matrix<3, 3, T>& m)
+    Qua<T> quaternion_cast(const Matrix<3, 3, T>& m)
     {
         const T four_x_squared_minus_1 = m[0][0] - m[1][1] - m[2][2];
         const T four_y_squared_minus_1 = m[1][1] - m[0][0] - m[2][2];
@@ -111,9 +111,9 @@ namespace osc
     }
 
     template<typename T>
-    Qua<T> quat_cast(const Matrix<4, 4, T>& m)
+    Qua<T> quaternion_cast(const Matrix<4, 4, T>& m)
     {
-        return quat_cast(Matrix<3, 3, T>(m));
+        return quaternion_cast(Matrix<3, 3, T>(m));
     }
 
     template<typename T>
@@ -153,7 +153,7 @@ namespace osc
     }
 
     template<typename T>
-    constexpr Qua<T> quat_identity()
+    constexpr Qua<T> quaternion_identity()
     {
         return Qua<T>::wxyz(
             static_cast<T>(1),
@@ -190,7 +190,7 @@ namespace osc
         const T cos_theta = dot(origin, destination);
         if (cos_theta >= static_cast<T>(1) - epsilon_v<T>) {
             // origin and destination point in the same direction
-            return quat_identity<T>();
+            return quaternion_identity<T>();
         }
 
         Vec<3, T> rotation_axis;
