+++ libopensimcreator/Documents/OutputExtractors/OutputExtractorDataType.h
@@ -5,7 +5,7 @@ namespace osc
     // indicates the datatype that an `OutputExtractor` emits at runtime
     enum class OutputExtractorDataType {
         Float,
-        Vec2,
+        Vector2,
         String,
         NUM_OPTIONS
     };
