### 2025-08-19 af0c4e5
Rename matrix ShaderPropertyTypes to match MatrixCxR naming
