+++ liboscar/Graphics/GraphicsImplementation.cpp
@@ -497,8 +497,8 @@ namespace
         case GL_FLOAT_VEC2:   return ShaderPropertyType::Vector2;
         case GL_FLOAT_VEC3:   return ShaderPropertyType::Vector3;
         case GL_FLOAT_VEC4:   return ShaderPropertyType::Vector4;
-        case GL_FLOAT_MAT3:   return ShaderPropertyType::Mat3;
-        case GL_FLOAT_MAT4:   return ShaderPropertyType::Mat4;
+        case GL_FLOAT_MAT3:   return ShaderPropertyType::Matrix3x3;
+        case GL_FLOAT_MAT4:   return ShaderPropertyType::Matrix4x4;
         case GL_INT:          return ShaderPropertyType::Int;
         case GL_BOOL:         return ShaderPropertyType::Bool;
         case GL_SAMPLER_2D:   return ShaderPropertyType::Sampler2D;
@@ -7048,7 +7048,7 @@ void osc::GraphicsBackend::bind_to_instanced_attributes(
 
     size_t byte_offset = 0;
     if (shader_impl.maybe_instanced_model_mat_attr_) {
-        if (shader_impl.maybe_instanced_model_mat_attr_->shader_type == ShaderPropertyType::Mat4) {
+        if (shader_impl.maybe_instanced_model_mat_attr_->shader_type == ShaderPropertyType::Matrix4x4) {
             const gl::AttributeMat4 model_matrix_attr{shader_impl.maybe_instanced_model_mat_attr_->location};
             gl::vertex_attrib_pointer(model_matrix_attr, false, instancing_state.stride, instancing_state.base_offset + byte_offset);
             gl::vertex_attrib_divisor(model_matrix_attr, 1);
@@ -7057,14 +7057,14 @@ void osc::GraphicsBackend::bind_to_instanced_attributes(
         }
     }
     if (shader_impl.maybe_instanced_normal_mat_attr_) {
-        if (shader_impl.maybe_instanced_normal_mat_attr_->shader_type == ShaderPropertyType::Mat4) {
+        if (shader_impl.maybe_instanced_normal_mat_attr_->shader_type == ShaderPropertyType::Matrix4x4) {
             const gl::AttributeMat4 normal_matrix_attr{shader_impl.maybe_instanced_normal_mat_attr_->location};
             gl::vertex_attrib_pointer(normal_matrix_attr, false, instancing_state.stride, instancing_state.base_offset + byte_offset);
             gl::vertex_attrib_divisor(normal_matrix_attr, 1);
             gl::enable_vertex_attrib_array(normal_matrix_attr);
             // unused: byteOffset += sizeof(float) * 16;
         }
-        else if (shader_impl.maybe_instanced_normal_mat_attr_->shader_type == ShaderPropertyType::Mat3) {
+        else if (shader_impl.maybe_instanced_normal_mat_attr_->shader_type == ShaderPropertyType::Matrix3x3) {
             const gl::AttributeMat3 normal_matrix_attr{shader_impl.maybe_instanced_normal_mat_attr_->location};
             gl::vertex_attrib_pointer(normal_matrix_attr, false, instancing_state.stride, instancing_state.base_offset + byte_offset);
             gl::vertex_attrib_divisor(normal_matrix_attr, 1);
@@ -7080,17 +7080,17 @@ void osc::GraphicsBackend::unbind_from_instanced_attributes(
     InstancingState&)
 {
     if (shader_impl.maybe_instanced_model_mat_attr_) {
-        if (shader_impl.maybe_instanced_model_mat_attr_->shader_type == ShaderPropertyType::Mat4) {
[…]
