+++ liboscar/Graphics/ShaderPropertyType.h
@@ -11,8 +11,8 @@ namespace osc
         Vector2,
         Vector3,
         Vector4,
-        Mat3,
-        Mat4,
+        Matrix3x3,
+        Matrix4x4,
         Int,
         Bool,
         Sampler2D,
