+++ liboscar/Graphics/Detail/ShaderPropertyTypeList.h
@@ -10,8 +10,8 @@ namespace osc::detail
         ShaderPropertyType::Vector2,
         ShaderPropertyType::Vector3,
         ShaderPropertyType::Vector4,
-        ShaderPropertyType::Mat3,
-        ShaderPropertyType::Mat4,
+        ShaderPropertyType::Matrix3x3,
+        ShaderPropertyType::Matrix4x4,
         ShaderPropertyType::Int,
         ShaderPropertyType::Bool,
         ShaderPropertyType::Sampler2D,
