### 2025-08-20 b0e81db
Rename Vec template base to Vector
