+++ liboscar/Maths/Vector3.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <liboscar/Maths/Scalar.h>
-#include <liboscar/Maths/Vec.h>
+#include <liboscar/Maths/Vector.h>
 
 #include <concepts>
 #include <cstddef>
@@ -11,7 +11,7 @@
 namespace osc
 {
     template<ScalarOrBoolean T>
-    struct Vec<3, T> {
+    struct Vector<3, T> {
         using value_type = T;
         using element_type = T;
         using size_type = size_t;
@@ -23,13 +23,13 @@ namespace osc
         using iterator = T*;
         using const_iterator = const T*;
 
-        constexpr Vec() = default;
-        explicit constexpr Vec(T value) :
+        constexpr Vector() = default;
+        explicit constexpr Vector(T value) :
             x{value},
             y{value},
             z{value}
         {}
-        constexpr Vec(T x_, T y_, T z_) :
+        constexpr Vector(T x_, T y_, T z_) :
             x{x_},
             y{y_},
             z{z_}
@@ -37,7 +37,7 @@ namespace osc
         template<typename X, typename Y, typename Z>
         requires std::constructible_from<T, X> and std::constructible_from<T, Y> and std::constructible_from<T, Z>
         explicit (not (std::convertible_to<T, X> and std::convertible_to<T, Y> and std::convertible_to<T, Z>))
-        constexpr Vec(X x_, Y y_, Z z_) :
+        constexpr Vector(X x_, Y y_, Z z_) :
             x{static_cast<T>(x_)},
             y{static_cast<T>(y_)},
             z{static_cast<T>(z_)}
@@ -45,7 +45,7 @@ namespace osc
         template<ScalarOrBoolean A, typename B>
         requires std::constructible_from<T, A> and std::constructible_from<T, B>
         explicit (not (std::convertible_to<T, A> and std::convertible_to<T, B>))
-        constexpr Vec(const Vec<2, A>& xy_, B z_) :
+        constexpr Vector(const Vector<2, A>& xy_, B z_) :
             x{static_cast<T>(xy_.x)},
             y{static_cast<T>(xy_.y)},
             z{static_cast<T>(z_)}
@@ -53,14 +53,14 @@ namespace osc
         template<typename A, ScalarOrBoolean B>
         requires std::constructible_from<T, A> and std::constructible_from<T, B>
         explicit (not (std::convertible_to<T, A> and std::convertible_to<T, B>))
-        constexpr Vec(A x_, const Vec<2, B>& yz_) :
+        constexpr Vector(A x_, const Vector<2, B>& yz_) :
             x{static_cast<T>(x_)},
             y{static_cast<T>(yz_.y)},
             z{static_cast<T>(yz_.z)}
         {}
         template<ScalarOrBoolean U>
         requires std::constructible_from<T, U>
-        explicit constexpr Vec(const Vec<4, U>& v) :
+        explicit constexpr Vector(const Vector<4, U>& v) :
             x{static_cast<T>(v.x)},
             y{static_cast<T>(v.y)},
             z{static_cast<T>(v.z)}
@@ -68,7 +68,7 @@ namespace osc
         template<ScalarOrBoolean U>
         requires std::constructible_from<T, U>
         explicit (not (std::convertible_to<T, U>))
-        constexpr Vec(const Vec<3, U>& v) :
+        constexpr Vector(const Vector<3, U>& v) :
             x{static_cast<T>(v.x)},
             y{static_cast<T>(v.y)},
             z{static_cast<T>(v.z)}
[…]
