+++ libopensimcreator/Utils/LandmarkPair3D.h
@@ -14,8 +14,8 @@ namespace osc
     struct LandmarkPair3D {
         friend bool operator==(const LandmarkPair3D&, const LandmarkPair3D&) = default;
 
-        Vec<3, T> source{};
-        Vec<3, T> destination{};
+        Vector<3, T> source{};
+        Vector<3, T> destination{};
     };
 
     std::ostream& operator<<(std::ostream&, const LandmarkPair3D<float>&);
