+++ liboscar/Maths/CoordinateDirection.h
@@ -84,12 +84,12 @@ namespace osc
             return static_cast<T>(direction_);
         }
 
-        // returns a direction that points in the direction stored by this `CoordinateDirection`
+        // returns a unit-length direction vector that points in the direction stored by this `CoordinateDirection`
         template<typename T = float>
         requires std::is_arithmetic_v<T>
-        constexpr Vec<3, T> vec() const
+        constexpr Vector<3, T> direction_vector() const
         {
-            return Vec<3, T>{}.with_element(axis().index(), direction<T>());
+            return Vector<3, T>{}.with_element(axis().index(), direction<T>());
         }
 
         // returns a `CoordinateDirection` that points along the same `CoordinateAxis`, but with its direction negated
