### 2025-08-20 #8284 @adamkewley
Hi, just as a note: I'm back but in a limited capacity until early-/mid-September.

---

@modenaxe thanks for reviewing the paper, and apologies for the delay in the response.

I agree with most of the points you have made. My only push back is that I think it's very important to keep the intent of this paper (a software announcement for an already-established open-source software project in JOSS) in mind. The paper isn't intended to make any new scientific or research claims (I would argue that's the users' jobs), so when we claim something like "users prefer this" it should be seen via the lens of "a qualitative claim by the software developers" rather than "a scientific claim that requires extensive evidence". Concrete examples of users stating claims like that are available in this review issue, even.

---

> I feel that the state of the field is not very well represented in the paper.

👍 Agreed: other reviewers have said the same, so it clearly needs to be addressed.

> [...] In summary: can you clarify the software scope, i.e. what a user is meant to do in this software environment and what they are supposed to do in OpenSim in your workflow design?

👍 I'll try and list some use-cases, but it's hard to write up 100 % of the available features.

To directly address your questions/concerns, the intent of OSC is to supply an integrated environment for model development, which requires developing hundreds of little features that might be useful for user-specific model building workflows. For example, the simulation support is there because some users required running a little forward dynamic simulation on their model to ensure it's balanced/stable. Other users don't need that, but do need the ability to edit muscle curves.

All of this is to say that workflows for building OpenSim models are extremely varied and user-specific, which makes designing an integrated platform challenging. I posit that the clunkiness you have detected is because OSC doesn't cover 100 % of your requirements yet. The two (opposed) solutions to that problem are to keep expanding the scope of OSC until it achieves better coverage, or to curate an ecosystem of separate scripts/libraries/UIs that, when combined, provide better coverage. Both solutions have benefits/pitfalls.

> Explicitly including an educational use in the statement of need, e.g. having students practicing building models and running basic simulations in the classroom as an additional target, would improve the consistency with the JOSS requirements in my opinion.

👍 Good suggestion, we've already been using it in undergraduate classes for several years so it's not too difficult to mention it.

> One of the main issues of other existing modelling options like NMSBuilder is that it is designed to “manually” build each model component separately [...]. For that kind of operation, the programmatic OpenSim APIs have always been more practical in my experience. Does OCC have some functionality to overcome this limitation?

I agree that programmatic APIs are the best way for a user to automatically define a model. A big reason why OSC is designed to be compatible with OpenSim (APIs, scripts, GUI) is so that users can choose a programmatic solution for parts of their specific model building workflow. It would be practically impossible for OSC to provide a UI that can act as a 100 % replacement for automated model building (to do so would require a Turing-complete scripting/node system in the UI).

> More generally, can you realistically use OCC for building “large scale” models in reasonable operator time? [...] Could OCC streamline a workflow of that type or is it designed more for adapting/morphing existing models?

The OSC project is currently midway through developing a specific workflow for adapting/morphing existing models ([The Model Warper](https://docs.opensimcreator.com/manual/en/latest/the-model-warper.html)) but that development has been very slow and full of rewrites because scaling/warping models is extremely user-specific and requires a large degree of flexibility that's hard to deliver via a UI. Every time I ship a model warping beta, a tester (end-user) inevitably asks for a new scaling trick that they would like the UI to support. It might be the case that scaling/warping/fitting is so user-specific and full of heuristics that it would be better to deliver it as a programming API (another, future, project).

"Reasonable operator time" is hard to define. A couple of hours of operator time might be reasonable if the alternative is a multi-year automated engine development - especially if there isn't a particularly large volume of models to make in the first place. Automating the process is desirable, but not always necessary.

> [Rapid releases] might be not ideal for users though, because they will have to manually update the software several times per year if they want to use the latest version. Have you considered including a mechanism for updating the software within the application?

We have considered it, and it's a good suggestion. The only reason it hasn't been done yet is because it will take some time + resources (hosting) to ensure it is robust on all three supported OSes. Different OSes have different constraints w.r.t. auto-updating native binaries.

Given the installation procedure is fairly short (<5 minutes), and users don't *have* to update OSC (you can even install [0.0.2](https://files.opensimcreator.com/releases/osmv-0.0.2-windows-amd64.exe), if you want a laugh), developing auto-updates is considered low-priority compared to other features at the moment.

> The use-cases are nice examples, and work well as an high level description, but it would be nice to have some additional details for potential users familiar with OpenSim.

It would be useful to have an explanation for every example model, but I don't think it's OSC's job to deliver that explanation as part of this paper announcement. It should be a feature provided by OpenSim (e.g. the description fields etc.) that OSC, and users, actually use. It could also be provided via a platform (e.g. a jacked-up SimTK).

Almost all of the provided example models are directly taken from [opensim-models](https://github.com/opensim-org/opensim-models). They are usually related to some sort of academic research paper, which does explain the model-building process. The fundamental issue you're (correctly!) encountering is that model files are usually separate from their associated SimTK pages, which are usually separate from academic paper hosts, which are usually separate from the underlying paper PDF, which is where the explanation you want is usually hiding.

> Could the second example focusing on model simplification while preserving moment arms not also be performed in OpenSim, through the muscle editing functionalities of their UI/visualiser? I am trying to identify the specific advantage in using OCC for this task: is it the real-time feedback and visualisation of the moment arms?

It's the real-time feedback, combined with the overlay of external data. While you can *theoretically* do the same job in OpenSim GUI via the UI/visualizer, it will take significantly more time and not include some extremely useful features (e.g. historical curve plotting, undo/redo to a specific curve).

A major part of why users prefer OSC to OpenSim GUI for some tasks isn't because it does something new, but because it does something old in a more robust/intuitive way.

> Claiming quick model creation is certainly a major achievement and appeal of a software a scope like OCC, but the sentence at line 94-96 in my opinion needs more evidence. [...] What were these users trying to do? How complex was the model?

Previous techniques include using the OpenSim GUI to edit model properties, writing XML, and running MATLAB scripts to build parts of a model. The users span a wide variety of use-cases and genuinely report an improved quality of life when using OSC. As examples, some users are complex robotic models of hands from scratch, while other users are undergraduates that are new to OpenSim and are editing a couple of muscles in an existing model.

It must be noted that this paper is a JOSS article announcing OSC, not a rigorous scientific study of usability. The claim on line 94-96 is accurate: OSC's users generally report what we have written. The claim is not intended to be a robust usability experiment in which we time a sample of users performing a rigidly-defined task (we do not claim this). Providing a metric for usability is extremely difficult.

> Similarly, “some researchers already have” used the Model Warper (from the documentation of the tool)…it would be nice to know who they are and what they have produced, so the reader can learn from that.

Those users are in TU Delft and are responsible for their own, currently unpublished, research based on the (very experimental) model warper. Right above the thing you are citing (in the same box, even) it is made very clear that **The model warper is experimental**. Eventually, when it is non-experimental, there may be things to cite, and that documentation may be updated when that's the case.

> An additional comment: although I understand why you are comparing OCC to a CAD software for building MSK models, I think some limitations, like the inability of deleting an arbitrary element of the model, or performing actual CAD operations on the meshes (cut, boolean etc), make it currently a bit stretched.  I am aware of the technical difficulties related to some of these limitations, but I would clarify the limits of this comparison to set the expectations right, especially for mechanical engineers that might expect tools like those available in Solidworks. in terms of functionalities or absence of limitations.

:+1: I'll try and outline some of the limitations in the article. Obviously, if users are expecting OSC (5 years, one developer) to compare to Solidworks (30 years, ~25-200 developers) as an actual CAD replacement then they might be disappointed!
