### 2025-08-21 141bfa9
Fix cancelling out of a file dialog in Ubuntu emits an empty path.
