### 2025-08-21 1a6984f
Add link to first practical materials
