+++ docs/source/make-a-knee.rst
@@ -90,3 +90,62 @@ TODO: explain adding a ``StationDefinedFrame`` to the femur condyls (repetition)
 
 Add a Tibia Body
 ----------------
+
+TODO: explain adding a tibia body associated with the tibia mesh, can initially be
+attached to ground if it makes the initial landmark/SDF setup easier.
+
+
+Add Tibia Landmarks
+-------------------
+
+TODO: as above, import the landmarks somehow (even if it literally means just adding them
+manually and typing in the relevant coordinates).
+
+
+Add a StationDefinedFrame to the Tibia Head
+-------------------------------------------
+
+TODO: as above, ensure there's a frame on the tibial head that makes sense for the joint
+
+
+Join the Tibia head SDF to the Femur Condyl SDF
+-----------------------------------------------
+
+TODO: Mess around with the joint (e.g. via socket editor) to rejoin the existing placeholder joint
+(e.g. between tibia and ground) to the femur
+
+
+Add a Muscle Between the Femur and the Tibia
+--------------------------------------------
+
+TODO: describe adding relevant landmarks/stations or whatever is necessary in order to
+create a muscle that has one point on the femur and one point on the tibia. Point out that
+it's going to look kind of shit initially because it will take the shortest path between the
+two (i.e. it'll clip through the meshes etc.)
+
+
+Add a Knee Wrap Cylinder Wrap Surface
+-------------------------------------
+
+TODO: describe adding a knee wrap surface to the relevant body (accessible via the UI by
+right-clicking the relevant body -> Add -> WrapSurface or similar).
+
+
+Associate the Muscle with the Wrap Surface
+------------------------------------------
+
+TODO: describe why wrap surfaces are separate from their interaction set (it makes logical sense
+but users routinely are confused by this extra step). Associated by right-clicking muscle
+and following the relevant menu option
+
+BONUS: Add Markers Ready for IK or similar
+------------------------------------------
+
+TODO: describe adding relevant markers, pretend this is how you'd link the model to
+lab measurements.
+
+ESMAC2025 Notes: What's Missing Here
+------------------------------------
+
+This tutorial isn't going to cover the mesh warper, Thin-Plate Spline, or warping because
+that's covered in the subsequent part of the workshop by :doc:`the-model-warper`.
