### 2025-08-21 0da1fab
Add ability to select which frame points are expressed in (#791)
