### 2025-08-22 9bbe20d
Fix OOB error when manually navigating the undo/redo history (#1083)
