### 2025-08-22 38c6d5f
Add manipulator gizmos to the mesh warping workflow (#1084)
