+++ libopensimcreator/UI/MeshWarper/MeshWarpingTabUserSelection.h
@@ -10,26 +10,13 @@ namespace osc
     // the user's current selection
     struct MeshWaringTabUserSelection final {
 
-        void clear()
-        {
-            m_SelectedSceneElements.clear();
-        }
+        size_t size() const { return m_SelectedSceneElements.size(); }
+        auto begin() const { return m_SelectedSceneElements.begin(); }
+        auto end() const { return m_SelectedSceneElements.end(); }
 
-        void select(TPSDocumentElementID el)
-        {
-            m_SelectedSceneElements.insert(std::move(el));
-        }
-
-        bool contains(const TPSDocumentElementID& el) const
-        {
-            return m_SelectedSceneElements.contains(el);
-        }
-
-        const std::unordered_set<TPSDocumentElementID>& getUnderlyingSet() const
-        {
-            return m_SelectedSceneElements;
-        }
-
+        void clear() { m_SelectedSceneElements.clear(); }
+        void select(TPSDocumentElementID el) { m_SelectedSceneElements.insert(std::move(el)); }
+        bool contains(const TPSDocumentElementID& el) const { return m_SelectedSceneElements.contains(el); }
     private:
         std::unordered_set<TPSDocumentElementID> m_SelectedSceneElements;
     };
