+++ CHANGELOG.md
@@ -30,6 +30,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The mesh warping UI now supports manually moving landmarks around using a UI gizmo,
   similar to how things can be moved in the mesh importer and model editor workflows
   (#1084).
+- The mesh warper UI now correctly rim-highlights selected landmarks on large
+  meshes (#1085).
 
 ## [0.5.25] - 2025/07/14
 
