### 2025-08-22 1325f8f
Add missing switch default handlers
