+++ docs/source/make-a-knee.rst
@@ -10,7 +10,10 @@ Make a Knee
 
 In this tutorial, we will be making a basic model of a knee using OpenSim Creator:
 
-:red:`TODO`: screenshot of the knee
+.. figure:: _static/make-a-bouncing-block/constraints-added.jpg
+    :width: 60%
+
+    TODO: screenshot of the final knee model
 
 This tutorial will primarily use the model editor workflow to build a new model that
 contains some of the steps/components necessary to build a human model. In essence, the
@@ -50,69 +53,133 @@ In order to follow this tutorial, you will need to download the associated
 resources :download:`download here <_static/the-model-warper/walkthrough-model.zip>`.
 
 
-Top-Level Process explanation
------------------------------
-
-TODO: briefly give readers an overview of what we want to achieve and how an opensim
-model is typically built (bodies, joints, muscles, etc.).
-
-
 Create a New Model
 ------------------
 
-TODO: explain creating a blank model from the home screen etc.
+Create a blank model from the home screen (explained in :ref:`create-new-model`).
 
+.. _add-femur-body:
 
 Add a Femur Body
 ----------------
 
-TODO: explain adding a femur body to the new model. Associate the body with
-the femur mesh.
+Add a femur body with the femur mesh attached to it to the model. For this model,
+use the following parameters:
 
+.. figure:: _static/make-a-knee/add-femur-body.jpeg
+    :width: 60%
 
-Add Femur Landmarks
--------------------
+    Create a body called ``femur`` and attach the ``femur_r.vtp`` geometry to it. The
+    mass and intertia can be handled later. The ``femur`` should initially be joined
+    to ``ground`` (the knee joint comes later in the process).
 
-TODO: explain adding markers for the femur to the model as stations/markers.
+Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
+:ref:`create-the-foot`.
 
+.. _import-femur-landmarks:
 
-Add a StationDefinedFrame for the Femur Head
---------------------------------------------
+Import Femur Landmarks
+----------------------
 
-TODO: explain adding a ``StationDefinedFrame`` to the femur head
+This model will use a landmark-defined approach to defining the knee frame (explained
+in :doc:`station-defined-frames`). To do that, we'll need landmarks on the femur that
+correspond to the points that the ISB REF TODO use to define the knee coordinate system.
 
+You can use the point importer in the model editor from the top menu bar, located at ``Tools`` ->
+``Import Points``. It will show a popup that you can use to import the source femur knee
+landmarks file (``femur_r_knee-frame.landmarks.csv``) as markers that are attached to
+the ``femur`` body:
 
-Add a StationDefinedFrame to the Femur Condyls
-----------------------------------------------
+.. figure:: _static/make-a-knee/import-femur-landmarks.jpeg
+    :width: 60%
 
-TODO: explain adding a ``StationDefinedFrame`` to the femur condyls (repetition)
+    The ``Import Points`` dialog, with ``femur_r_knee-frame.landmarks.csv``. Make sure to
+    select ``femur`` as the body to attach the landmarks to. Otherwise, they will end up
+    attached to ``ground``.
+
+.. _add-sdfs-to-femur-condyls:
+
+Add a StationDefinedFrame for the Femur Condyls
+-----------------------------------------------
+
+Now that the appropriate landmarks are imported into the model, you can now add a
+``StationDefinedFrame`` to ``femur`` that computes the knee's coordinate system
+(frame) from those landmarks.
+
+To do that, you'll need to right-click the ``femur`` body and then ``Add`` a
+``StationDefinedFrame`` component to it that connects to the appropriate (imported)
+markers:
[…]
