### 2025-08-23 c230e06
Fix device_pixel_ratio for custom render target being ignored when scaling scissor rect
