### 2025-08-24 338c6bd
Make conversion explicit in RenderTextureParams initializer
