### 2025-08-25 93310d8
minor fixes and clarifications
