### 2025-08-25 7aee5cc
minor fixes and clarifications
