### 2025-08-25 7e33cc4
Add caption to opening figure in make-a-knee
