### 2025-08-25 0743802
Make mesh import errors into a log error (#1088)
