### 2025-08-26 55a6541
Reorder Vector template arguments to match std::array/std::span
