+++ liboscar/Maths/Vector.h
@@ -11,13 +11,13 @@
 namespace osc
 {
     // specialized by `Vector2`, `Vector3`, etc.
-    template<size_t L, Scalar T>
+    template<Scalar T, size_t N>
     struct Vector;
 
-    template<size_t L, Scalar T>
-    std::ostream& operator<<(std::ostream& out, const Vector<L, T>& vec)
+    template<Scalar T, size_t N>
+    std::ostream& operator<<(std::ostream& out, const Vector<T, N>& vec)
     {
-        out << "Vector" << L << '(';
+        out << "Vector" << N << '(';
         std::string_view delimiter;
         for (const T& el : vec) {
             out << delimiter << el;
@@ -29,32 +29,32 @@ namespace osc
 
     // when handled as a tuple-like object, a `Vec` decomposes into its elements
 
-    template<size_t I, size_t L, Scalar T>
-    constexpr const T& get(const Vector<L, T>& vec) { return vec[I]; }
+    template<size_t I, Scalar T, size_t N>
+    constexpr const T& get(const Vector<T, N>& vec) { return vec[I]; }
 
-    template<size_t I, size_t L, Scalar T>
-    constexpr T& get(Vector<L, T>& vec) { return vec[I]; }
+    template<size_t I, Scalar T, size_t N>
+    constexpr T& get(Vector<T, N>& vec) { return vec[I]; }
 
-    template<size_t I, size_t L, Scalar T>
-    constexpr T&& get(Vector<L, T>&& vec) { return std::move(vec[I]); }
+    template<size_t I, Scalar T, size_t N>
+    constexpr T&& get(Vector<T, N>&& vec) { return std::move(vec[I]); }
 
-    template<size_t I, size_t L, Scalar T>
-    constexpr const T&& get(const Vector<L, T>&& vec) { return std::move(vec[I]); }
+    template<size_t I, Scalar T, size_t N>
+    constexpr const T&& get(const Vector<T, N>&& vec) { return std::move(vec[I]); }
 }
 
-template<size_t L, osc::Scalar T>
-struct std::tuple_size<osc::Vector<L, T>> {
-    static inline constexpr size_t value = L;
+template<osc::Scalar T, size_t N>
+struct std::tuple_size<osc::Vector<T, N>> {
+    static inline constexpr size_t value = N;
 };
 
-template<size_t I, size_t L, osc::Scalar T>
-struct std::tuple_element<I, osc::Vector<L, T>> {
+template<size_t I, osc::Scalar T, size_t N>
+struct std::tuple_element<I, osc::Vector<T, N>> {
     using type = T;
 };
 
-template<size_t L, osc::Scalar T>
-struct std::hash<osc::Vector<L, T>> final {
-    size_t operator()(const osc::Vector<L, T>& vec) const noexcept(noexcept(osc::hash_range(vec)))
+template<osc::Scalar T, size_t N>
+struct std::hash<osc::Vector<T, N>> final {
+    size_t operator()(const osc::Vector<T, N>& vec) const noexcept(noexcept(osc::hash_range(vec)))
     {
         return osc::hash_range(vec);
     }
