+++ liboscar/Maths/Matrix3x3.h
@@ -10,11 +10,11 @@ namespace osc
 {
     // a 3x3 column-major matrix
     template<typename T>
-    struct Matrix<3, 3, T> {
+    struct Matrix<T, 3, 3> {
         using col_type = Vector<T, 3>;
         using row_type = Vector<T, 3>;
-        using transpose_type = Matrix<3, 3, T>;
-        using type = Matrix<3, 3, T>;
+        using transpose_type = Matrix<T, 3, 3>;
+        using type = Matrix<T, 3, 3>;
         using value_type = col_type;
         using element_type = T;
         using size_type = size_t;
@@ -86,7 +86,7 @@ namespace osc
         {}
 
         template<typename U>
-        explicit constexpr Matrix(const Matrix<3, 3, U>& m) :
+        explicit constexpr Matrix(const Matrix<U, 3, 3>& m) :
             value{
                 col_type{m[0]},
                 col_type{m[1]},
@@ -94,7 +94,7 @@ namespace osc
             }
         {}
 
-        explicit constexpr Matrix(const Matrix<4, 4, T>& m) :
+        explicit constexpr Matrix(const Matrix<T, 4, 4>& m) :
             value{
                 col_type{m[0]},
                 col_type{m[1]},
@@ -103,7 +103,7 @@ namespace osc
         {}
 
         template<typename U>
-        Matrix<3, 3, T>& operator=(const Matrix<3, 3, U>& m)
+        Matrix<T, 3, 3>& operator=(const Matrix<U, 3, 3>& m)
         {
             this->value[0] = m[0];
             this->value[1] = m[1];
@@ -124,7 +124,7 @@ namespace osc
         friend constexpr bool operator==(const Matrix&, const Matrix&) = default;
 
         template<typename U>
-        Matrix<3, 3, T>& operator+=(U s)
+        Matrix<T, 3, 3>& operator+=(U s)
         {
             this->value[0] += s;
             this->value[1] += s;
@@ -133,7 +133,7 @@ namespace osc
         }
 
         template<typename U>
-        Matrix<3, 3, T>& operator+=(const Matrix<3, 3, U>& m)
+        Matrix<T, 3, 3>& operator+=(const Matrix<U, 3, 3>& m)
         {
             this->value[0] += m[0];
             this->value[1] += m[1];
@@ -142,7 +142,7 @@ namespace osc
         }
 
         template<typename U>
-        Matrix<3, 3, T>& operator-=(U s)
+        Matrix<T, 3, 3>& operator-=(U s)
         {
             this->value[0] -= s;
             this->value[1] -= s;
@@ -151,7 +151,7 @@ namespace osc
         }
 
         template<typename U>
-        Matrix<3, 3, T>& operator-=(const Matrix<3, 3, U>& m)
+        Matrix<T, 3, 3>& operator-=(const Matrix<U, 3, 3>& m)
         {
             this->value[0] -= m[0];
             this->value[1] -= m[1];
@@ -160,7 +160,7 @@ namespace osc
         }
 
         template<typename U>
-        Matrix<3, 3, T>& operator*=(U s)
+        Matrix<T, 3, 3>& operator*=(U s)
         {
             this->value[0] *= s;
             this->value[1] *= s;
@@ -169,13 +169,13 @@ namespace osc
         }
[…]
