### 2025-08-27 2baefa6
Clean up Matrix type aliases (e.g. col_type -> column_type)
