+++ liboscar/Maths/Matrix4x4.h
@@ -11,29 +11,27 @@ namespace osc
     // a 4x4 column-major matrix
     template<typename T>
     struct Matrix<T, 4, 4> {
-        using col_type = Vector<T, 4>;
+        using column_type = Vector<T, 4>;
         using row_type = Vector<T, 4>;
-        using transpose_type = Matrix<T, 4, 4>;
-        using type = Matrix<T, 4, 4>;
-        using value_type = col_type;
+        using value_type = column_type;
         using element_type = T;
         using size_type = size_t;
         using difference_type = ptrdiff_t;
-        using reference = col_type&;
-        using const_reference = const col_type&;
-        using pointer = col_type*;
-        using const_pointer = const col_type*;
-        using iterator = col_type*;
-        using const_iterator = const col_type*;
+        using reference = column_type&;
+        using const_reference = const column_type&;
+        using pointer = column_type*;
+        using const_pointer = const column_type*;
+        using iterator = column_type*;
+        using const_iterator = const column_type*;
 
         constexpr Matrix() = default;
 
         explicit constexpr Matrix(T s) :
             value{
-                col_type{s, T{}, T{}, T{}},
-                col_type{T{}, s, T{}, T{}},
-                col_type{T{}, T{}, s, T{}},
-                col_type{T{}, T{}, T{}, s}
+                column_type{s, T{}, T{}, T{}},
+                column_type{T{}, s, T{}, T{}},
+                column_type{T{}, T{}, s, T{}},
+                column_type{T{}, T{}, T{}, s}
             }
         {}
 
@@ -44,18 +42,18 @@ namespace osc
             const T& x3, const T& y3, const T& z3, const T& w3) :
 
             value{
-                col_type{x0, y0, z0, w0},
-                col_type{x1, y1, z1, w1},
-                col_type{x2, y2, z2, w2},
-                col_type{x3, y3, z3, w3}
+                column_type{x0, y0, z0, w0},
+                column_type{x1, y1, z1, w1},
+                column_type{x2, y2, z2, w2},
+                column_type{x3, y3, z3, w3}
             }
         {}
 
         constexpr Matrix(
-            const col_type& v0,
-            const col_type& v1,
-            const col_type& v2,
-            const col_type& v3) :
+            const column_type& v0,
+            const column_type& v1,
+            const column_type& v2,
+            const column_type& v3) :
 
             value{v0, v1, v2, v3}
         {}
@@ -72,10 +70,10 @@ namespace osc
             const X3& x3, const Y3& y3, const Z3& z3, const W3& w3) :
 
             value{
-                col_type{x0, y0, z0, w0},
-                col_type{x1, y1, z1, w1},
-                col_type{x2, y2, z2, w2},
-                col_type{x3, y3, z3, w3}
+                column_type{x0, y0, z0, w0},
+                column_type{x1, y1, z1, w1},
+                column_type{x2, y2, z2, w2},
+                column_type{x3, y3, z3, w3}
             }
         {}
 
@@ -87,19 +85,19 @@ namespace osc
             const Vector<V4, 4>& v4) :
 
             value{
-                col_type{v1},
-                col_type{v2},
-                col_type{v3},
-                col_type{v4}
+                column_type{v1},
+                column_type{v2},
+                column_type{v3},
+                column_type{v4}
             }
         {}
 
[…]
