### 2025-08-27 5ed7aaf
Rename ImGuizmo namespace to osc::ui::gizmo::detail to reflect its place
