### 2025-08-27 13841e1
Add search bar to add body modal (#21)
