### 2025-08-28 58ce707
Add support for visualizing inbound connections (#1090)
