### 2025-08-28 4fbfe19
Fix minor comment lint issue found by gcc -pedantic
