+++ libopensimcreator/UI/ModelWarper/ModelWarperTab.cpp
@@ -1104,20 +1104,61 @@ Uses the Thin-Plate Spline (TPS) warping algorithm to scale `WrapCylinder`s in t
     };
 
     // A `ScalingStep` that scales the masses of bodies in the model.
-    class BodyMassesScalingStep final : public ScalingStep {
-        OpenSim_DECLARE_CONCRETE_OBJECT(BodyMassesScalingStep, ScalingStep)
+    class ModelMassToSubjectMassScalingStep final : public ScalingStep {
+        OpenSim_DECLARE_CONCRETE_OBJECT(ModelMassToSubjectMassScalingStep, ScalingStep)
     public:
-        explicit BodyMassesScalingStep() :
-            ScalingStep{"TODO: Scale Body Masses to Subject Mass"}
+        explicit ModelMassToSubjectMassScalingStep() :
+            ScalingStep{"Scale Model Mass to Subject Mass"}
         {
-            setDescription("Scales the masses of bodies in the model to match the subject's mass");
+            setDescription("Scales the masses of bodies in the model to match the subject mass, while preserving the overall mass disribution of each body in the model.");
         }
 
     private:
+        std::vector<ScalingStepValidationMessage> implValidate(
+            ScalingCache&,
+            const ScalingParameters& parameters,
+            const OpenSim::Model& model) const final
+        {
+            std::vector<ScalingStepValidationMessage> messages;
+
+            if (parameters.lookup<double>("subject_mass").value_or(0.0) <= 0.0) {
+                std::stringstream msg;
+                msg << "The subject_mass scaling parameter must be greater than zero";
+                messages.emplace_back(ScalingStepValidationState::Error, std::move(msg).str());
+            }
+
+            if (model.getTotalMass(model.getWorkingState()) <= 0.0) {
+                std::stringstream msg;
+                msg << "Cannot scale the model's mass to the subject's mass because the model itself has a mass of zero";
+                messages.emplace_back(ScalingStepValidationState::Error, std::move(msg).str());
+            }
+
+            return messages;
+        }
+
         void implForEachScalingParameterDeclaration(const std::function<void(const ScalingParameterDeclaration&)>& callback) const final
         {
             callback(ScalingParameterDeclaration{"blending_factor", 1.0});
             callback(ScalingParameterDeclaration{"subject_mass", 75.0});
+        }
+
+        void implApplyScalingStep(
+            ScalingCache&,
+            const ScalingParameters& parameters,
+            const OpenSim::Model&,
+            OpenSim::Model& resultModel) const final
+        {
+            const auto subjectMass = parameters.lookup<double>("subject_mass").value_or(0.0);
+            OSC_ASSERT_ALWAYS(subjectMass > 0.0 && "Subject mass must be greater than zero");
+
+            const std::optional<double> blendingFactor = parameters.lookup<double>("blending_factor");
+            OSC_ASSERT_ALWAYS(blendingFactor && "blending_factor was not set by the warping engine");
+
+            const auto effectiveTargetMass = lerp(resultModel.getTotalMass(resultModel.getWorkingState()), subjectMass, *blendingFactor);
+
+            ScaleModelMassPreserveMassDistribution(resultModel, resultModel.getWorkingState(), effectiveTargetMass);
+            InitializeModel(resultModel);
+            InitializeState(resultModel);
         }
     };
 
@@ -1283,7 +1324,7 @@ Uses the Thin-Plate Spline (TPS) warping algorithm to scale `WrapCylinder`s in t
         ThinPlateSplineOffsetFrameTranslationScalingStep,
         ThinPlateSplineMeshSubstitutionScalingStep,
         ThinPlateSplineWrapCylinderScalingStep,
-        BodyMassesScalingStep,
+        ModelMassToSubjectMassScalingStep,
[…]
