### 2025-08-28 5cbc604
Add blending_factor support to ManuallyScaleBodySegmentsScalingStep
