### 2025-08-29 189d5b0
Replace include(CTest) with enable_testing()
