### 2025-08-29 90af3e6
Fix BUILD_TESTING typo
