### 2025-08-29 1aa53da
Make OSC_BUILD_OSIM_INLINED not-advanced
