### 2025-08-29 c9e1c0b
Clean up build slightly, move compile defines to a config header.
