### 2025-08-29 c431055
Add -Wno-array-bounds when compiling with gcc-12 (spurious)
