### 2025-08-31 e2b95fa
Patch simbody oob casts and reorganize compiler warning flags
