+++ liboscar/Graphics/GraphicsImplementation.cpp
@@ -8153,7 +8153,7 @@ void osc::GraphicsBackend::copy_texture(
     // - https://registry.khronos.org/OpenGL-Refpages/es2.0/xhtml/glTexImage2D.xml
     const auto max_mipmap_level = static_cast<size_t>(max(
         0,
-        static_cast<int>(std::bit_width(static_cast<size_t>(destination.width()))) - 1  // casted to int because of LWG3656
+        static_cast<int>(std::bit_width(static_cast<size_t>(destination.width()))) - 1  // NOLINT(readability-redundant-casting) casted to int because of LWG3656
     ));
 
     OSC_ASSERT(source.dimensionality() == TextureDimensionality::Cube && "provided render texture must be a cubemap to call this method");
