+++ liboscar-demos/learnopengl/PBR/LOGLPBRSpecularIrradianceTexturedTab.cpp
@@ -126,7 +126,7 @@ namespace
 
         constexpr size_t max_mipmap_level = static_cast<size_t>(max(
             0,
-            static_cast<int>(std::bit_width(static_cast<size_t>(level_zero_width)) - 1)  // casted to int because of LWG3656
+            static_cast<int>(std::bit_width(static_cast<size_t>(level_zero_width)) - 1)  // NOLINT(readability-redundant-casting) casted to int because of LWG3656
         ));
         static_assert(max_mipmap_level == 7);
 
