+++ docs/source/make-a-knee.rst
@@ -13,7 +13,7 @@ In this tutorial, we will be making a basic model of a knee using OpenSim Creato
 .. figure:: _static/make-a-knee/after-adding-path-wrap-to-muscle.jpeg
     :width: 60%
 
-    The model created by this tutorial. It contains two bodies, one joint, one muscle,
+    The model created by this tutorial. It contains three bodies, two joints, one muscle,
     and a wrap surface. This covers the basics of using OpenSim Creator's model editor
     to build a model with biological components (see :doc:`make-a-bouncing-block` for
     a more mechanical example).
@@ -56,25 +56,26 @@ In order to follow this tutorial, you will need to download the associated
 resources :download:`download here <_static/the-model-warper/walkthrough-model.zip>`.
 
 
-Create a New Model
+Open Pelvis Model
 ------------------
 
-Create a blank model from the home screen (explained in :ref:`create-new-model`).
+In OpenSim Creator open the pelvis model by clicking ``File`` ->
+``Open``. Select the file ``model_building_leg.osim``.
 
 .. _add-femur-body:
 
 Add a Femur Body
 ----------------
 
-Add a femur body with the femur mesh (``femur_r.vtp``) attached to it to the
+Add a femur body with the femur mesh (``femur_r.vtp``) attached to the pelvis
 model. For this model, use the following parameters:
 
-.. figure:: _static/make-a-knee/add-femur-body.jpeg
+.. figure:: _static/make-a-knee/add-femur-body-to-pelvis-model.jpeg
     :width: 60%
 
-    Create a body called ``femur`` and attach the ``femur_r.vtp`` geometry to it. The
-    mass and intertia can be handled later. The ``femur`` should initially be joined
-    to ``ground`` (the knee joint comes later in the process).
+    Create a body called ``femur_r`` and attach the ``femur_r.vtp`` geometry to it. The
+    mass and intertia can be handled later. The ``femur_r`` should initially be joined
+    to ``hip_r_frame`` (the knee joint comes later in the process).
 
 Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
 :ref:`create-the-foot`.
@@ -84,53 +85,85 @@ Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
 Import Femur Landmarks
 ----------------------
 
-This model will use a landmark-defined approach to defining the knee frame (explained
+This model will use a landmark-defined approach to defining the hip and knee frame (explained
 in :doc:`station-defined-frames`). To do that, we'll need landmarks on the femur that
-correspond to the points that can be used to define the knee's parent frame. The landmarks
-we will use roughly correspond to those explained in `Grood et. al.`_, but our definition
+correspond to the points that can be used to define the parent frames. The landmarks
+we will use roughly correspond to those explained in `Grood et. al.`_. For the knee joint our definition
 will use the Z axis to define knee extension/flexion (Grood et. al. use the X axis) because
 OpenSim's ``PinJoint`` always uses the Z axis for rotation.
 
 You can use the point importer in the model editor from the top menu bar, located at ``Tools`` ->
 ``Import Points``. It will show a popup that you can use to import the source femur knee
 landmarks file (``femur_r.landmarks.csv``) as markers that are attached to
-the ``femur`` body:
+the ``femur_r`` body:
 
-.. figure:: _static/make-a-knee/import-femur-landmarks.jpeg
[…]
