### 2025-09-02 01eabcf
Perform minor fixes to make-a-knee documentation (#1096)
