+++ docs/source/make-a-knee.rst
@@ -3,10 +3,10 @@ Make a Knee
 
 .. warning::
 
-    This tutorial is new ⭐, and ``StationDefinedFrame`` s require OpenSim >= v4.5.1. The content
-    of this tutorial should be valid long-term, but we are waiting for OpenSim GUI v4.6 to be
-    released before we remove any "experimental" labelling. We also anticipate adding some handy
-    tooling around re-socketing existing joints and defining ``StationDefinedFrame`` s.
+    This tutorial is new ⭐, and uses ``StationDefinedFrame``\s, which require OpenSim >= v4.5.1.
+    The content of this tutorial should be valid long-term, but we are waiting for OpenSim GUI
+    v4.6 to be released before we remove any "experimental" labelling. We also anticipate
+    adding some handy tooling around re-socketing existing joints and defining ``StationDefinedFrame``\s.
 
 In this tutorial, we will be making a basic model of a knee using OpenSim Creator:
 
@@ -18,7 +18,7 @@ In this tutorial, we will be making a basic model of a knee using OpenSim Creato
     to build a model with biological components (see :doc:`make-a-bouncing-block` for
     a more mechanical example).
 
-This tutorial will primarily use the model editor workflow to build a new model that
+This tutorial will primarily use the model editor workflow to build a model that
 contains some of the steps/components necessary to build a human model. In essence, the
 content here will be similar to that in :doc:`make-a-bouncing-block`, but with a focus
 on using landmark data, :doc:`station-defined-frames`, and wrap surfaces to build a
@@ -42,10 +42,10 @@ This tutorial assumes you:
 Topics Covered by this Tutorial
 -------------------------------
 
-* Creating an OpenSim model by adding bodies and joints
+* Creating an OpenSim model by adding bodies and joints.
 * Adding a ``StationDefinedFrame`` to the model in order to define anatomically
   representative joint frames.
-* Adding a muscle to the model
+* Adding a muscle to the model.
 * Adding a wrap surface to the model and associating muscles to that surface.
 
 
@@ -53,14 +53,16 @@ Download Resources
 ------------------
 
 In order to follow this tutorial, you will need to download the associated
-resources :download:`download here <_static/the-model-warper/walkthrough-model.zip>`.
+resources (:download:`download here <_static/the-model-warper/walkthrough-model.zip>`)
+and unzip them on your computer.
 
 
 Open Pelvis Model
 ------------------
 
 In OpenSim Creator open the pelvis model by clicking ``File`` ->
-``Open``. Select the file ``model_building_leg.osim``.
+``Open``. Select the file ``model_building_leg.osim`` in the resources directory (see
+above).
 
 .. _add-femur-body:
 
@@ -74,7 +76,7 @@ model. For this model, use the following parameters:
     :width: 60%
 
     Create a body called ``femur_r`` and attach the ``femur_r.vtp`` geometry to it. The
-    mass and intertia can be handled later. The ``femur_r`` should initially be joined
+    mass and intertia can be handled later. ``femur_r`` should initially be joined
     to ``hip_r_frame`` (the knee joint comes later in the process).
 
 Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
@@ -87,7 +89,7 @@ Import Femur Landmarks
 
 This model will use a landmark-defined approach to defining the hip and knee frame (explained
 in :doc:`station-defined-frames`). To do that, we'll need landmarks on the femur that
[…]
