### 2025-09-02 0dd933b
Fix mismatched push-/pop- style (#1097)
