### 2025-09-02 f5f4874
Fix 'Export Points' popup to actually write the exported points (#1098)
