### 2025-09-02 f21dc77
Rename make-a-knee to make-a-lower-leg and ensure it is end-to-end
