+++ docs/source/make-a-lower-leg.rst
@@ -1,5 +1,5 @@
-Make a Knee
-===========
+Make a Lower Leg
+================
 
 .. warning::
 
@@ -8,9 +8,9 @@ Make a Knee
     v4.6 to be released before we remove any "experimental" labelling. We also anticipate
     adding some handy tooling around re-socketing existing joints and defining ``StationDefinedFrame``\s.
 
-In this tutorial, we will be making a basic model of a knee using OpenSim Creator:
+In this tutorial, we will be making a basic model of a lower leg using OpenSim Creator:
 
-.. figure:: _static/make-a-knee/after-adding-path-wrap-to-muscle.jpeg
+.. figure:: _static/make-a-lower-leg/after-adding-path-wrap-to-muscle.jpeg
     :width: 60%
 
     The model created by this tutorial. It contains three bodies, two joints, one muscle,
@@ -18,11 +18,11 @@ In this tutorial, we will be making a basic model of a knee using OpenSim Creato
     to build a model with biological components (see :doc:`make-a-bouncing-block` for
     a more mechanical example).
 
-This tutorial will primarily use the model editor workflow to build a model that
-contains some of the steps/components necessary to build a human model. In essence, the
-content here will be similar to that in :doc:`make-a-bouncing-block`, but with a focus
-on using landmark data, :doc:`station-defined-frames`, and wrap surfaces to build a
-model of a biological system.
+This tutorial will primarily use the model editor workflow to build a model from scratch
+that contains some of the steps/components necessary to build a human model. In
+essence, the content here will be similar to that in :doc:`make-a-bouncing-block`, but
+with a focus on using landmark data, :doc:`station-defined-frames`, and wrap surfaces
+to build a model of a biological system.
 
 
 Prerequisites
@@ -43,11 +43,13 @@ Topics Covered by this Tutorial
 -------------------------------
 
 * Creating an OpenSim model by adding bodies and joints.
-* Adding a ``StationDefinedFrame`` to the model in order to define anatomically
+* Adding ``StationDefinedFrame``\s to the model in order to define anatomically
   representative joint frames.
 * Adding a muscle to the model.
 * Adding a wrap surface to the model and associating muscles to that surface.
 
+
+.. _make-a-lower-leg-resources-link:
 
 Download Resources
 ------------------
@@ -57,22 +59,114 @@ resources (:download:`download here <_static/the-model-warper/walkthrough-model.
 and unzip them on your computer.
 
 
-Open Pelvis Model
+Create a New Model
 ------------------
 
-In OpenSim Creator open the pelvis model by clicking ``File`` ->
-``Open``. Select the file ``model_building_leg.osim`` in the resources directory (see
-above).
+Create a new model, as described in :doc:`make-a-pendulum` (:ref:`create-new-model`).
+
+
+Add a ``pelvis`` Body
+---------------------
+
+Add a pelvis body. For this model, use the following parameters:
+
+.. figure:: _static/make-a-lower-leg/add-pelvis-body.jpeg
+    :width: 60%
+
+    Create a body called ``pelvis``. The mass and intertia can be handled later.
+    ``pelvis`` should be joined to ``ground`` with a ``WeldJoint``.
+
+Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
+:ref:`create-the-foot`.
+
+
+Attach Pelvis Meshes to the ``pelvis`` Body
+-------------------------------------------
+
[…]
