+++ docs/source/index.rst
@@ -34,7 +34,7 @@ Table of Contents
    the-mesh-warper
    preview-experimental-data
    station-defined-frames
-   make-a-knee
+   make-a-lower-leg
    the-model-warper
 
 .. toctree::
