### 2025-09-03 b5027bb
Rework docs with reoriented meshes etc.
