+++ docs/source/make-a-lower-leg.rst
@@ -74,7 +74,8 @@ Add a pelvis body. For this model, use the following parameters:
     :width: 60%
 
     Create a body called ``pelvis``. The mass and intertia can be handled later.
-    ``pelvis`` should be joined to ``ground`` with a ``WeldJoint``.
+    ``pelvis`` should be joined to ``ground`` with a ``WeldJoint``. Geometry is
+    attached in the next step.
 
 Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
 :ref:`create-the-foot`.
@@ -85,7 +86,7 @@ Attach Pelvis Meshes to the ``pelvis`` Body
 
 The resources zip described in :ref:`make-a-lower-leg-resources-link` contain two
 separate pelvis meshes for the left- and right-side. For this model, we are simplifiying
-the pelvis to a single rigid body (``pelvis``), so both meshes need to be attached to it.
+the pelvis to a single rigid body (``pelvis``). Both meshes need to be attached to it.
 
 To attach meshes to ``pelvis``, right-click it in the ``Navigator`` panel and use
 the ``Add > Geometry`` context menu to attach each pelvis mesh:
@@ -94,12 +95,15 @@ the ``Add > Geometry`` context menu to attach each pelvis mesh:
     :width: 60%
 
     Use ``pelvis``'s context menu to ``Add > Geometry`` to it, then select the
-    two pelvis meshes (``pelvis_l.vtp`` and ``pelvis_r.vtp``).
+    two pelvis meshes (``pelvis_l.obj`` and ``pelvis_r.obj``).
 
 .. figure:: _static/make-a-lower-leg/after-attaching-both-pelvis-meshes-to-pelvis.jpeg
     :width: 60%
 
-    The model, after attaching both pelvis meshes to the ``pelvis`` body.
+    The model, after attaching both pelvis meshes to the ``pelvis`` body. The subject
+    was lying down in a scanner when the bone meshes were created. A major part of
+    the model building procedure involves defining frames that transform experimental
+    measurements into standardized coordinate systems.
 
 
 .. _import-pelvis-landmarks:
@@ -107,13 +111,13 @@ the ``Add > Geometry`` context menu to attach each pelvis mesh:
 Import Pelvis Landmarks
 -----------------------
 
-This model will use a landmark-defined approach to define the hip and knee joint
-frames (explained in :doc:`station-defined-frames`). To do that, we'll initially
-import landmarks on the ``pelvis`` body and (later) on a femur body. The landmarks
-we will use roughly correspond to those explained in `Grood et. al.`_; however, our
-knee joint definition will use the Z axis to define knee extension/flexion (Grood
-et. al. use the X axis) because OpenSim's ``PinJoint`` always uses the Z axis
-for rotation.
+This model will use a landmark-defined approach to define the pelvis frame and
+the hip/knee joint frames (explained in :doc:`station-defined-frames`). To do
+that, we'll initially import landmarks on the ``pelvis`` body and (later) on the
+femur body. The landmarks we will use roughly correspond to those explained
+in `Grood et. al.`_; however, our knee joint definition will use the Z axis to
+define knee extension/flexion (Grood et. al. use the X axis) because OpenSim's
+``PinJoint`` always uses the Z axis for rotation.
 
 To import the landmarks, you can use the point importer in the model editor from
 the top menu bar, located at ``Tools > Import Points``. It will show a popup
@@ -127,15 +131,15 @@ markers that are attached to the ``pelvis`` body:
[…]
