### 2025-09-04 a78f8e5
Squashed 'third_party/spdlog/' changes from 453be2e08..6fa36017c
6fa36017c Version 1.15.3
c73b8cc40 Update comment
7ca6a4fb2 Update commemt
070e1c974 Update comment
0d31acae2 Fmt 11.2.0 (#3399)
943fcbd76 Register replace logger (#3398)
7e022c430 Feature 3379 (#3397)
548b26425 Fix warning C4530 (#3393)
847db3375 dup_filter_sink: remove notification_level argument; use last message log level for notification instead (#3390)
bb8694b50 Fix links for #3380 (#3381)
cec28bf83 Fix links to local reference. (#3378)
bd0609d7a Update README.md
1f4959c83 Fix link to wiki. (#3377)
48bcf39a6 Version 1.15.2

[…]
