+++ tests/test_misc.cpp
@@ -1,11 +1,12 @@
+#ifdef _WIN32  // to prevent fopen warning on windows
+    #define _CRT_SECURE_NO_WARNINGS
+#endif
+
 #include "includes.h"
 #include "test_sink.h"
-#include "spdlog/fmt/bin_to_hex.h"
 
-template<class T>
-std::string log_info(const T &what, spdlog::level::level_enum logger_level = spdlog::level::info)
-{
-
+template <class T>
+std::string log_info(const T& what, spdlog::level::level_enum logger_level = spdlog::level::info) {
     std::ostringstream oss;
     auto oss_sink = std::make_shared<spdlog::sinks::ostream_sink_mt>(oss);
 
@@ -17,15 +18,14 @@ std::string log_info(const T &what, spdlog::level::level_enum logger_level = spd
     return oss.str().substr(0, oss.str().length() - strlen(spdlog::details::os::default_eol));
 }
 
-TEST_CASE("basic_logging ", "[basic_logging]")
-{
+TEST_CASE("basic_logging ", "[basic_logging]") {
     // const char
     REQUIRE(log_info("Hello") == "Hello");
-    REQUIRE(log_info("") == "");
+    REQUIRE(log_info("").empty());
 
     // std::string
     REQUIRE(log_info(std::string("Hello")) == "Hello");
-    REQUIRE(log_info(std::string()) == std::string());
+    REQUIRE(log_info(std::string()).empty());
 
     // Numbers
     REQUIRE(log_info(5) == "5");
@@ -35,17 +35,15 @@ TEST_CASE("basic_logging ", "[basic_logging]")
     // REQUIRE(log_info(some_logged_class("some_val")) == "some_val");
 }
 
-TEST_CASE("log_levels", "[log_levels]")
-{
-    REQUIRE(log_info("Hello", spdlog::level::err) == "");
-    REQUIRE(log_info("Hello", spdlog::level::critical) == "");
+TEST_CASE("log_levels", "[log_levels]") {
+    REQUIRE(log_info("Hello", spdlog::level::err).empty());
+    REQUIRE(log_info("Hello", spdlog::level::critical).empty());
     REQUIRE(log_info("Hello", spdlog::level::info) == "Hello");
     REQUIRE(log_info("Hello", spdlog::level::debug) == "Hello");
     REQUIRE(log_info("Hello", spdlog::level::trace) == "Hello");
 }
 
-TEST_CASE("level_to_string_view", "[convert_to_string_view")
-{
+TEST_CASE("level_to_string_view", "[convert_to_string_view]") {
     REQUIRE(spdlog::level::to_string_view(spdlog::level::trace) == "trace");
     REQUIRE(spdlog::level::to_string_view(spdlog::level::debug) == "debug");
     REQUIRE(spdlog::level::to_string_view(spdlog::level::info) == "info");
@@ -55,8 +53,7 @@ TEST_CASE("level_to_string_view", "[convert_to_string_view")
     REQUIRE(spdlog::level::to_string_view(spdlog::level::off) == "off");
 }
 
-TEST_CASE("to_short_c_str", "[convert_to_short_c_str]")
-{
+TEST_CASE("to_short_c_str", "[convert_to_short_c_str]") {
     REQUIRE(std::string(spdlog::level::to_short_c_str(spdlog::level::trace)) == "T");
     REQUIRE(std::string(spdlog::level::to_short_c_str(spdlog::level::debug)) == "D");
     REQUIRE(std::string(spdlog::level::to_short_c_str(spdlog::level::info)) == "I");
[…]

[…]
