### 2025-09-04 60d1c05
Squashed 'third_party/libosim/opensim-core/' changes from 28e14e6a7..7f3ec440a
7f3ec440a Merge pull request #4115 from nickbianco/contact_geometry_updates
4da0f387c Merge pull request #4141 from opensim-org/fix_4138-oob-assert-in-remove
c128acc26 Fix incorrect assertion in OpenSim::Array::remove (#4138)
d9c7da15f Add new ContactGeometry concrete implementations
9fd295df4 Merge pull request #4131 from francoisk/static-init-order
c34fa052f Merge pull request #4130 from nickbianco/test_serialize_opensim_objects_patch
40696bc13 Add warning to Logger to not use it from static object destructors
1d0
[…]
