+++ Simbody/src/CableSpan.cpp
@@ -18,6 +18,7 @@
 
 #include "CableSpan_SubsystemTestHelper_Impl.h"
 
+#include "simbody/internal/common.h"
 #include "simbody/internal/CableSpan.h"
 #include "simbody/internal/MultibodySystem.h"
 #include "simbody/internal/SimbodyMatterSubsystem.h"
@@ -26,8 +27,14 @@
 
 using namespace SimTK;
 
+// Define a unique index for cable segments.
+namespace {
+SimTK_DEFINE_UNIQUE_INDEX_TYPE(CableSegmentIndex);
+}
+
 // For convenience.
 using ObstacleIndex = CableSpanObstacleIndex;
+using ViaPointIndex = CableSpanViaPointIndex;
 
 //==============================================================================
 //                            Data Structures
@@ -124,13 +131,13 @@ struct MatrixWorkspace {
     explicit MatrixWorkspace(int problemSize)
     {
         // Total number of obstacles in contact with cable.
-        const int n = problemSize;
+        numObstaclesInContact = problemSize;
         // Dimension of the free variable vector: all geodesic corrections.
-        const int nQ = n * c_GeodesicDOF;
+        const int nQ = numObstaclesInContact * c_GeodesicDOF;
         // Total number of constraints (2 per obstacle).
-        const int nC = n * 2;
+        const int nC = numObstaclesInContact * 2;
 
-        lineSegments.resize(n + 1);
+        lineSegments.resize(numObstaclesInContact + 1);
         pathCorrection = Vector(nQ, 0.);
 
         normalPathError   = Vector(nC, 0.);
@@ -158,10 +165,8 @@ struct MatrixWorkspace {
         const Vector& pathCorrection,
         int activeCurveIx)
     {
-        SimTK_ASSERT(
-            pathCorrection.size() <= (activeCurveIx + 1) * c_GeodesicDOF,
-            "Invalid size of pathCorrection vector.");
         const int eltIx = activeCurveIx * c_GeodesicDOF;
+        SimTK_ASSERT(0 <= eltIx && eltIx + 3 < pathCorrection.size(), "Invalid size of pathCorrection vector.");
         return {
             pathCorrection[eltIx],
             pathCorrection[eltIx + 1],
@@ -169,6 +174,8 @@ struct MatrixWorkspace {
             pathCorrection[eltIx + 3]};
     }
 
+    /* The number of obstacles in contact with the cable. */
+    int numObstaclesInContact;
     /* All straight line segments in the current path. */
     std::vector<LineSegment> lineSegments;
     /* The path correction vector contains the NaturalGeodesicCorrection
@@ -302,6 +309,12 @@ struct CableSpanData {
         // Tangent vector to the cable at the termination point, in the ground
         // frame.
         UnitVec3 terminationTangent_G{NaN, NaN, NaN};
+        // Tangent vectors to the cable segments incoming to (i.e., immediately
+        // prior to) each via point in the path, in the ground frame.
+        Array_<UnitVec3, ViaPointIndex> viaPointInTangents_G;
+        // Tangent vectors to the cable segments outgoing from (i.e.,
+        // immediately following) each via point in the path, in the ground frame.
+        Array_<UnitVec3, ViaPointIndex> viaPointOutTangents_G;
         // The total cable length.
         Real cableLength = NaN;
         // The path smoothness, computed as the maximum angular misalignment at
@@ -585,6 +598,22 @@ private:
 
     int realizeSubsystemTopologyImpl(State& state) const override;
 
+    int realizeSubsystemModelImpl(State& state) const override;
+
+    int realizeSubsystemInstanceImpl(const State& state) const override;
+
+    int realizeSubsystemTimeImpl(const State& state) const override;
+
[…]
