### 2025-09-04 4a89431
Remove unneeded compile define
