### 2025-09-04 52050c3
Patch OpenSim::Marker::generateDecorations invalid when marker attached to custom frame (#1101)
