+++ CHANGELOG.md
@@ -50,6 +50,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed a styling bug in the `Export Points` popup that shows a styling error message
   rather than the appropriate tooltip (#1097).
 - Fixed the `Export Points` popup not writing the exported CSV file (#1098).
+- Patched a bug in OpenSim that causes `OpenSim::Marker`s to emit incorrect 3D decorations
+  when they are attached to custom frames (#1101).
 
 ## [0.5.25] - 2025/07/14
 
