### 2025-09-05 7c88caf
Add Move To > /markerset to Marker context menu (#1102)
