### 2025-09-05 8b50722
Update imgui_base_config.ini to correctly dock model warper control panel by default
