+++ resources/OpenSimCreator/imgui_base_config.ini
@@ -1,6 +1,6 @@
 [Window][DockSpaceViewport_11111111]
-Pos=0,77
-Size=1920,903
+Pos=0,83
+Size=1919,997
 Collapsed=0
 
 [Window][Debug##Default]
@@ -9,7 +9,7 @@ Size=400,400
 Collapsed=0
 
 [Window][Splash screen]
-Pos=640,317
+Pos=639,352
 Size=640,512
 Collapsed=0
 
@@ -62,7 +62,7 @@ DockId=0x00000020,0
 
 [Window][Coordinates]
 Pos=0,77
-Size=270,484
+Size=270,522
 Collapsed=0
 DockId=0x00000020,1
 
@@ -215,13 +215,13 @@ Collapsed=0
 
 [Window][viewer0]
 Pos=272,77
-Size=1316,672
+Size=1315,743
 Collapsed=0
 DockId=0x00000034,0
 
 [Window][Log]
-Pos=272,751
-Size=1316,229
+Pos=460,851
+Size=1459,229
 Collapsed=0
 DockId=0x0000000A,0
 
@@ -261,8 +261,8 @@ Collapsed=0
 DockId=0x0000002A,0
 
 [Window][Loading Message]
-Pos=672,482
-Size=576,56
+Pos=671,518
+Size=575,56
 Collapsed=0
 
 [Window][select experiment]
@@ -310,7 +310,7 @@ DockId=0x00000015,0
 
 [Window][wizard_3dViewer]
 Pos=269,46
-Size=1382,732
+Size=1650,1034
 Collapsed=0
 DockId=0x00000034,0
 
@@ -321,7 +321,7 @@ Collapsed=0
 DockId=0x0000002E,0
 
 [Window][Save Changes?]
-Pos=704,443
+Pos=703,479
 Size=512,122
 Collapsed=0
 
@@ -360,20 +360,20 @@ Collapsed=0
 DockId=0x0000002C,0
 
 [Window][Output Watches]
-Pos=1590,77
-Size=330,903
+Pos=1589,77
+Size=330,974
 Collapsed=0
 DockId=0x0000002A,0
 
 [Window][Navigator]
 Pos=0,77
-Size=270,484
+Size=270,522
 Collapsed=0
 DockId=0x00000020,0
 
 [Window][Properties]
-Pos=0,563
-Size=270,417
+Pos=0,601
+Size=270,450
 Collapsed=0
 DockId=0x0000000B,0
 
@@ -391,7 +391,7 @@ DockId=0x00000024,0
 
 [Window][navigator]
 Pos=0,46
-Size=267,963
+Size=267,1034
 Collapsed=0
 DockId=0x0000001D,0
 
@@ -440,27 +440,28 @@ Collapsed=0
 DockId=0x00000034,0
 
 [Window][Result]
-Pos=1327,83
-Size=593,897
+Pos=1326,83
+Size=593,968
 Collapsed=0
 DockId=0x00000029,0
 
 [Window][Destination Mesh]
 Pos=572,83
-Size=753,897
+Size=752,968
 Collapsed=0
 DockId=0x00000034,0
 
 [Window][Source Mesh]
 Pos=0,83
-Size=570,897
+Size=570,968
 Collapsed=0
 DockId=0x00000026,0
 
 [Window][Controls]
-Pos=60,75
-Size=54,58
+Pos=1686,46
+Size=233,1034
 Collapsed=0
+DockId=0x00000018,0
 
 [Window][Perf]
 Pos=60,75
@@ -468,8 +469,8 @@ Size=965,612
 Collapsed=0
 
 [Window][muscleplot0]
-Pos=1590,531
-Size=330,225
+Pos=1589,567
+Size=330,484
 Collapsed=0
 DockId=0x00000032,0
 
[…]
