### 2025-09-05 dcc8528
Add 'Bake StationDefinedFrames checkbox to model warper (#1104)
