### 2025-09-05 c2ff0a3
Add Sabrina's manual scale factors to the-model-warper docs
