### 2025-09-05 236a424
Fix issues found by clang-tidy (#1105)
