+++ docs/source/make-a-lower-leg.rst
@@ -13,8 +13,8 @@ In this tutorial, we will be making a basic model of a lower leg using OpenSim C
 .. figure:: _static/make-a-lower-leg/after-adding-path-wrap-to-muscle.jpeg
     :width: 60%
 
-    The model created by this tutorial. It contains three bodies, two joints, one muscle,
-    and a wrap surface. This covers the basics of using OpenSim Creator's model editor
+    The model created by this tutorial. It contains three bodies, two joints, three muscles,
+    and a wrap cylinder. This covers the basics of using OpenSim Creator's model editor
     to build a model with biological components (see :doc:`make-a-bouncing-block` for
     a more mechanical example).
 
@@ -34,9 +34,9 @@ This tutorial assumes you:
   and :doc:`make-a-bouncing-block`.
 - (*optional*) The modelling process will also include adding a ``StationDefinedFrame`` to
   the model. The details of how they work is explained in :doc:`station-defined-frames`.
-- (*optional*) The building process uses externally-provided landmarks in a CSV. If you
-  would like to know how to manually place landmarks on a mesh, we recommend reading
-  through :doc:`the-mesh-importer`.
+- (*optional*) The building process uses externally-provided landmarks from CSV files (e.g.
+  ``femur_r.landmarks.csv``). If you would like to know how to manually place landmarks
+  on a mesh, we recommend reading through :doc:`the-mesh-importer`.
 
 
 Topics Covered by this Tutorial
@@ -74,8 +74,9 @@ Add a pelvis body. For this model, use the following parameters:
     :width: 60%
 
     Create a body called ``pelvis``. The mass and intertia can be handled later.
-    ``pelvis`` should be joined to ``ground`` with a ``WeldJoint``. Geometry is
-    attached in the next step.
+    ``pelvis`` should directly (no offset frames) be joined to ``ground`` with
+    a ``FreeJoint`` called ``pelvis_to_ground``. Pelvis meshes are attached in
+    the next step.
 
 Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
 :ref:`create-the-foot`.
@@ -94,16 +95,18 @@ the ``Add > Geometry`` context menu to attach each pelvis mesh:
 .. figure:: _static/make-a-lower-leg/add-geometry-to-pelvis-context-menu.jpeg
     :width: 60%
 
-    Use ``pelvis``'s context menu to ``Add > Geometry`` to it, then select the
-    two pelvis meshes (``pelvis_l.obj`` and ``pelvis_r.obj``).
+    Use ``pelvis``'s context menu to ``Add > Geometry`` to it, then select one
+    of the pelvis meshes (``pelvis_l.obj`` or ``pelvis_r.obj``). Repeat this process
+    for the other pelvis mesh.
 
 .. figure:: _static/make-a-lower-leg/after-attaching-both-pelvis-meshes-to-pelvis.jpeg
     :width: 60%
 
-    The model, after attaching both pelvis meshes to the ``pelvis`` body. The subject
-    was lying down in a scanner when the bone meshes were created. A major part of
-    the model building procedure involves defining frames that transform experimental
-    measurements into standardized coordinate systems.
+    The model after attaching both ``pelvis_l.obj`` and ``pelvis_r.obj`` to the
+    ``pelvis`` body. For context, assume the reference subject was lying down
+    when these bones were scanned. A major part of the model building procedure involves defining
+    frames that transform experimental measurements into standardized coordinate
+    systems.
 
 
 .. _import-pelvis-landmarks:
[…]
