### 2025-09-06 6f0e9e7
Make 'Add Body' attach button only show mesh filename rather than full path
