+++ docs/source/make-a-lower-leg.rst
@@ -8,9 +8,9 @@ Make a Lower Leg
     v4.6 to be released before we remove any "experimental" labelling. We also anticipate
     adding some handy tooling around re-socketing existing joints and defining ``StationDefinedFrame``\s.
 
-In this tutorial, we will be making a basic model of a lower leg using OpenSim Creator:
+In this tutorial, we will be making a model of a lower leg using OpenSim Creator:
 
-.. figure:: _static/make-a-lower-leg/after-adding-path-wrap-to-muscle.jpeg
+.. figure:: _static/make-a-lower-leg/final-model-screenshot.jpeg
     :width: 60%
 
     The model created by this tutorial. It contains three bodies, two joints, three muscles,
@@ -73,7 +73,7 @@ Add a pelvis body. For this model, use the following parameters:
 .. figure:: _static/make-a-lower-leg/add-pelvis-body.jpeg
     :width: 60%
 
-    Create a body called ``pelvis``. The mass and intertia can be handled later.
+    Create a body called ``pelvis``. The mass and inertia can be handled later.
     ``pelvis`` should directly (no offset frames) be joined to ``ground`` with
     a ``FreeJoint`` called ``pelvis_to_ground``. Pelvis meshes are attached in
     the next step.
@@ -86,7 +86,7 @@ Attach Pelvis Meshes to the ``pelvis`` Body
 -------------------------------------------
 
 The resources zip described in :ref:`make-a-lower-leg-resources-link` contain two
-separate pelvis meshes for the left- and right-side. For this model, we are simplifiying
+separate pelvis meshes for the left- and right-side. For this model, we are simplifying
 the pelvis to a single rigid body (``pelvis``). Both meshes need to be attached to it.
 
 To attach meshes to ``pelvis``, right-click it in the ``Navigator`` panel and use
@@ -156,8 +156,8 @@ Adding a ``pelvis_frame`` is described in the following two figures:
 
     When creating the ``StationDefinedFrame``, call it ``pelvis_frame``, make ``ASIS_midpoint``
     the frame ``origin_point`` and ``point_a``, ``PSIS_midpoint`` ``point_b``, and ``ASIS_r`` ``point_c``
-    Addtionally, ensure that ``ab_axis`` is ``-x`` and ``ab_x_ac_axis`` is ``+y``. The :doc:`station-defined-frames`
-    page explains ``StationDefinedFrame``\s in more detail.
+    Additionally, ensure that ``ab_axis`` is ``-x`` and ``ab_x_ac_axis`` is ``+y``. The
+    :doc:`station-defined-frames` page explains ``StationDefinedFrame``\s in more detail.
 
 .. _reassign-pelvis-root-joint:
 
@@ -196,7 +196,7 @@ define a ``StationDefinedFrame`` on ``pelvis`` called ``hip_r_frame`` as follows
     Right-click the ``pelvis`` body and add a ``StationDefinedFrame``. Call it
     ``hip_r_frame``, make ``Acetabulum_centre`` the frame ``origin_point``,
     ``PSIS_midpoint`` ``point_a``, ``ASIS_midpoint`` ``point_b``, and ``ASIS_l``
-    ``point_c``. Addtionally, ensure that ``ab_axis`` is ``+x`` and ``ab_x_ac_axis``
+    ``point_c``. Additionally, ensure that ``ab_axis`` is ``+x`` and ``ab_x_ac_axis``
     is ``+y``.
 
 .. figure:: _static/make-a-lower-leg/after-adding-hip-sdf.jpeg
@@ -255,7 +255,7 @@ the landmarks attached to it:
     Right-click the ``femur_r`` body and add a ``StationDefinedFrame``. Call it
     ``hip_r_child_frame``, make  ``femur_r_head_centre`` the ``origin_point`` and
[…]
