### 2025-09-08 10968d8
Update CHANGELOG.md for 0.6.0 (#1105)
