### 2025-09-08 b8f2f8c
Strip trailing whitespace etc. in docs/
