### 2025-09-08 b3eb2d7
Update format_sources.py
