### 2025-09-09 799ce05
Add signing/notarization steps to MacOS packaging (#1106)
