+++ osc/MacOS/codesign_osc.dmg.cmake.in
@@ -1,0 +1,20 @@
+execute_process(
+        COMMAND codesign
+        --verbose
+        --timestamp
+        --sign "@OSC_CODESIGN_DEVELOPER_ID@"
+        "${CPACK_PACKAGE_FILES}"
+        RESULT_VARIABLE res
+        COMMAND_ECHO STDERR
+)
+if(NOT res EQUAL 0)
+    message(FATAL_ERROR "codesign failed!")
+endif()
+execute_process(
+        COMMAND codesign --verify --verbose=4 "${CPACK_PACKAGE_FILES}"
+        RESULT_VARIABLE res
+        COMMAND_ECHO STDERR
+)
+if(NOT res EQUAL 0)
+    message(FATAL_ERROR "codesign failed!")
+endif()
