### 2025-09-09 62723b4
Add documentation note about xz tarballing on MacOS
