### 2025-09-09 cffc29e
Update codemeta.json, CITATION.cff, and README.md for 0.6.0 (#1105)
