### 2025-09-16 34ac0d8
Restructure paper to address reviewer comments, add TODOs mentioning unresolved reviewer concerns
