+++ paper/paper.md
@@ -8,7 +8,6 @@ tags:
   - biomechanics
   - rigid body dynamics
   - user-interface tooling
-  - cad tooling
   - imgui
   - C++
 authors:
@@ -36,39 +35,68 @@ bibliography: paper.bib
 
 # Summary
 
-OpenSim Creator is an open-source "biomechanical CAD" user interface (UI) that focuses on building and evaluating musculoskeletal models in OpenSim, something that previously required manually editing extensible markup language (XML). In this paper, we outline its architecture and provide concrete examples of how OpenSim Creator has already been used to build new biomechanical models.
+OpenSim Creator is an open-source graphical interface that lowers the barrier to building and evaluating musculoskeletal models in OpenSim, which previously required editing XML files or custom code. This paper describes its architecture and features. It also demonstrates its use in developing new biomechanical models.
 
 
-# Statement of need
+# Statement of Need
 
-OpenSim [@Delp:2007; @Seth:2018] is an open-source musculoskeletal modeling platform with a large and growing number of users in diverse fields. This follows the general trend of musculoskeletal modeling being increasingly used in clinical [@Marconi:2023], biomechanical [@Porsa:2016], and palaeobiological [@Beesel:2021; @Demuth:2023] studies. However, creating a biomechanical model remains a challenge because the process is both laborious, as it requires the input of a large amount of biological data, and complex, as it requires expertise in biomedical imaging, model/data file formats, and musculoskeletal anatomy.
+OpenSim [@Delp:2007; @Seth:2018] is an open-source musculoskeletal modeling platform with a large and growing user base across diverse fields, including clinical [@Marconi:2023], biomechanical [@Porsa:2016], and palaeobiological [@Beesel:2021; @Demuth:2023] studies. However, creating a biomechanical model remains labor-intensive, requiring the careful integration of biological data, and technically demanding, requiring expertise in biomedical imaging, model/data file formats, and musculoskeletal anatomy.
 
-As an illustrative example, @Bishop:2021 outline a workflow for building a biomechanical model of a dinosaur. It involves specialized steps such as a) acquiring and segmenting bone morphology, centers of mass, and inertia tensors from fossil records; b) rigging the bones into a skeletal model; c) reconstructing musculature from soft tissue volumes; and d) using the resulting model in a variety of biomechanical simulations. Specialized procedures like these may need to be repeated for each subject in an experiment because animals are unique, which can significantly increase the amount of labor required. Scaling existing models can minimize that work, but designing a correct, unbiased [@vanDerKruk:2025], scaling procedure that is capable of handling a variety of shapes, injuries, and deformities is also challenging.
+For example, building a biomechanical model of a dinosaur involves segmenting bone morphology, rigging skeletal structures, reconstructing musculature, and running biomechanical simulations [@Bishop:2021]. Such workflows often need to be repeated for each subject, and while scaling existing models can reduce this effort, developing robust and unbiased procedures that handle diverse shapes, injuries, or deformities remains challenging [@vanDerKruk:2025]. Real-time visual tools can provide valuable feedback during model building, scaling, and validation.
 
-Similar challenges have been addressed in fields such as 3D design, engineering, and game development. For example, engineers commonly use computer-aided design (CAD) software both to reduce labor, by automating actions such as 3D object placement; and to increase the user's expertise, by enabling model editing in a safe undoable area with immediate, useful, visual feedback. As musculoskeletal modelling with OpenSim expands to fields that are not traditionally occupied by engineers (e.g., medicine, palaeobiology), real-time model development with visual feedback becomes paramount. OpenSim Creator provides that tooling by building on ideas from existing CAD software - with a biomechanical twist.
[…]
