+++ paper/paper.bib
@@ -116,6 +116,24 @@
   url = {http://www.blender.org},
 }
 
+@misc{Godot,
+  title = {Godot: a Free, Open-Source Game Engine},
+  author = {Godot, Online Community},
+  organization = {Godot Foundation},
+  address = {Stichting Godot, Spoorlaan 5, 1911HA Uitgeest, The Netherlands},
+  year = {2014},
+  url = {https://godotengine.org/},
+}
+
+@misc{FreeCAD,
+  title = {FreeCAD: Your own 3D parametric modeler},
+  author = {FreeCAD, Online Community},
+  organization = {FreeCAD Project Association},
+  address = {Rue Docteur Kuborn 31 / 46, 1070 Anderlecht, Belgium},
+  year = {2002},
+  url = {https://www.freecad.org/},
+}
+
 @misc{ImGui,
   title = {Dear ImGui: Bloat-free Graphical User interface for C++ with minimal dependencies},
   author = {Omar Cornut},
@@ -151,4 +169,58 @@
   title = {Addition of torso musculature to a simplified musculoskeletal model for 2D predictive simulation of the sit-to-walk movement},
   year = {2023},
   url = {http://resolver.tudelft.nl/uuid:91925ea8-199e-4304-8b8f-8f52519a7931},
+}
+@article{Besier:2014,
+  author="Zhang, Ju and Sorby, Hugh and Clement, John and Thomas, C. David L. and Hunter, Peter and Nielsen, Poul and Lloyd, David and Taylor, Mark and Besier, Thor",
+  title="The MAP Client: User-Friendly Musculoskeletal Modelling Workflows",
+  year="2014",
+  doi = {10.1007/978-3-319-12057-7_21},
+  isbn="978-3-319-12057-7",
+  booktitle={International Symposium on Biomedical Simulation},
+  pages="182--192",
+  publisher="Springer International Publishing"
+}
+@article{Hammond:2025,
+  title={The Neuromusculoskeletal Modeling Pipeline: MATLAB-based model personalization and treatment optimization functionality for OpenSim},
+  author={Hammond, Claire V and Williams, Spencer T and Vega, Marleny M and Ao, Di and Li, Geng and Salati, Robert M and Pariser, Kayla M and Shourijeh, Mohammad S and Habib, Ayman W and Patten, Carolynn and others},
+  journal={Journal of NeuroEngineering and Rehabilitation},
+  volume={22},
+  number={1},
+  pages={112},
+  year={2025},
+  publisher={Springer},
+  doi={10.1186/s12984-025-01629-5}
+}
+@article{Valente:2017,
+  title = {nmsBuilder: Freeware to create subject-specific musculoskeletal models for OpenSim},
+  journal = {Computer Methods and Programs in Biomedicine},
+  volume = {152},
+  pages = {85-92},
+  year = {2017},
+  issn = {0169-2607},
+  doi = {https://doi.org/10.1016/j.cmpb.2017.09.012},
+  url = {https://www.sciencedirect.com/science/article/pii/S016926071730367X},
+  author = {Giordano Valente and Gianluigi Crimi and Nicola Vanella and Enrico Schileo and Fulvia Taddei},
+  keywords = {Musculoskeletal model, Subject-specific model, Image-based model, Imaging data processing, Freeware, OpenSim}
+}
+@article {Bijlert:2024,
+	author = {van Bijlert, Pasha A.},
+	title = {MuSkeMo: Open-source software to construct, analyze, and visualize human and animal musculoskeletal models and movements in Blender},
+	elocation-id = {2024.12.10.627828},
+	year = {2024},
+	doi = {10.1101/2024.12.10.627828},
+	publisher = {Cold Spring Harbor Laboratory},
[…]
