### 2025-09-16 ComputationalBiomechanicsLab/opensim-creator#1078 @adamkewley
As of 0.6.0, this is no longer relevant because the "Make a Leg" and "The Model Warper" tutorials were mostly rewritten from scratch so that users can (and do!) reproduce all parts of the tutorial themselves.
