### 2025-09-16 c27d79e
Address most TODOs apart from cite-able example use-cases
