+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 # Summary
 
-OpenSim Creator is an open-source graphical interface that lowers the barrier to building and evaluating musculoskeletal models in OpenSim, which previously required editing XML files or custom code. This paper describes its architecture and features. It also demonstrates its use in developing new biomechanical models.
+OpenSim Creator is an open-source graphical interface that lowers the barrier to building and evaluating musculoskeletal models in OpenSim, which previously required editing XML files or custom code. This paper describes its architecture and features, and demonstrates its application in developing new biomechanical models.
 
 
 # Statement of Need
@@ -44,7 +44,7 @@ OpenSim [@Delp:2007; @Seth:2018] is an open-source musculoskeletal modeling plat
 
 For example, building a biomechanical model of a dinosaur involves segmenting bone morphology, rigging skeletal structures, reconstructing musculature, and running biomechanical simulations [@Bishop:2021]. Such workflows often need to be repeated for each subject, and while scaling existing models can reduce this effort, developing robust and unbiased procedures that handle diverse shapes, injuries, or deformities remains challenging [@vanDerKruk:2025]. Real-time visual tools can provide valuable feedback during model building, scaling, and validation.
 
-These challenges not only affect research but also limit the accessibility of musculoskeletal modelling in education. As OpenSim use expands beyond engineering (e.g., medicine, palaeobiology) and open education becomes more widespread (e.g., online and open courses [@Weller:2018]), there is a growing need for open-source, visual, and easy-to-install tools that educators can use to teach musculoskeletal model-building concepts clearly — tools that lower barriers to entry compared to traditional code- or engineering-focused workflows.
+These challenges not only affect research but also limit accessibility for students and new researchers learning musculoskeletal modeling. As OpenSim use expands beyond engineering (e.g., medicine, palaeobiology) and open education becomes more widespread (e.g., online and open courses [@Weller:2018]), there is a growing need for open-source, visual, and easy-to-install tools that educators can use to teach musculoskeletal model-building concepts clearly — tools that lower barriers to entry compared to traditional code- or engineering-focused workflows.
 
 
 # State of the Field
@@ -53,9 +53,9 @@ Similar modeling challenges are addressed in fields such as 3D design (e.g., Ble
 
 OpenSim includes a user interface (UI) that can edit existing OpenSim models, but it is specialized around providing tools that accept existing model files as input. For example, it includes tools for scaling models, computing inverse kinematics (IK) from marker motion recordings, and computed muscle control (CMC), but it does not provide tools for adding bodies or joints to an existing model. Therefore, to create new models, users typically use a combination of the following three options:
 
-- **Manually write the model file (XML)**: It is possible to build any model manually, but it requires understanding OpenSim's XML datastructure, is laborious, and does not provide immediate or visual feedback.
+- **Manually write the model file (XML)**: It is possible to build any model manually, but it requires understanding OpenSim's XML data structure, is laborious, and does not provide immediate or visual feedback.
 - **Use code/libraries to generate or optimize a model**: Examples include the MAP Client [@Besier:2014] and the Neuromusculoskeletal Modeling Pipeline [@Hammond:2025]. These can greatly accelerate specific tasks but require coding knowledge and are not general-purpose model-building suites.
-- **Use visual tools that can export OpenSim models**: Examples include NMSBuilder [@Valente:2017] and MuSkeMo [@Bijlert:2024]. Most visual tools manipulate a separate datastructure exported to OpenSim and do not support direct editing, viewing, or simulation in-UI.
[…]
