### 2025-09-24 892b6f3
Remove redundant variable setters from ci_build_mac-ARCH.sh
