+++ scripts/ci_build_mac-amd64.sh
@@ -6,14 +6,4 @@
 
 set -xeuo pipefail
 
-# Hide window creation in CI, because CI runners typically do not have
-# a desktop environment.
-export OSC_INTERNAL_HIDE_WINDOW="1"
-
-# Ensure dependencies are re-checked/re-built if the CI script is ran on
-# a potentially stale/cached workspace directory.
-export OSCDEPS_BUILD_ALWAYS=${OSCDEPS_BUILD_ALWAYS:-ON}
-
-export OSC_OSX_ARCHITECTURES=x86_64
-
-./scripts/ci_build_mac.sh
+OSC_OSX_ARCHITECTURES=x86_64 ./scripts/ci_build_mac.sh
