### 2025-09-24 1fc5e9c
Minorly format hellooscar.cpp
