### 2025-09-24 a6a3918
Drop SimTKsimbody.h header usage from ForwardDynamicSimulator
