### 2025-09-24 5070f1d
Address final TODO
