+++ paper/paper.bib
@@ -163,13 +163,6 @@
   doi = {10.3389/fnbot.2019.00090},
   issn = {1662-5218},
 }
-
-@misc{Heijerman:2023,
-  author = {Heijerman, Scott},
-  title = {Addition of torso musculature to a simplified musculoskeletal model for 2D predictive simulation of the sit-to-walk movement},
-  year = {2023},
-  url = {http://resolver.tudelft.nl/uuid:91925ea8-199e-4304-8b8f-8f52519a7931},
-}
 @article{Besier:2014,
   author="Zhang, Ju and Sorby, Hugh and Clement, John and Thomas, C. David L. and Hunter, Peter and Nielsen, Poul and Lloyd, David and Taylor, Mark and Besier, Thor",
   title="The MAP Client: User-Friendly Musculoskeletal Modelling Workflows",
@@ -198,7 +191,7 @@
   pages = {85-92},
   year = {2017},
   issn = {0169-2607},
-  doi = {https://doi.org/10.1016/j.cmpb.2017.09.012},
+  doi = {10.1016/j.cmpb.2017.09.012},
   url = {https://www.sciencedirect.com/science/article/pii/S016926071730367X},
   author = {Giordano Valente and Gianluigi Crimi and Nicola Vanella and Enrico Schileo and Fulvia Taddei},
   keywords = {Musculoskeletal model, Subject-specific model, Image-based model, Imaging data processing, Freeware, OpenSim}
@@ -223,4 +216,89 @@
   month = {Apr},
   title = {Mapping the Open Education Landscape: Citation Network Analysis of Historical Open and Distance Education Research},
   year = {2018}
+},
+@INPROCEEDINGS{Belli:2023,
+  author={Beck, Irene and Belli, Italo and Peternel, Luka and Seth, Ajay and Prendergast, J. Micah},
+  booktitle={2023 IEEE-RAS 22nd International Conference on Humanoid Robots (Humanoids)}, 
+  title={Real-Time Tendon Strain Estimation of Rotator-Cuff Muscles during Robotic-Assisted Rehabilitation}, 
+  year={2023},
+  volume={},
+  number={},
+  pages={1-8},
+  keywords={Shoulder;Medical treatment;Kinematics;Muscles;Real-time systems;Robots;Injuries},
+  doi={10.1109/Humanoids57100.2023.10375158},
+}
+@article{Modenese:2016,
+  title = {Estimation of musculotendon parameters for scaled and subject specific musculoskeletal models using an optimization technique},
+  journal = {Journal of Biomechanics},
+  volume = {49},
+  number = {2},
+  pages = {141-148},
+  year = {2016},
+  issn = {0021-9290},
+  doi = {10.1016/j.jbiomech.2015.11.006},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021929015006363},
+  author = {Luca Modenese and Elena Ceseracciu and Monica Reggiani and David G. Lloyd},
+  keywords = {Optimal fiber length, Tendon slack length, Scaling, Subject specific, Muscle models, Hip joint, Muscle–tendon parameters, Parameter optimization},
[…]
