+++ paper/paper.md
@@ -4,7 +4,7 @@ tags:
   - opensim
   - musculoskeletal modeling
   - musculoskeletal mechanics
-  - simulation and modelling
+  - simulation and modeling
   - biomechanics
   - rigid body dynamics
   - user-interface tooling
@@ -35,16 +35,16 @@ bibliography: paper.bib
 
 # Summary
 
-OpenSim Creator is an open-source graphical interface that lowers the barrier to building and evaluating musculoskeletal models in OpenSim, which previously required editing XML files or custom code. This paper describes its architecture and features, and demonstrates its application in developing new biomechanical models.
+OpenSim Creator is an open-source integrated graphical interface that lowers the barrier to building and evaluating musculoskeletal models in OpenSim, which often required editing XML files or custom code. This paper describes its architecture and features, and demonstrates its application in developing new biomechanical models.
 
 
 # Statement of Need
 
-OpenSim [@Delp:2007; @Seth:2018] is an open-source musculoskeletal modeling platform with a large and growing user base across diverse fields, including clinical [@Marconi:2023], biomechanical [@Porsa:2016], and palaeobiological [@Beesel:2021; @Demuth:2023] studies. However, creating a biomechanical model remains labor-intensive, requiring the careful integration of biological data, and technically demanding, requiring expertise in biomedical imaging, model/data file formats, and musculoskeletal anatomy.
+OpenSim [@Delp:2007; @Seth:2018] is an open-source musculoskeletal modeling platform with a large and growing user base across diverse fields, including clinical [@Marconi:2023], biomechanical [@Porsa:2016], robotics [@Belli:2023], and paleobiological [@Beesel:2021; @Demuth:2023] studies. However, creating a biomechanical model remains labor-intensive, requiring the careful integration of biological data, and technically demanding, requiring expertise in biomedical imaging, model/data file formats, and musculoskeletal anatomy.
 
 For example, building a biomechanical model of a dinosaur involves segmenting bone morphology, rigging skeletal structures, reconstructing musculature, and running biomechanical simulations [@Bishop:2021]. Such workflows often need to be repeated for each subject, and while scaling existing models can reduce this effort, developing robust and unbiased procedures that handle diverse shapes, injuries, or deformities remains challenging [@vanDerKruk:2025]. Real-time visual tools can provide valuable feedback during model building, scaling, and validation.
 
-These challenges not only affect research but also limit accessibility for students and new researchers learning musculoskeletal modeling. As OpenSim use expands beyond engineering (e.g., medicine, palaeobiology) and open education becomes more widespread (e.g., online and open courses [@Weller:2018]), there is a growing need for open-source, visual, and easy-to-install tools that educators can use to teach musculoskeletal model-building concepts clearly — tools that lower barriers to entry compared to traditional code- or engineering-focused workflows.
+These challenges not only affect research but also limit accessibility for students and new researchers learning musculoskeletal modeling. As OpenSim use expands beyond engineering (e.g., medicine, paleobiology) and open education becomes more widespread (e.g., online and open courses [@Weller:2018]), there is a growing need for open-source, visual, and easy-to-install tools that educators can use to teach musculoskeletal model-building concepts clearly — tools that lower barriers to entry compared to traditional code- or engineering-focused workflows.
 
 
 # State of the Field
@@ -53,60 +53,57 @@ Similar modeling challenges are addressed in fields such as 3D design (e.g., Ble
 
 OpenSim includes a user interface (UI) that can edit existing OpenSim models, but it is specialized around providing tools that accept existing model files as input. For example, it includes tools for scaling models, computing inverse kinematics (IK) from marker motion recordings, and computed muscle control (CMC), but it does not provide tools for adding bodies or joints to an existing model. Therefore, to create new models, users typically use a combination of the following three options:
 
[…]
