### 2025-09-27 34f6fdd
Add LinearAlgebra.h uses
