+++ libopensimcreator/UI/TPS2DTab.cpp
@@ -24,6 +24,7 @@
 #include <liboscar/Utils/Assertions.h>
 #include <liboscar/Utils/StdVariantHelpers.h>
 #include <liboscar/Utils/StringHelpers.h>
+#include <simmath/LinearAlgebra.h>
 #include <Simbody.h>
 
 #include <cmath>
