### 2025-09-28 be85067
Remove Simbody.h uses
