+++ libopensimcreator/Utils/TPS3D.cpp
@@ -9,7 +9,6 @@
 #include <liboscar/Utils/ParalellizationHelpers.h>
 #include <liboscar/Utils/Perf.h>
 #include <simmath/LinearAlgebra.h>
-#include <Simbody.h>
 
 #include <ostream>
 #include <ranges>
