+++ libopensimcreator/Documents/CustomComponents/SphereLandmark.cpp
@@ -2,7 +2,6 @@
 
 #include <libopensimcreator/Documents/FrameDefinition/FrameDefinitionHelpers.h>
 
-#include <Simbody.h>  // TODO: OpenSim's Appearance header doesn't include simbody properly
 #include <OpenSim/Simulation/Model/Appearance.h>
 
 osc::fd::SphereLandmark::SphereLandmark()
