### 2025-10-01 8f34564
Adjust headers/variables slightly to ensure compatibility with windows macros/minimal OpenSim usage
