### 2025-10-06 58069e6
De-inline osc::Variant methods
