### 2025-10-06 b58c1ad
Clean up CMakeLists.txt and github workflow slightly
