+++ hellooscar/CMakeLists.txt
@@ -1,20 +1,10 @@
 include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/osc_strict_compiler_options.cmake)  # OSC_STRICT_COMPILER_OPTIONS
 
-add_executable(hellooscar
-    hellooscar.cpp
-)
-set_target_properties(hellooscar PROPERTIES
-    CXX_EXTENSIONS OFF
-)
-target_compile_features(hellooscar PUBLIC
-    cxx_std_23
-)
-target_compile_options(hellooscar PRIVATE
-    ${OSC_STRICT_COMPILER_OPTIONS}
-)
-target_link_libraries(hellooscar PRIVATE
-    oscar
-)
+add_executable(hellooscar hellooscar.cpp)
+set_target_properties(hellooscar PROPERTIES CXX_EXTENSIONS OFF)
+target_compile_features(hellooscar PUBLIC cxx_std_23)
+target_compile_options(hellooscar PRIVATE ${OSC_STRICT_COMPILER_OPTIONS})
+target_link_libraries(hellooscar PRIVATE oscar)
 
 # -------------- installation/packaging ------------- #
 if(DEFINED OSC_EMSCRIPTEN AND ${OSC_EMSCRIPTEN})
