### 2025-10-06 864feca
Make multi-config building easier, add CMakePresets.json
