### 2025-10-06 3caa39b
Add executable flag to scripts, make build analsis run in Debug mode
