+++ docs/source/make-a-leg.rst
@@ -1,5 +1,5 @@
-Make a Lower Leg
-================
+Make a Leg
+==========
 
 .. warning::
 
@@ -8,9 +8,9 @@ Make a Lower Leg
     v4.6 to be released before we remove any "experimental" labelling. We also anticipate
     adding some handy tooling around re-socketing existing joints and defining ``StationDefinedFrame``\s.
 
-In this tutorial, we will be making a model of a lower leg using OpenSim Creator:
+In this tutorial, we will be making a model of a leg using OpenSim Creator:
 
-.. figure:: _static/make-a-lower-leg/final-model-screenshot.jpeg
+.. figure:: _static/make-a-leg/final-model-screenshot.jpeg
     :width: 60%
 
     The model created by this tutorial. It contains three bodies, two joints, three muscles,
@@ -49,13 +49,13 @@ Topics Covered by this Tutorial
 * Adding a wrap surface to the model and associating muscles to that surface.
 
 
-.. _make-a-lower-leg-resources-link:
+.. _make-a-leg-resources-link:
 
 Download Resources
 ------------------
 
 In order to follow this tutorial, you will need to download the associated
-resources (:download:`download here <_static/the-model-warper/make-a-lower-leg_resources.zip>`)
+resources (:download:`download here <_static/the-model-warper/make-a-leg_resources.zip>`)
 and unzip them on your computer.
 
 
@@ -70,7 +70,7 @@ Add a ``pelvis`` Body
 
 Add a pelvis body. For this model, use the following parameters:
 
-.. figure:: _static/make-a-lower-leg/add-pelvis-body.jpeg
+.. figure:: _static/make-a-leg/add-pelvis-body.jpeg
     :width: 60%
 
     Create a body called ``pelvis``. The mass and inertia can be handled later.
@@ -85,21 +85,21 @@ Adding bodies is explained in more detail in :ref:`add-body-with-weldjoint` and
 Attach Pelvis Meshes to the ``pelvis`` Body
 -------------------------------------------
 
-The resources zip described in :ref:`make-a-lower-leg-resources-link` contain two
+The resources zip described in :ref:`make-a-leg-resources-link` contain two
 separate pelvis meshes for the left- and right-side. For this model, we are simplifying
 the pelvis to a single rigid body (``pelvis``). Both meshes need to be attached to it.
 
 To attach meshes to ``pelvis``, right-click it in the ``Navigator`` panel and use
 the ``Add > Geometry`` context menu to attach each pelvis mesh:
 
-.. figure:: _static/make-a-lower-leg/add-geometry-to-pelvis-context-menu.jpeg
+.. figure:: _static/make-a-leg/add-geometry-to-pelvis-context-menu.jpeg
     :width: 60%
 
     Use ``pelvis``'s context menu to ``Add > Geometry`` to it, then select one
     of the pelvis meshes (``pelvis_l.obj`` or ``pelvis_r.obj``). Repeat this process
     for the other pelvis mesh.
 
-.. figure:: _static/make-a-lower-leg/after-attaching-both-pelvis-meshes-to-pelvis.jpeg
+.. figure:: _static/make-a-leg/after-attaching-both-pelvis-meshes-to-pelvis.jpeg
     :width: 60%
 
     The model after attaching both ``pelvis_l.obj`` and ``pelvis_r.obj`` to the
@@ -127,7 +127,7 @@ the top menu bar, located at ``Tools > Import Points``. It will show a popup
[…]
