+++ CHANGELOG.md
@@ -9,6 +9,9 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The MacOS DMG installers should now be signed and notarized by Adam Kewley, which should make
   installing OpenSim Creator considerably easier on MacOS (#1106).
 - The "Make a Lower Leg" tutorial was renamed to "Make a Leg" (#1108).
+- The `knee_frame` definition in the "Make a Leg" tutorial was changed such that the
+  along-knee (along the condyls) axis its primary axis (for the `StationDefinedFrame`),
+  which is the correct formulation and behaves better when warping (#1112).
 
 ## [0.6.0] - 2025/09/08
 
