+++ CHANGELOG.md
@@ -12,6 +12,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The `knee_frame` definition in the "Make a Leg" tutorial was changed such that the
   along-knee (along the condyls) axis its primary axis (for the `StationDefinedFrame`),
   which is the correct formulation and behaves better when warping (#1112).
+- A marker-fixing step was added to the "Make a Leg" tutorial so that it works better
+  with OpenSim's IK solver (#1109).
 
 ## [0.6.0] - 2025/09/08
 
