### 2025-10-28 e87fcd4
Squashed 'third_party/imgui/' changes from 4806a1924f..e7d2d636af
e7d2d636af Merge branch 'master' into docking
9a5d5c45f5 Version 1.92.4
25e5baf5f0 (Breaking) Clipper: commented out legacy IncludeRangeByIndices name obsoleted in 1.89.9. (#6424, #3841)
c8db91b7b4 (Breaking) TreeNode, Selectable: commented out legacy ImGuiTreeNodeFlags_AllowItemOverlap, ImGuiSelectableFlags_AllowItemOverlap names obsoleted in 1.89.7.
94ed5b1408 Docs: improve docking API comments and demo. (#9000)
702cc9993a Relaxed internal assert in MarkItemEdited() some more. (#8997)
96ad003000 Nav, Docking, Selection: Fixed 
[…]
