### 2025-11-16 #8284 @Kevin-Mattheus-Moerman
@davidpagnon thanks. You have left one box unticked (State of the field), is that deliberate please elaborate. If not, please tick it now if possible. Thanks again for your help.
