### 2025-12-17 #8284 @adamkewley
Hi @modenaxe , I've implemented the suggestions as three separate commits, here:

- [c246383](https://github.com/ComputationalBiomechanicsLab/opensim-creator/commit/c24638384bd0cb2f60715c2b4c6f446771601ffb)
- [9420d33](https://github.com/ComputationalBiomechanicsLab/opensim-creator/commit/9420d33b29aba016c413904bf3f20fd16a650328)
- [6cfc43a](https://github.com/ComputationalBiomechanicsLab/opensim-creator/commit/6cfc43a2468a4d6781a418c6c511acfbadf7ed19)

I added the Thelen citation. For the dynamics callout, I've tried to mention that dynamics simulations can be necessary for ensuring the model is balanced/stable and then ensured the features section uses "dynamic simulation" rather than just "simulation" to make the connection between the two sections a little clearer. For the CT scans, I used the terminology "bone meshes" rather than (as you suggest) "bone geometries", not because your suggestion is bad/wrong (it isn't) but purely because the previous sentence is talking about the "mesh importer" and I figured it would be easier to read if I stuck to the meshes naming.

Please let met know if there's anything else!
