### 2025-12-17 #8284 @adamkewley
@Kevin-Mattheus-Moerman , sorry to hassle you with this but I don't think I can change it myself.

The title for the paper was changed to:

> OpenSim Creator: An Interactive User Interface for Building OpenSim Musculoskeletal Models
>
> (*previously*): OpenSim Creator: A Graphical User Interface for Building OpenSim Musculoskeletal Models

It's a minor stylistic change that was made a few versions ago, but I realize that the GitHub issue, any databasing by JOSS, etc. might now be out of sync: is it possible to update it or shall I revert the name (it's no big)?
