### 2026-01-09 #8284 @adamkewley
- Double check authors and affiliations (including ORCIDs):

  - 0000-0002-6505-5374 -> [yes, this is me](https://orcid.org/0000-0002-6505-5374)
  - 0000-0001-5457-0103 -> [yes, this is Julia](https://orcid.org/0000-0001-5457-0103)
  - 0000-0003-4217-1580 -> [yes, this is Ajay](https://orcid.org/0000-0003-4217-1580)

  - Use [0.7.0](https://github.com/ComputationalBiomechanicsLab/opensim-creator/releases/tag/0.7.0) (current latest release) - it contains any software changes since 0.5.21 that were the result of this review process.

- Archive the release on Zenodo/figshare/etc and post the DOI here:
  - All releases are released to Zenodo, 0.7.0 is:
> Kewley, A., Beesel, J., & Seth, A. (2026). OpenSim Creator (0.7.0). Zenodo. https://doi.org/10.5281/zenodo.18172189

- Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper:
  - Uh, not applicable @Kevin-Mattheus-Moerman ? O_o? A software release of OpenSim Creator doesn't include a JOSS paper. It's not a typical research codebase ("oh, here's my Jupyter script, the paper's in `paper/`) unless JOSS's opinion is that all resultant papers must be baked into the repository? Is that something I should handle?

  -   The Zenodo entry has the same license as the software
