### 2026-01-09 #8284 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
